% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{add_data_col_type}
\alias{add_data_col_type}
\title{Add data of col type to datatable}
\usage{
add_data_col_type(dt, data_col_name, data, indices = "")
}
\description{
Accepts an existing dataframe, colname and data to insert.  Adds data
as type \code{char} to avoid issues when adding data with default or error values
from combining with furhter datatables with specific data types.
}
\details{
User should convert cols to appropriate datatypes as needed in analyses.

This function adds a char column, renames the col to \code{data_col_name}, and
finally converts data to char datatype and inserts into datatable.

DT row length MUST equal the length of data.  \code{data} and \code{data_col_name} must be
character vectors.

Can optionally use indices to insert data into specific row indices in new col
in dt.
}
