% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_content}
\alias{addin_create_content}
\title{Create New Content Addin}
\usage{
addin_create_content()
}
\description{
Shiny Gadget generated to create new Insertable Content in a Project Note
at the currently selected line.
}
\details{
Content is any templated insertable text for boilerplate content.  Typically
in ProjectManage it is used to define a set of Standard Operating
Procedures (SOPs) & LOG Sections that record Protocol execution.  The
Content is defined between specific delimiters, defined in
`CONTENT_SEP.txt`: by default the delimiter is a series of `====`.

User defines the following in the gadget interface:

* `contentName` : A descriptive name for the block of insertable content.

* `contentDescription` : A concise description of the block of insertable
  content - what will this content be used for?

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.

This addin calls `create_content()` with appropriate ARGs.
}
