% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-functions.R
\name{addin_create_dialog_viewer}
\alias{addin_create_dialog_viewer}
\title{Create Dialog Viewer for Addin}
\usage{
addin_create_dialog_viewer(title, settings)
}
\description{
Create Dialog Viewer for Addin
}
