% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_programme}
\alias{addin_create_programme}
\title{Create a New Programme Addin}
\usage{
addin_create_programme()
}
\description{
Generates a Shiny Gadget for creating a new Programme inside
a Project Organisation.  User selects a destination in
the file system, Programme name (for the directory),
and Programme title (for the html page).
}
