% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_programme_section}
\alias{addin_create_programme_section}
\title{Create a New Programme Section Addin}
\usage{
addin_create_programme_section()
}
\description{
Generates a Shiny Gadget for creating a new Programme Section inside
a Programme.  User selects a destination in
the file system, Programme Section name (for the directory),
and Programme Section title (for the html page).
}
