% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_doc_2}
\alias{addin_create_project_doc_2}
\title{Create a New Project Document Addin}
\usage{
addin_create_project_doc_2()
}
\description{
TRIED SPLITTING UI AND SERVER - but cannot get this to work with shinyDirChoose
So abandoning this for now...
}
\details{
Generates a Shiny Gadget for creating a new Project Doc inside
a Programme.

User selects a Programme, then destination in the file system (MUST be within
a Programme Dir), Project name, and Project title (for the html page).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
