% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_doc_ui}
\alias{addin_create_project_doc_ui}
\title{User Interface for Creat Project Doc shiny gadget}
\usage{
addin_create_project_doc_ui(orgPath, settings)
}
\arguments{
\item{orgPath}{The Organisation Path - derived from the current working directory.}

\item{settings}{}
}
\description{
Passes some important instance variables:
}
