% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_note_from_doc_gdt}
\alias{addin_create_project_note_from_doc_gdt}
\title{Add a single or group note to doc at selected GDT}
\usage{
addin_create_project_note_from_doc_gdt(selection, settings, orgPath)
}
\description{
Add a single or group note to doc at selected GDT
}
