% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_note_from_single}
\alias{addin_create_project_note_from_single}
\title{Add new note from existing single note - inheriting all note GDTs}
\usage{
addin_create_project_note_from_single(selection, settings, orgPath)
}
\description{
Add new note from existing single note - inheriting all note GDTs
}
