% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_org}
\alias{addin_create_project_org}
\title{Create a New Project Org Addin}
\usage{
addin_create_project_org()
}
\description{
Generates a Shiny Gadget for initialising
a new Project Organisation.  User selects a destination in
the file system, Organisation name (for the directory) and
Organisation title (for the html page).
}
