% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_project_org_server}
\alias{addin_create_project_org_server}
\title{Create Project Org: Shiny Server}
\usage{
addin_create_project_org_server(input, output, session)
}
\description{
Creates a Proect Org using \code{create_project_org()} function, with reactive
inputs: input$dir, input$organisationName, input$organisationTitle.
}
