% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_protocol}
\alias{addin_create_protocol}
\title{Add a New Protocol Addin}
\usage{
addin_create_protocol()
}
\description{
Generates a Shiny Gadget for adding a new Protocol inside the `protocol`
directory of the containing Programme, and inserting a link to this
inside the active Project Note.
}
\details{
The User must be in an active Project Note, and the selected
line and column is where the Protocol Link will be inserted.
This command works for all Project Notes.  A reciprocal link is formed
from the Protocol to the Project Note.

User selects a Protocol name, and (Optionally) Protocol
 title (filled by default with SPACES replacing "_" and "-" from name).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
