% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_create_subnote_from_group}
\alias{addin_create_subnote_from_group}
\title{Add new subnote under group note - inheriting all group note GDTs}
\usage{
addin_create_subnote_from_group(selection, settings, orgPath)
}
\description{
Add new subnote under group note - inheriting all group note GDTs
}
