% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-addins.R
\name{addin_datatable_add_groups}
\alias{addin_datatable_add_groups}
\title{Add Datatable containing groups to existing samples}
\usage{
addin_datatable_add_groups()
}
\description{
Generates a Shiny Gadget for adding group assignment data to existing samples
in a new datatable.
}
\details{
User selects a datatable name : list of existing datatables
in the current Rmd.

User defines the data table column names and values : these MUST begin with
the prefix \code{group} to define each group set.  Then the group codes are
entered, separated by spaces after the group column name.

Multiple groups can be entered by repeating this layout.  The groups are
automatically distributed across the samples, which may need re-ordering.
The columns are spaced according to the max width of the group column names
and the group codes.

A datatable with the function GROUP is added with sample-first layout.
}
