% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-addins.R
\name{addin_datatable_create}
\alias{addin_datatable_create}
\title{Create Datatable from File}
\usage{
addin_datatable_create()
}
\description{
Generates a Shiny Gadget for initialising
a new datatable from ID list, plus optional cols with vals.  A datatable
is formed and inserted in current Rmd at the current line.
}
