% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-addins.R
\name{addin_datatable_create_from_file}
\alias{addin_datatable_create_from_file}
\title{Create a New Datatable from a CSV File in R Markdown}
\usage{
addin_datatable_create_from_file()
}
\value{
This function does not return any value. It modifies the Rmd document by creating
and inserting a new datatable (derived from the selected CSV file) at the cursor's
current position.
}
\description{
Generates a Shiny Gadget for initializing a new datatable from a CSV file selected
by the user. The content of the CSV file is converted to a datatable format and
inserted at the current cursor position in the active R Markdown document.
}
\details{
This function allows the user to:
\itemize{
\item Select a CSV file from their file system through a Shiny Gadget.
\item The selected file is then read, converted to a datatable format, and inserted
at the current cursor position in the R Markdown document.
\item The user can also select from different datatable types using a set of checkbox options.
\item The function ensures the Rmd document is saved before any modifications are made.
}
}
\examples{
# Example usage:
# Run this function to create a new datatable from a selected CSV file and insert it in the Rmd.
addin_datatable_create_from_file()

# During usage:
# - Select a CSV file using the Shiny Gadget.
# - The CSV file content will be converted to a datatable format and inserted at the cursor position.

}
