% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-addins.R
\name{addin_datatable_import}
\alias{addin_datatable_import}
\title{Import EXISTING samples and reps from Datatable in source Project Note}
\usage{
addin_datatable_import()
}
\description{
Generates a Shiny Gadget for importing existing samples and reps derived from
datatables that exist in a source Project Note into the currently active
destination Project Note.
}
\details{
* User selects the line in the current Project Note where the IMPORT will
be written.

* User selects a source Project Note, from the current ORGANISATION: This
note is read for all EXISTING samples/reps in all datatables, which are
presented in SUMMARY FORMAT

* User filters the datatable for desired characteristics of samples/reps for
import.

* User selects the samples by clicking on the rows in the datatable.

* User selects reps by filling in the rep number in the left hand column of
selected sample rows.

* User selects DONE : samples/reps are first

  + EXPORTED from the source Project Note : Export datatable is updated/written
    as needed in the source note, pointing to the destination Project Note

  + IMPORTED to the destination Project Note : Import datatable is written
    which includes the samples/reps selected, pointing to the source Project
    Note.
}
