% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-addins.R
\name{addin_datatable_insert_tibble}
\alias{addin_datatable_insert_tibble}
\title{Insert datatable from tibble in active RStudio file}
\usage{
addin_datatable_insert_tibble(
  tb,
  dt_name,
  dt_function = "CREATE",
  dt_length = 100
)
}
\description{
Insert datatable from tibble in active RStudio file
}
