% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-addins.R
\name{addin_datatable_resample}
\alias{addin_datatable_resample}
\title{Add Resample Datatable to existing samples}
\usage{
addin_datatable_resample()
}
\description{
Generates a Shiny Gadget for adding data to existing samples in a new
datatable.
}
\details{
User selects a datatable name : list of existing datatables
in the current Rmd.

User selects datatable layout : samples, variables, timetable.

User selects the ID format to add : ALL, all-IDs, or a group set.

The datatable with function ADD_DATA is added with appropriate layout.

If the cursor is already INSIDE a GENERIC ADD_DATA, which has a valid
ADD_DATA layout (and is already sample, variable or timetable), then
user just needs to select the IDs to add : ALL, all-IDs, or a group set.

The generic ADD_DATA table will be updated to contain the correct IDs
from existing samples for adding data.
}
