% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-addins.R
\name{addin_extract_todos_ui}
\alias{addin_extract_todos_ui}
\title{UI for the Extract TODOs Shiny Gadget (Write to File)}
\usage{
addin_extract_todos_ui(orgPath, fromFilePath, settings)
}
\arguments{
\item{orgPath}{The organisation path, used as a root in \code{shinyDirChoose}.}

\item{fromFilePath}{The path of the currently active RStudio document (display only).}

\item{settings}{A list of settings from \code{get_settings_yml(orgPath)}.}
}
\description{
Creates a \code{miniUI::miniPage} with:
\itemize{
\item The path of the active RStudio file (read-only).
\item A directory chooser for \code{location}.
\item A date input for choosing \code{date}.
\item Checkboxes for \code{onlyFirstTodoPerFile}, \code{sortByFileModTimeDesc},
and \code{priorityFirst}.
}
}
\keyword{internal}
