% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_content}
\alias{addin_insert_content}
\title{Select & Insert Content}
\usage{
addin_insert_content()
}
\description{
A Directory Tree within the Organisation is used to screen all Project Notes
for all declared Contents, based on the Content Delimiter Syntax (defined
in `settings[["ContentSep"]]`).  This Directory Tree can be adjusted.
}
\details{
All Insertable Content present in Project Notes in the Directory Tree are
shown in a DT datatable by content title.

Individual Content can be selected to show the summary information.

A Preview view of the Content can be opened to read through the entire
Content in the ADDIN.

Selected content can be inserted into the current selection in a
Destination Project Note via the DONE button.

Using DT to display a datatable in the shiny gadget here.  See the tutorial
at https://rstudio.github.io/DT/shiny.html
}
