% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_date}
\alias{addin_insert_date}
\title{Insert Date Addin}
\usage{
addin_insert_date()
}
\description{
Inserts the current date into the current selection in RStudio Source
Editor.  Uses \code{YYYY/MM/DD} format.
}
\details{
The timezone used is specified in settings.yml under the \code{DateTimeZone}
parameter if settings.yml can be found, otherwise it uses the \code{UTC}
timezone.
}
