% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_datetime}
\alias{addin_insert_datetime}
\title{Insert Datetime Addin}
\usage{
addin_insert_datetime()
}
\description{
Inserts the current datetime into the current selection in RStudio Source
Editor. Uses \code{YYYY/MM/DD:hh:mm} format.
}
\details{
The timezone used is specified in settings.yml under the \code{DateTimeZone}
parameter if settings.yml can be found, otherwise it uses the \code{UTC}
timezone.
}
