% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_doc_deliverable_section}
\alias{addin_insert_doc_deliverable_section}
\title{Insert Deliverable Section to end of Project Doc GDT}
\usage{
addin_insert_doc_deliverable_section()
}
\description{
Collects selection object from active document and attempts to insert a
new deliverable section to the end of this Project Doc's selected
Goal Section.
}
