% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_doc_goal_section}
\alias{addin_insert_doc_goal_section}
\title{Insert Goal Section to end of Project Doc GDT}
\usage{
addin_insert_doc_goal_section()
}
\description{
Collects selection object from active document and attempts to insert a
new goal section to the end of this document IF a Project Doc.
}
