% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-addins.R, R/insert-addins.R
\name{addin_insert_hyperlink}
\alias{addin_insert_hyperlink}
\title{Insert a Hyperlink}
\usage{
addin_insert_hyperlink()

addin_insert_hyperlink()
}
\description{
Insert a RELATIVE Hyperlink from an Open RStudio File to the currently
Active RStudio file, at the current cursor position.

Insert a RELATIVE Hyperlink from an Open RStudio File to the currently
Active RStudio file, at the current cursor position.
}
