% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_hyperlink_server}
\alias{addin_insert_hyperlink_server}
\title{Insert Hyperlink: Shiny Server}
\usage{
addin_insert_hyperlink_server(input, output, session)
}
\description{
Insert hyperlink using \code{create_hyperlink_no_ext()} function.
}
