% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_hyperlink_ui}
\alias{addin_insert_hyperlink_ui}
\title{User Interface for Insert Hyperlink}
\usage{
addin_insert_hyperlink_ui(filename, line, column, numberedFileList)
}
\arguments{
\item{filename}{Name of file where hyperlink will be written to}

\item{line}{Line in file where hyperlink will be written}

\item{column}{Column in file where hyperlink will be written}

\item{numberedFileList}{FileList of open Rmd files in RStudio, numbered.}
}
\description{
Passes some important instance variables:
}
