% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_protocol}
\alias{addin_insert_protocol}
\title{Select & Insert Protocol}
\usage{
addin_insert_protocol()
}
\description{
Using DT to display a datatable in the shiny gadget here.  See the tutorial
at https://rstudio.github.io/DT/shiny.html
}
