% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_section_sep}
\alias{addin_insert_section_sep}
\title{Insert Section Sep Addin}
\usage{
addin_insert_section_sep(index)
}
\description{
Load section `sep\if{html}{\out{<index>}}`` from templates DIR and insert into current
cursor position.
}
