% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_open_daily_journal_server}
\alias{addin_open_daily_journal_server}
\title{Server Logic for Addin: Open Daily Journal}
\usage{
addin_open_daily_journal_server(input, output, session)
}
\arguments{
\item{input}{Reactive input object. Contains inputs from the Shiny gadget UI, including the selected date from the calendar (\code{input$calendar}) and the "Done" button action (\code{input$done}).}

\item{output}{Reactive output object. Used to render outputs in the UI, such as the text displaying the selected day (\code{output$selected_day_text}).}

\item{session}{Reactive session object. Used to manage the Shiny gadget's session.}
}
\value{
This function does not return a value. It is invoked as part of the Shiny server logic for the gadget.
}
\description{
Implements the server logic for the "Open Daily Journal" Shiny gadget.
Handles date selection from the calendar, determines the selected day,
and creates or navigates to the journal file as needed.
}
\details{
This function processes user interactions with the "Open Daily Journal" gadget:
\itemize{
\item Monitors the calendar input for date selection.
\item Calculates the date selected.
\item Displays the calculated date in the UI.
\item Responds to the "Done" button click:
\itemize{
\item If a day is selected, it creates the journal for that day (if necessary) and navigates to the journal file in RStudio.
\item If no day is selected, the gadget is closed without any action.
}
}
}
\note{
This function is designed to be used exclusively within a Shiny gadget as part of the "Open Daily Journal" add-in.
}
\examples{
# Example usage within a Shiny gadget:
server <- function(input, output, session) {
  addin_open_daily_journal_server(input, output, session)
}

}
\seealso{
\itemize{
\item \code{\link{addin_open_daily_journal_ui}} for the corresponding UI function.
\item \code{\link{create_daily_journal}} for creating daily journal files.
\item \code{\link{addin_rstudio_nav}} for navigating to files in RStudio.
}
}
