% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-addins.R
\name{addin_open_weekly_journal_server}
\alias{addin_open_weekly_journal_server}
\title{Server Logic for Addin: Open Weekly Journal}
\usage{
addin_open_weekly_journal_server(input, output, session)
}
\arguments{
\item{input}{Reactive input object. Contains inputs from the Shiny gadget UI, including the selected date from the calendar (\code{input$calendar}) and the "Done" button action (\code{input$done}).}

\item{output}{Reactive output object. Used to render outputs in the UI, such as the text displaying the Monday of the selected week (\code{output$selected_week_text}).}

\item{session}{Reactive session object. Used to manage the Shiny gadget's session.}
}
\value{
This function does not return a value. It is invoked as part of the Shiny server logic for the gadget.
}
\description{
Implements the server logic for the "Open Weekly Journal" Shiny gadget.
Handles date selection from the calendar, determines the corresponding week's
Monday, and creates or navigates to the weekly journal file as needed.
}
\details{
This function processes user interactions with the "Open Weekly Journal" gadget:
\itemize{
\item Monitors the calendar input for date selection.
\item Calculates the Monday of the selected week using \code{lubridate::floor_date}.
\item Displays the calculated Monday date in the UI.
\item Responds to the "Done" button click:
\itemize{
\item If a week is selected, it creates the journal for that week (if necessary) and navigates to the journal file in RStudio.
\item If no week is selected, the gadget is closed without any action.
}
}
}
\note{
This function is designed to be used exclusively within a Shiny gadget as part of the "Open Weekly Journal" add-in.
}
\examples{
# Example usage within a Shiny gadget:
server <- function(input, output, session) {
  addin_open_weekly_journal_server(input, output, session)
}

}
\seealso{
\itemize{
\item \code{\link{addin_open_weekly_journal_ui}} for the corresponding UI function.
\item \code{\link{create_weekly_journal}} for creating weekly journal files.
\item \code{\link{addin_rstudio_nav}} for navigating to files in RStudio.
}
}
