% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_project_doc_deliverable}
\alias{addin_rename_project_doc_deliverable}
\title{ADDIN to rename Project Doc Deliverable}
\usage{
addin_rename_project_doc_deliverable(selection, settings, orgPath)
}
\arguments{
\item{selection}{Selection made on Project Doc Deliverable.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\description{
ADDIN to rename Project Doc Deliverable
}
