% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_project_doc_goal}
\alias{addin_rename_project_doc_goal}
\title{ADDIN to rename Project Doc GOAL}
\usage{
addin_rename_project_doc_goal(selection, settings, orgPath)
}
\arguments{
\item{selection}{Selection made on Project Doc Goal.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\description{
ADDIN to rename Project Doc GOAL
}
