% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_project_doc_task}
\alias{addin_rename_project_doc_task}
\title{ADDIN to rename Project Doc Task}
\usage{
addin_rename_project_doc_task(selection, settings, orgPath)
}
\arguments{
\item{selection}{Selection made on Project Doc Task.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\description{
ADDIN to rename Project Doc Task
}
