% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_project_file}
\alias{addin_rename_project_file}
\title{Rename a Project File}
\usage{
addin_rename_project_file()
}
\description{
This Addin renames the currently active project file - Note or Doc - with a
new filename, and a new title.
}
\details{
All links are updated throughout the organisation to the project file.
}
