% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_project_file_header}
\alias{addin_rename_project_file_header}
\title{ADDIN to rename Project File Section Header}
\usage{
addin_rename_project_file_header(selection, settings, orgPath)
}
\arguments{
\item{selection}{Selection made on Project Doc Task.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\description{
Section Headers defined in Markdown plasintext files, beginning wiht \verb{#}.
}
