% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-addins.R
\name{addin_rename_projectmanagr_component}
\alias{addin_rename_projectmanagr_component}
\title{Rename a ProjectManagr Component}
\usage{
addin_rename_projectmanagr_component()
}
\description{
This Addin provides access to renaming any projectmanagr component,
including:
}
\details{
* ProjectManagr Project Note filenames

* ProjectManagr Project Doc filenames

* Directories inside a ProjectManagr Organisation

This Addin ensures any renaming will automatically update all ProjectManagr
components to this change, including hyperlinks, GDT links, and paths, as
needed:

* ProjectManagr Project Note renaming will update all hyperlinks to the note
to use the new name.

* ProjectManagr Project Doc renaming will update all hyperlinks & GDT links
to the Doc to use the new name.

* Directory renaming will depend on the directory type:
    + If a user-defined directory containing project notes, all note prefixes
      (inherited from the directory name) will update, as well as any strings
      that contain /{{DIRNAME}}/?? To update all links that contain the dir?
}
