% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-functions.R
\name{addin_rstudio_nav}
\alias{addin_rstudio_nav}
\title{Navigate to file in RStudio}
\usage{
addin_rstudio_nav(filePath)
}
\description{
Uses \code{rstudioapi} to navigate to filePath and its parent, and set the working
directory, once files have been created by ADDIN.
}
