% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-addins.R
\name{addin_search_project_org}
\alias{addin_search_project_org}
\title{Create a New Project Org Addin}
\usage{
addin_search_project_org()
}
\description{
Generates a Shiny Gadget for searching a ProjectManagr Organisation.
User can select the organisations & the scope for searching within
it - default is the root of organisation of current working directory.
}
\details{
Search terms cover only the ProjectManagr files by default - this scope
can be expanded to include other files as set by the user.
}
