% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinAddHyperLink.R
\name{addinAddHyperLink}
\alias{addinAddHyperLink}
\title{Add a RELATIVE Hyperlink from an Open RStudio File to the currently Active RStudio
file, at the current cursor position.}
\usage{
addinAddHyperLink()
}
\description{
Add a RELATIVE Hyperlink from an Open RStudio File to the currently Active RStudio
file, at the current cursor position.
}
