% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinAddProjectLink.R
\name{addinAddProjectLink}
\alias{addinAddProjectLink}
\title{Add a New Project Link}
\usage{
addinAddProjectLink()
}
\description{
Generates a Shiny Gadget for adding a new Project Link inside
a Project Doc.  The Link must be FROM a (Source) Project Docs'
Goal/Del/Task TO a Project Note (Single, SubNote, or a whole
Group of Notes), or a Destination Project Doc.
}
\details{
The User must select a Task in  the Source Project Document.
A link will then be made to a User selected Project Note or
Project Doc in the file system.

The ADDIN stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
