% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinAddProjectNote.R
\name{addinAddProjectNote}
\alias{addinAddProjectNote}
\title{Add a New Project Note}
\usage{
addinAddProjectNote()
}
\description{
Generates a Shiny Gadget for adding a new Project Note inside
a Project Doc.
}
\details{
User selects a destination in
the file system (MUST be a Programme Dir), Project name, and
Project title (for the html page).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
