% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinAddProtocolCheatSheet.R
\name{addinAddProtocolCheatSheet}
\alias{addinAddProtocolCheatSheet}
\title{Add a New Protocol or CheatSheet}
\usage{
addinAddProtocolCheatSheet()
}
\description{
Generates a Shiny Gadget for adding a new Protocol ot CheatSheet inside
the SOP/ dir of the containing Programme, and inserting a link to this
inside the active Proejct Note.
}
\details{
The User must be in an active Project Note, and the selected
line and column is where the Protocol/CheatSheet Link will be inserted.
This command works for all Project Notes.  A reciprocal link is formed
from the Protocol/Cheatsheet to the Project Note.

User selects a Protocol/CheatSheet name, and (Optionally) Protocol/CheatSheet
 title (filled by default with SPACES replacing "_" and "-" from name).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
