% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinAddProtocolCheatSheet.R,
%   R/addinAddProtocolReference.R
\name{addinAddProtocolReference}
\alias{addinAddProtocolReference}
\title{Add a New Protocol or Reference}
\usage{
addinAddProtocolReference()

addinAddProtocolReference()
}
\description{
Generates a Shiny Gadget for adding a new Protocol or Reference inside
the SOP/ dir of the containing Programme, and inserting a link to this
inside the active Project Note.

Generates a Shiny Gadget for adding a new Protocol or Reference inside
the SOP/ dir of the containing Programme, and inserting a link to this
inside the active Project Note.
}
\details{
The User must be in an active Project Note, and the selected
line and column is where the Protocol/Reference Link will be inserted.
This command works for all Project Notes.  A reciprocal link is formed
from the Protocol/Reference to the Project Note.

User selects a Protocol/Reference name, and (Optionally) Protocol/Reference
 title (filled by default with SPACES replacing "_" and "-" from name).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.


NO LONGER exporting this

The User must be in an active Project Note, and the selected
line and column is where the Protocol/Reference Link will be inserted.
This command works for all Project Notes.  A reciprocal link is formed
from the Protocol/Reference to the Project Note.

User selects a Protocol/Reference name, and (Optionally) Protocol/Reference
 title (filled by default with SPACES replacing "_" and "-" from name).

Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
