% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinCreateProgramme.R
\name{addinCreateProgramme}
\alias{addinCreateProgramme}
\title{Create a New Programme}
\usage{
addinCreateProgramme()
}
\description{
Generates a Shiny Gadget for creating a new Programme inside
a Project Organisation.  User selects a destination in
the file system, Programme name (for the directory),
Programme prefix (used to identify Projects with this
Programme), and Programme title (for the html page).
}
