% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinCreateProjectDoc.R
\name{addinCreateProjectDoc}
\alias{addinCreateProjectDoc}
\title{Create a New Project Document}
\usage{
addinCreateProjectDoc()
}
\description{
Generates a Shiny Gadget for creating a new Project Doc inside
a Programme.  User selects a destination in
the file system (MUST be a Programme Dir), Project name, and
Project title (for the html page).
}
\details{
Stipulates any errors in the input, and can only be completed
when these errors have been resolved.
}
