% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinCreateProjectOrg.R
\name{addinCreateProjectOrg}
\alias{addinCreateProjectOrg}
\title{Create a New Project Org}
\usage{
addinCreateProjectOrg()
}
\description{
Generates a Shiny Gadget for initialising
a new Project Organisation.  User selects a destination in
the file system, Organisation name (for the directory) and
Organisation title (for the html page).
}
