% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinCreateVolume.R
\name{addinCreateVolume}
\alias{addinCreateVolume}
\title{Create a New Volume}
\usage{
addinCreateVolume()
}
\description{
Generates a Shiny Gadget for creating a new Volume inside
a Project Organisation.  User selects a destination in
the file system, Volume name (for the symlink), and
optionally the current working directory (set to
current WD).  The addin will call createVolume.
}
