% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinCreateVolumeDir.R
\name{addinCreateVolumeDir}
\alias{addinCreateVolumeDir}
\title{Create a Dir on a Volume}
\usage{
addinCreateVolumeDir()
}
\description{
Generate a DIR on a Volume, and symlinks it inside the active PRoject Notes' DIR.
}
\details{
Generates a Shiny Gadget for creating a new Programme inside
a Project Organisation.  User selects a destination in
the file system, Programme name (for the directory),
Programme prefix (used to identify Projects with this
Programme), and Programme title (for the html page).
}
