% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinDatatableCreate.R
\name{addinDatatableCreate}
\alias{addinDatatableCreate}
\title{Create Datatable from File}
\usage{
addinDatatableCreate()
}
\description{
Generates a Shiny Gadget for initialising
a new datatable from ID list, plus optional cols with vals.  A datatable
is formed and inserted in current Rmd at the current line.
}
