% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinInsertDateTimeString.R
\name{addinInsertDateTimeString}
\alias{addinInsertDateTimeString}
\title{Insert DateTime}
\usage{
addinInsertDateTimeString()
}
\description{
Inserts DateTime, rounded to nearest 5 min interval,
at the current cursor location.
}
