% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinInsertProtocol.R
\name{addinInsertProtocol}
\alias{addinInsertProtocol}
\title{Insert Protocol into Experimental Note}
\usage{
addinInsertProtocol()
}
\description{
Generates a Shiny Gadget for inserting an SOP Protocol into an Experimental Note.  SOPs
from the Experimental Note's Programme are provided in a dropdown list for selection.
The PROTOCOL from the SOP (excluding Notes by default) will be pasted at the cursor
location.
}
\details{
This allows standard operating procedures to be defined and inserted for repeatable and
reproducible experiments and documentation.  The SOPs can include Procedures and detailed
steps to follow, and R chunks of code for encoding data on the processes and samples.

The active document MUST be a Project Note, MUST be inside a PROGRAMME, and there MUST
be SOPs to add!  User can specify to include all notes in the Experimental Note if
desired - these can provide tips and tricks for successfully implementing the procedures.
}
