% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinInsertToDoString.R
\name{addinInsertToDoString}
\alias{addinInsertToDoString}
\title{Insert ToDo String}
\usage{
addinInsertToDoString()
}
\description{
Inserts TODO, bullet with header and sub-bullet,
at the current cursor location.
}
