% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLinkProjectDoc.R
\name{addLinkProjectDoc}
\alias{addLinkProjectDoc}
\title{Add a New Link from a Project Doc Goal/Del/Task to another Project Doc}
\usage{
addLinkProjectDoc(projectDocPath, selection)
}
\arguments{
\item{projectDocPath}{The path to the DESTINATION Project Doc.  This MUST be in a sub-directory or lower inside a
PROGRAMME Directory.  This MUST be an ABSOLUTE PATH.  This is the Project Doc that FULFILLS another Project Docs'
Goal/Del/Task.}

\item{selection}{List containing the Goal, Del, Task selected from the Project Doc, as well as other useful
information - lines of Task/Del/Goal, SOURCE projectDoc path, content of selection line.  See cursorSelection()
or userSelection().}
}
\description{
This Function adds a new Link from one Project Doc (source) to another Project Doc (destination).  The link is made to a
SOURCE Project Doc GOAL/DEL/TASK, and this GOAL/DEL/TASK is added to the DESTINATIONS Project Docs' OBJECTIVES.  If the
Project Doc does not contain ANY objectives (by default they do not), an Objectives Section is added to the start of
the Document.
}
