% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLinkProjectGroup.R
\name{addLinkProjectGroup}
\alias{addLinkProjectGroup}
\title{Add a New Link from a Project Doc Goal/Del/Task to a Group of Project Notes}
\usage{
addLinkProjectGroup(headerNotePath, selection)
}
\arguments{
\item{headerNotePath}{The path to the Header Note.  This MUST be in a sub-directory or lower inside a
PROGRAMME Directory.  This MUST be an ABSOLUTE PATH.}

\item{selection}{List containing the Goal, Del, Task selected from the Project Doc, as well as other useful
information - lines of Task/Del/Goal, projectDoc path, content of selection line.  See cursorSelection()
or userSelection().}
}
\description{
This Function adds a new Link from an Existing Group of Project Notes (including the Header Note and
all existing Sub Notes) to a Project Doc.  The link is made to a Project Doc GOAL/DEL/TASK, and this
GOAL/DEL/TASK is added to the Project Notes' OBJECTIVES.
}
