% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{addLinks}
\alias{addLinks}
\title{Add Project Doc Links to Project Note.}
\usage{
addLinks(projNoteContents, projNotePath, projDocList)
}
\arguments{
\item{projNoteContents}{character vector containing the initial contents of a project note.}

\item{projNotePath}{ABSOLUTE path to the location on the filesystem where project note will be saved.}

\item{projDocList}{List containing all Project Docs - Absolute path, Goal/Del/Task.  Like the list
returned by getProjectNoteDocLinkList().}

\item{summaryBullet}{The Contents of the Summary Bullet - initial default is "* [POSIXct date-time]"}
}
\description{
Assumes the Project Note MAY not have ANY Project Doc links written to it, and
therefore still contain the original Objective.Rmd template Variables.
}
