% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReference.R
\name{addReference}
\alias{addReference}
\title{Add a New Reference}
\usage{
addReference(
  projectNotePath,
  referenceName,
  referenceTitle = "",
  referenceTemplate = "Reference-Template.Rmd"
)
}
\arguments{
\item{projectNotePath}{The ABSOLUTE path of the Project Note.}

\item{referenceName}{The name of the reference, a Title with all SPACES replaced
with - or _.}

\item{referenceTitle}{The title of the reference, by default the name with all - and _ replaced
with SPACES.}

\item{referenceTemplate}{Template to use, as found in the `config/templates/` directory.  Default is
"Reference-Template.Rmd"}
}
\description{
This Function adds a new Reference to a Project Note - saved in the PROGRAMMES' REF/ Dir.
The Reference is formed using an Rmd template: Reference-Template.Rmd from the projectmanagr
package. Cheatheets compile to PDF as standard, and can therefore be used as independent files.
}
\details{
References are stored in the REF/ directory inside the PROGRAMME Directory.  Each Reference
exists in its own directory, to keep its compiled files together.

The Reference source Rmd will link to its creating Project Note, and the Project Note will link to the
compiled PDF of the Reference.

References can further be inserted into new Project Notes, using the insertProcotol() function, which
allows convenient insertion of documentation into new Notes.
}
