% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{addSubNoteLinkToDocs}
\alias{addSubNoteLinkToDocs}
\title{Add Project Doc Links to Project Note.}
\usage{
addSubNoteLinkToDocs(projDocList, subNotePath, headerPath)
}
\arguments{
\item{projDocList}{List containing all Project Docs - Absolute path, Goal/Del/Task.  Like the list
returned by getProjectNoteDocLinkList().}

\item{subNotePath}{ABSOLUTE path to the location on the filesystem where sub note is saved.}

\item{headerPath}{the Absolute path to the header note, which the subnote is stored under.}
}
\description{
Assumes the Project Note MAY not have ANY Project Doc links written to it, and
therefore still contain the original Objective.Rmd template Variables.
}
