% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSubNoteToGroup.R
\name{addSubNoteToGroup}
\alias{addSubNoteToGroup}
\title{Add a New Sub Note to a Project Group}
\usage{
addSubNoteToGroup(subNoteName, subNotePrefix, subNoteDir, selection,
  subNoteTitle = "", subNoteTemp = "Project-Note-Sub-Template.Rmd")
}
\description{
This Function adds a Sub Project Note to a Project Group:
}
\details{
- The subNote is placed into the Project Group Directory

- Links to the subNote are added to the project doc and the project group header
note.

subNoteName - the name of the Project Sub Note, a Title with all SPACES replaced
with - or _.

subNotePrefix - the whole subNotePrefix, including identifier and Major
Numbering, separated by ~, and finally the Minor Numbering system (or subNote numbering),
separated by -.

subNoteDir - the directory where the Sub Note will be stored.  This will be the Project Group
Note Directory.

selection - List containing the Goal, Del, Task selected from the Project Doc, as well as other useful
information - lines of Task/Del/Goal, projectDoc path content of selection line.  See cursorSelection()
or userSelection().

subNoteTitle - OPTIONAL title for the Project Note.  Default is to use subNoteName and replace
all _ and - with SPACES.

subNoteTemp - OPTIONAL template to use, as found in the `templates/` directory.  Default is
"Project-Note-Simple-Template.Rmd"
}
