% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{build_datatable_from_dataframe}
\alias{build_datatable_from_dataframe}
\title{Build Datatable from Dataframe}
\usage{
build_datatable_from_dataframe(
  ID_col,
  IDs,
  data_cols,
  data,
  dt_function,
  datatable_name,
  MAX_DATATABLE_LENGTH,
  DATATABLE_SPACER_CHAR
)
}
\arguments{
\item{ID_col}{Character vector of ID col name.}

\item{IDs}{Character vector of sample IDs}

\item{data_cols}{Character vector of data table column titles}

\item{data}{A list of vectors containing all the data to insert into the
datatable.  Length of list must match length of data_cols vector.
This method now supports adding multi-obs data.  Length of each
vector in data must be a MULTIPLE of IDs length.  If 2 or more multiples,
the data is added in a multi-obs set of rows for each ID: Added in data-first
order - so ID\link{1} gets data\link{1}, data\link{2}; ID\link{2} gets data\link{3}, data\link{4} etc.}

\item{dt_function}{The FUNCTION of the datatable to be built.}

\item{datatable_name}{String of data table name - default "samples".}

\item{MAX_DATATABLE_LENGTH}{Int of data table max length in characters.
Typically 100.}

\item{DATATABLE_SPACER_CHAR}{Character used to underline column headers.
Typically '='.}
}
\description{
This handles multi columns correctly - when spaces exist between strings in
one data value, the data is split and correctly placed into a datatable
on separate rows.
}
\details{
Handles dataframes with differnet numbers of multi-rows across different IDs.

Creates a new Sample DataTable vector. The Sample DataTable will contain an
initial ID column containing the IDs vector, and an optional set of extra
data columns as specified in the data_cols vector.  If the table exceeds
\code{dt_length} characters (default 100), then the table is split into multiple
tables, with IDs as first col, and subsequent data_cols given in subsequent
tables.

This method also supports adding Multi-Observations:  Multiple data points
for a given ID.  These are entered under the data_col and split across
lines next to an ID.  The length of each vector in \code{data} must be a multiple
of length of IDs.  If more than 1, the data is split between IDs in
data-first order.
}
