% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{build_datatable_from_tibble}
\alias{build_datatable_from_tibble}
\title{Return a datatable Rmd character vector from a tibble}
\usage{
build_datatable_from_tibble(
  tb,
  datatable_name,
  dt_function = "CREATE",
  dt_length = 100
)
}
\arguments{
\item{tb}{A tibble to parse into projectmanagr datatable format.}

\item{datatable_name}{The name to set in the datatable.}

\item{dt_function}{CREATE ADD_DATA RESAMPLE IMPORT EXPORT}
}
\description{
Assumes the tibble has initial col ID with all IDs in it. Spaces the
columns based on the col header and contents widths.
}
\details{
NOT CLEAR THIS FUNCTION HANDLES VARIALBE MULTI-ROWS CORRECTLY!

Use `build_datatable_from_dataframe()`
}
