% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{build_note_first_layout}
\alias{build_note_first_layout}
\title{Build the final output with note-first layout}
\usage{
build_note_first_layout(
  allTodos,
  sortByFileModTimeDesc,
  itemLines,
  noteLines,
  fromFilePath,
  orgPath,
  date,
  location,
  collectionLines
)
}
\arguments{
\item{allTodos}{The data frame of all collected TODO items.}

\item{sortByFileModTimeDesc}{Boolean indicating whether to sort files by newest first.}

\item{itemLines}{The lines from \code{itemTemplate}.}

\item{noteLines}{The lines from \code{noteTemplate}.}

\item{fromFilePath}{Path from which links are computed.}

\item{orgPath}{Path to the project root for relative references.}

\item{date}{The user's "today" date string.}

\item{location}{The original location passed to \code{extract_todos()}.}

\item{collectionLines}{The lines from \code{collectionTemplate}.}
}
\value{
A character vector of the final doc in Rmd form.
}
\description{
In "note-first" mode, we group items by file, sorted by newest file first
(if \code{sortByFileModTimeDesc=TRUE}), then within the file we sort items by
descending priority, ascending date/time, and ascending index in file.
}
\keyword{internal}
