% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{check_assigned_ids_reps}
\alias{check_assigned_ids_reps}
\title{return assigned IDs : that already have data in col}
\usage{
check_assigned_ids_reps(dt, datacol)
}
\description{
return assigned IDs : that already have data in col
}
