% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{check_datatable_validity}
\alias{check_datatable_validity}
\title{check syntax of dt_vector for validity}
\usage{
check_datatable_validity(dt_vector, table_start_line)
}
\description{
check syntax of dt_vector for validity
}
