% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{check_dirs_org}
\alias{check_dirs_org}
\title{Check Dirs for Organisations}
\usage{
check_dirs_org(dirs)
}
\description{
Simple wuick check of directories in \code{dirs} for presence of any ProjectManagr
Organisations.
}
