% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{check_divisions_ids}
\alias{check_divisions_ids}
\title{Check Divisions of sample IDs only in a datatable}
\usage{
check_divisions_ids(df)
}
\arguments{
\item{df}{A dataframe derived from Rmd which contains sample IDs from which
a new datatable will be created.  This is screened to remove all IDs that
have been previously resampled, disposed, or exported.}
}
\value{
IDs vector.
}
\description{
check and remove all samples IDs that no longer exist in datatable.  Sample
IDs are excluded if they have been resampled, disposed or exported in the
current datatable.
}
\details{
This returns a character vector of sample IDs for use in generating a new
datatable consisting of only EXISTING IDs.
}
