% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{check_divisions_ids_reps}
\alias{check_divisions_ids_reps}
\title{Check Divisions of sample IDs & reps in a datatable}
\usage{
check_divisions_ids_reps(df)
}
\arguments{
\item{df}{A dataframe derived from Rmd which contains sample IDs/reps from
which a new datatable will be created.  This is screened to remove all
IDs/REPs that have been previously resampled, disposed, or exported.}
}
\value{
A list containing the IDs vector and REPs vector - all IDs/REPs that
still EXIST in the df.
}
\description{
check and remove all samples IDs & reps that no longer exist in datatable.
Sample IDs/reps are excluded if they have been resampled, disposed or
exported in the current datatable.
}
\details{
This returns a list of two character vectors - IDs and REPs - for use in
generating a new datatable consisting of only EXISTING IDs/REPs.
}
