% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{check_org_dir}
\alias{check_org_dir}
\title{Check a path is the org root dir}
\usage{
check_org_dir(fileSystemPath)
}
\arguments{
\item{fileSystemPath}{String of a filesystem path.}
}
\description{
Check a path is the org root dir
}
