% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{check_prog_file}
\alias{check_prog_file}
\title{Check Programme File}
\usage{
check_prog_file(fileSystemPath)
}
\description{
Checks fileSystemPath is inside a PROGRAMME directory.
A Programme Directory is a direct sub-dir to a Organisation
directory, so this method establishes this is the case by checking
the parent Dir to fileSystemPath, to see if 'config/' and
'config/templates/' exist.
}
\details{
Secondly, it checks the 'PROJECTS/' directory exists in this
putative Programme Directory.

If a Programme path is confirmed, it returns the fileSystemPath, otherwise
the function returns a BLANK string "".
}
