% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{check_prog_sub_dir}
\alias{check_prog_sub_dir}
\title{Check Programme Sub Dir}
\usage{
check_prog_sub_dir(fileSystemPath)
}
\description{
Searches fileSystemPath's parent directories to identify
a Project Organisation directory.  This is identified by
finding a 'config/' directory and a 'config/templates/' directory.
}
\details{
For the fileSystemPath to be successfully returned, the directory
MUST be at least in level below a PROGRAMME directory.

The original path is returned if identified, otherwise
the function returns a BLANK string "".
}
