% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{check_prog_subdir}
\alias{check_prog_subdir}
\title{Check Programme SubDir}
\usage{
check_prog_subdir(fileSystemPath, settings)
}
\arguments{
\item{fileSystemPath}{Path to possible programme directory}

\item{settings}{List from \code{.config/settings.yml} file.}
}
\description{
Checks fileSystemPath is pointing inside a PROGRAMME directory.
A Programme Directory is a direct sub-dir to a Organisation
directory, so this method establishes this is the case by recursively checking
the parent Dir to fileSystemPath, to see if '.config/' and
'.config/templates/' exist.
}
\details{
Secondly, it checks the 'PROJECTS/' directory exists in this
putative Programme Directory.

If a Programme path is confirmed, it is returned, otherwise
the function returns a BLANK string "".
}
