% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{check_project_note_dir}
\alias{check_project_note_dir}
\title{Check Project Note DIR exists from path}
\usage{
check_project_note_dir(projectNotePath)
}
\description{
From a putative projectNotePath, extract the Project Note DIR name from this (the prefix of the
file name - up to '~_'), check this
}
