% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{check_sample_data_frame}
\alias{check_sample_data_frame}
\title{Check sampleDataFrame}
\usage{
check_sample_data_frame(sampleDataFrame, DATA_DIR)
}
\description{
Checks the sampleDataFrame is a tibble, has attribute 'path', exists in current Rmd's DATA_DIR.
}
