% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{check_section_name_pir}
\alias{check_section_name_pir}
\title{check programme section name & parent path}
\usage{
check_section_name_pir(sectionParentPath, sectionName)
}
\description{
check programme section name & parent path
}
