% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{check_table_name}
\alias{check_table_name}
\title{Check Table Name}
\usage{
check_table_name(table_name, datatables, i, table_start_line)
}
\description{
Ensure \code{table_name} exists in datatables list, otherwise fail gracefully.
}
