% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-rmd-functions.R
\name{check_template_indices}
\alias{check_template_indices}
\title{Check Template Indices}
\usage{
check_template_indices(rmd_contents, rmd_line)
}
\description{
Check Template Indices
}
