% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProgDir.R
\name{checkProgDir}
\alias{checkProgDir}
\title{Check Programme Dir}
\usage{
checkProgDir(fileSystemPath)
}
\description{
Checks fileSystemPath is pointing to a PROGRAMME directory.
A Programme Directory is a direct sub-dir to a Organisation
directory, so this method establishes this is the case by checking
the parent Dir to fileSystemPath, to see if 'config/' and
'config/templates/' exist.
}
\details{
Secondly, it checks the 'PROJECTS/' directory exists in this
putative Programme Directory.

If a Programme path is confirmed, it is returned, otherwise
the function returns a BLANK string "".
}
