% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProgSubDir.R
\name{checkProgSubDir}
\alias{checkProgSubDir}
\title{Check Programme Sub Dir}
\usage{
checkProgSubDir(fileSystemPath)
}
\description{
Searches fileSystemPath's parent directories to identify
a Project Organisation directory.  This is identified by
finding a 'config/' directory and a 'config/templates/' directory.
}
\details{
For the fileSystemPath to be successfully returned, the directory
MUST be at least in level below a PROGRAMME directory.

If an Organisation path is identified, it is returned, otherwise
the function returns a BLANK string "".
}
