% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateProjectOrg.R
\name{checkProjectNote}
\alias{checkProjectNote}
\title{Check Project Note}
\usage{
checkProjectNote(path, todoHeader, goalHeader, delHeader, taskHeader)
}
\description{
Opens Project Note, collects its Project Doc Links, and then compares
the Summary information between the Note and its Doc Links.
}
