% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{checkProjNote}
\alias{checkProjNote}
\title{Check Project Note}
\usage{
checkProjNote(fileSystemPath)
}
\description{
Checks fileSystemPath is inside a sub-dir to a PROGRAMME directory.
A Programme Directory is a direct sub-dir to a Organisation
directory, so this method establishes this is the case by checking
the parent Dir to fileSystemPath, to see if 'config/' and
'config/templates/' exist.
}
\details{
Secondly, it checks the 'PROJECTS/' directory exists in this
putative Programme Directory.

Finally, it confirms the Project Note is of type NOTE.

If a Programme path is confirmed, it returns the fileSystemPath, otherwise
the function returns a BLANK string "".
}
