% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{coloured_print}
\alias{coloured_print}
\title{Print (cat) progress text as coloured text}
\usage{
coloured_print(text, colour = "green")
}
\arguments{
\item{text}{string to print using cat()}

\item{colour}{plain text colour ("red", "green", etc.). Defaults to "green"}
}
\description{
Print (cat) progress text as coloured text
}
\examples{
coloured_print("This is a test", colour = "blue")
}
