% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateProjectOrg.R
\name{compareGoalDelTaskSummary}
\alias{compareGoalDelTaskSummary}
\title{Compare Project Goal/Del/Task Summary}
\usage{
compareGoalDelTaskSummary(projDocLinkList, projNotePath)
}
\description{
Open Project Doc and extract the Goal/Del/Task from it.  Compare its summary information
to the Project Note, and if they do not match, write the summary information in the
Project Note to the Project Doc.
}
