% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{compute_deliverable_link}
\alias{compute_deliverable_link}
\title{compute deliverable link}
\usage{
compute_deliverable_link(del, DocLink, settings)
}
\arguments{
\item{del}{Deliverable Header String}

\item{DocLink}{Link to Project Doc.}

\item{settings}{projectmanagr settings list.}
}
\description{
generate new link from project doc with html # tag for deliverable section.
}
