% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{compute_doc_GDT_link}
\alias{compute_doc_GDT_link}
\title{Compute Project Doc Goal Del Task Link}
\usage{
compute_doc_GDT_link(
  projectDocPath,
  projNoteRmdPath,
  settings,
  goal,
  deliverable,
  task
)
}
\description{
Compute the components that form the Project Doc Goal Del Task Link, for
insertion into a linked Project Note.
}
