% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{compute_goal_link}
\alias{compute_goal_link}
\title{compute goal link}
\usage{
compute_goal_link(goal, DocLink, settings)
}
\description{
generate new link from project doc with html # tag for goal section.
}
