% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{compute_project_index}
\alias{compute_project_index}
\title{Compute the next available project index}
\usage{
compute_project_index(projsPath, programmePrefix)
}
\description{
Returns first result from compute_project_indices, appending a 0 if
the returned number is below 10.
}
