% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{compute_project_indices}
\alias{compute_project_indices}
\title{Compute Project Indices}
\usage{
compute_project_indices(projsPath, programmePrefix)
}
\description{
From the PROJECTS dir in a programme (\code{projsPath}), compute the available indices
for creating new project doc and DIR, based on all DIRs in \code{projsPath} that start
with \code{programmePrefix} and return a numeric array of these.
}
