% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{compute_task_link}
\alias{compute_task_link}
\title{compute deliverable link}
\usage{
compute_task_link(task, DocLink, settings)
}
\description{
generate new link from project doc with html # tag for deliverable section.
}
