% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeHeaderLineUnderProjectDocGoalDelTask}
\alias{computeHeaderLineUnderProjectDocGoalDelTask}
\title{Compute Next Line under Header Links, in Project Doc Goal Del Task}
\usage{
computeHeaderLineUnderProjectDocGoalDelTask(headerName, projDocContents,
  GOAL, DELIVERABLE, TASK)
}
\arguments{
\item{headerName}{name of header note to search for}

\item{projDocContents}{character vector containing the Project Doc}

\item{GOAL}{the GOAL character string to find in projDocContents}

\item{DELIVERABLE}{the DELIVERABLE character string to find in projDocContents}

\item{TASK}{the TASK character string to find in projDocContents}
}
\description{
In projDocContents, sequentially identifies lines containing GOAL, DELIVERABLE,
TASK, headerName.  Then identifies the next line AFTER all subnote links under the headerName
link.  This returned index is where the NEXT subnote should be inserted.
}
