% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeLastLineIndex}
\alias{computeLastLineIndex}
\title{Compute Last Line Index}
\usage{
computeLastLineIndex(contents)
}
\description{
Returns the index of the LAST Line in contents that contains any content ("[A-z,0-9]")
}
\details{
contents - a vector of strings, the index of the first instance of line in this vector is returned.
}
