% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeLineIndex}
\alias{computeLineIndex}
\title{Compute Line Index}
\usage{
computeLineIndex(line, contents)
}
\description{
Returns the index of the first EXACT MATCH of line in contents, if not found returns -1.
}
\details{
line - a String to be found

contents - a vector of strings, the index of the first instance of line in this vector is returned.
}
