% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNextLine.R
\name{computeNextLine}
\alias{computeNextLine}
\title{Returns the first blank line after the last content under a Task
in a Project Document,}
\usage{
computeNextLine(line, projDocContents)
}
\description{
Returns the first blank line after the last content under a Task
in a Project Document,
}
