% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeNextLineHeader}
\alias{computeNextLineHeader}
\title{Compute Next Line Header}
\usage{
computeNextLineHeader(line, projDocContents)
}
\description{
Returns the first blank line after the last Project SubNote in a Project Group, under a Task
in a Project Document.
}
