% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeNextLineIndex}
\alias{computeNextLineIndex}
\title{Compute Next Line Index}
\usage{
computeNextLineIndex(lineIndex, contents)
}
\arguments{
\item{lineIndex}{the index to start looking from.}

\item{contents}{a vector of strings, the index of the first instance of any content in this vector is returned.}
}
\description{
Returns the index of the Next Line in contents that contains any content ("[A-z,0-9]"), looking
from lineIndex to length of contents.
}
