% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{confirm_dir}
\alias{confirm_dir}
\title{Confirm a Single Directory Exists}
\usage{
confirm_dir(dirPath, errMsg, orgPath)
}
\arguments{
\item{dirPath}{Character. Path to the directory to check.}

\item{errMsg}{Character. Error message to display if the directory does not
exist.}

\item{orgPath}{Character. The root path of the organisation, which will be
removed if the check fails.}
}
\description{
Checks if a given directory exists. If not, deletes the organisation and
raises an error.
}
