% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{confirm_dirs}
\alias{confirm_dirs}
\title{Confirm Directories Exist}
\usage{
confirm_dirs(dirPaths, errMsg, orgPath)
}
\arguments{
\item{dirPaths}{Character vector. A list of directory paths to check.}

\item{errMsg}{Character. Error message to display if a directory does not exist.}

\item{orgPath}{Character. The root path of the organisation, which will be removed if any check fails.}
}
\description{
Checks if all directories in the given list exist. If any are missing, the
organisation directory is deleted and an error is raised.
}
