% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{confirm_file}
\alias{confirm_file}
\title{Confirm a File Exists}
\usage{
confirm_file(filePath, errMsg, orgPath)
}
\arguments{
\item{filePath}{Character. Path to the file to check.}

\item{errMsg}{Character. Error message to display if the file does not exist.}

\item{orgPath}{Character. The root path of the organisation, which will be
removed if the check fails.}
}
\description{
Checks if a given file exists. If not, deletes the organisation and raises an
error.
}
