% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{confirm_rmd_path}
\alias{confirm_rmd_path}
\title{Path in org && Project Doc or Note (subdir of programme)
&& absolute + normalised}
\usage{
confirm_rmd_path(rmd_path)
}
\arguments{
\item{rmd_path}{Path to Rmd file to check.}
}
\description{
Path in org && Project Doc or Note (subdir of programme)
&& absolute + normalised
}
