% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{correct_data_vals}
\alias{correct_data_vals}
\title{ensure data vals match data_cols length for add variables datatable}
\usage{
correct_data_vals(default_data_vals, data_cols)
}
\description{
ensure data vals match data_cols length for add variables datatable
}
