% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_addins_json}
\alias{create_addins_json}
\title{Create Addins JSON File}
\usage{
create_addins_json(projectmanagrPath, orgPath, settings)
}
\arguments{
\item{projectmanagrPath}{Character. Path to the projectmanagr package.}

\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings containing file structure configurations.}
}
\value{
Character. The path to the created addins.json file.
}
\description{
Copies the default addins.json file from the projectmanagr package into the organisation's config directory.
}
