% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_config_path}
\alias{create_config_path}
\title{Create Configuration Directory}
\usage{
create_config_path(orgPath, settings)
}
\arguments{
\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Settings containing the configuration directory structure.}
}
\value{
Character. The path to the created configuration directory.
}
\description{
Creates a directory for storing configuration files.
}
