% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_content}
\alias{create_content}
\title{Create Insertable Content in a Project Note}
\usage{
create_content(
  selection,
  contentName,
  contentDescription,
  contentSourcePath,
  contentTitle = "",
  contentDeclarationTemplate = "Content-Declaration-Template.Rmd",
  contentSourceTemplate = "Content-Source-Template.Rmd"
)
}
\arguments{
\item{selection}{A list representing the current Project Note selection.
It must include:
- filePath: The path of the current Project Note Rmd.
- originalLineNumber: The line number where content is to be inserted.
- rmdType: The type of Rmd file; must be "NOTE" or "SUB".}

\item{contentName}{A string specifying the name of the content.
It must not contain any spaces.}

\item{contentDescription}{A string describing the content. This is shown
to users when choosing content to insert.}

\item{contentSourcePath}{A string specifying the path where the content
directory and its Rmd file will be created. It is recommended to place
it within the Project Note's directory.}

\item{contentDeclarationTemplate}{A template file for content declaration,
located in the .config/templates/ directory. Default is
"Protocol-Declaration-Template.Rmd".}

\item{contentSourceTemplate}{A template file for the content source, found
in the .config/templates/ directory. Default is
"Protocol-Source-Template.Rmd".}

\item{projectNoteTitle}{(Optional) A title for the content. If left blank,
contentName is used with underscores and dashes replaced by spaces.}
}
\value{
A string representing the path to the created content Rmd file.
}
\description{
This function declares new insertable content based on a content
declaration template and inserts it into a selected Project Note.
It creates a separate Rmd file (the content source) that initially is blank,
which can be modified to include specific documentation such as SOPs or
protocol logs.
}
\details{
The function executes the following steps:
\enumerate{
\item Determines the content directory and Rmd file path using contentName and
contentSourcePath.
\item Creates a relative hyperlink from the Project Note to the new content Rmd.
\item Reads the content declaration and content source templates from the
templates directory.
\item Replaces placeholders in the content declaration template with the
provided content title, description, and the hyperlink.
\item Inserts the modified content declaration into the Project Note at the
specified insertion point.
\item Creates the content directory and writes the content source template to
a new Rmd file.
\item Updates the Project Note with the new content declaration.
}
}
\note{
The function assumes that the current file in the selection is a Project Note.
The selection must include keys 'filePath' and 'originalLineNumber'. The
contentName must not have spaces, and the content declaration template must
define placeholders for CONTENT_TITLE_FIELD, CONTENT_TITLE,
CONTENT_DESCRIPTION_FIELD, CONTENT_DESCRIPTION, CONTENT_SOURCE_FIELD, and
CONTENT_SOURCE_LINK.
}
\examples{
\dontrun{
  create_content(
    selection = list(
      filePath = "project_note.Rmd",
      originalLineNumber = 25,
      rmdType = "NOTE"
    ),
    contentName = "SOP_Fix",
    contentDescription = "Standard operating procedure for fixing the pump.",
    contentSourcePath = "/projects/my_project/notes",
    projectNoteTitle = "Pump Fix SOP",
    contentDeclarationTemplate = "Protocol-Declaration-Template.Rmd",
    contentSourceTemplate = "Protocol-Source-Template.Rmd"
  )
}

}
\seealso{
\code{\link{create_hyperlink}}, \code{\link{read_file}},
\code{\link{write_file}}, \code{\link{get_settings_yml}},
\code{\link{find_org_directory}}, \code{\link{insert_at_indices}}
}
