% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_html_site_path}
\alias{create_html_site_path}
\title{Create HTML Site Directory}
\usage{
create_html_site_path(orgPath, settings)
}
\arguments{
\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Settings containing the site directory structure.}
}
\value{
Character. The path to the created site directory.
}
\description{
Creates a directory for compiled HTML output.
}
