% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlink-functions.R
\name{create_hyperlink_no_ext}
\alias{create_hyperlink_no_ext}
\title{Create Hyperlink Without File Extension in Link Text}
\usage{
create_hyperlink_no_ext(toFilePath, fromFilePath)
}
\arguments{
\item{toFilePath}{Character. Full or relative path to the \strong{target} \code{.Rmd} file.}

\item{fromFilePath}{Character. Full or relative path to the \strong{source} \code{.Rmd} file
from which the relative link should be computed.}
}
\value{
A character string containing a properly formatted Markdown hyperlink,
using the file name without extension as the link text.
}
\description{
Generates a Markdown-formatted hyperlink pointing from one \code{.Rmd} file
(\code{fromFilePath}) to another (\code{toFilePath}), using the \strong{base name without file
extension} as the link text.
}
\details{
This is a convenience wrapper for \code{\link{create_hyperlink}} used when the
display name of the link should omit \code{.Rmd} or other file extensions.
}
