% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{create_hyperlink_section}
\alias{create_hyperlink_section}
\title{Create Hyperlink Section}
\usage{
create_hyperlink_section(toFileName, toFileSection, toFilePath, fromFilePath)
}
\description{
creates string for hyperlink in Rmd, using `toFileName : toFileSection` as
the hyperlink text, and generating a RELATIVE LINK to
`toFilePath#toFileSection` from `fromFilePath`.
}
