% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{create_no_selection}
\alias{create_no_selection}
\title{Create No Selection}
\usage{
create_no_selection(rmdType, errorMessage, filePath)
}
\description{
Selection object that signifies No Selection, with error message.
}
\details{
Consists of a list with first item the string "FALSE", and second item
an `errorMessage`.
}
