% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_org_dir}
\alias{create_org_dir}
\title{Create Organisation Directory}
\usage{
create_org_dir(orgPath)
}
\arguments{
\item{orgPath}{Character. The full path to the organisation directory.}
}
\description{
Creates the root directory for the organisation if it does not exist.
}
