% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_org_index}
\alias{create_org_index}
\title{Create Organisation Index File}
\usage{
create_org_index(
  orgName,
  tempPath,
  orgTemplate,
  orgPath,
  settings,
  orgTitle,
  authorValue = Sys.info()["user"]
)
}
\arguments{
\item{orgName}{Character. Name of the organisation.}

\item{tempPath}{Character. Path to the templates directory.}

\item{orgTemplate}{Character. Name of the template file.}

\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings used for configuring the index file.}

\item{orgTitle}{Character. Title of the organisation.}

\item{authorValue}{Character. Author of the organisation. Defaults to the system username.}
}
\value{
Character. The path to the created organisation index file.
}
\description{
Generates the index Rmd file for the organisation based on a template.
}
