% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_prog_index}
\alias{create_prog_index}
\title{Create Programme Index Rmd}
\usage{
create_prog_index(
  progPath,
  programmeTitle,
  tempPath,
  progTemplate,
  authorValue,
  orgPath,
  settings
)
}
\description{
Fill with contents, return contents
}
