% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_programme_section}
\alias{create_programme_section}
\title{Create Programme Section}
\usage{
create_programme_section(
  sectionName,
  sectionParentPath,
  authorValue = Sys.info()["user"],
  sectionTitle = "",
  sectTemplate = "Programme-Section-Template.Rmd",
  sectSummaryTemplate = "Programme-Section-Summary-Template.Rmd"
)
}
\arguments{
\item{sectionName}{Character. The name of the new programme section. Must not
contain spaces.}

\item{sectionParentPath}{Character. The file path to an existing organisation
directory, within a programme.}

\item{authorValue}{Character. The author of the section. Defaults to system
username.}

\item{sectionTitle}{Character. The title of the programme. If empty, it will
be derived from \code{programmeName}.}

\item{sectTemplate}{Character. The filename of the programme template. Default
is \code{"Programme-Section-Template.Rmd"}.}

\item{sectSummaryTemplate}{Character. The filename of the programme summary
template. Default is \code{"Programme-Section-Summary-Template.Rmd"}.}
}
\value{
Character. The file path of the created programme index Rmd file.
}
\description{
Generates a new programme section within an existing programme directory,
creating the necessary directory structure, R Markdown index file, and
linking to its programme or parent section index file.
}
\details{
This function performs the following steps:
\enumerate{
\item Ensures \code{sectionName} does not contain spaces.
\item Identifies the root organisation directory from \code{sectionParentPath}.
\item Creates a new programme sdction directory within the organisation.
\item Generates a new programme section index Rmd file from a template.
\item Links and inserts the programme section into its parent's index file.
}

The function ensures that all directories and files are properly created and
linked within the project structure.
}
\note{
\itemize{
\item The \code{sectionName} should not contain spaces.
\item The function modifies files directly on disk. Ensure backups are created
before running.
\item The function relies on a structured organisation directory; incorrect
structures may cause failures.
}
}
\examples{
\dontrun{
# Create a new programme in an organisation
create_programme_section(
  programmeName = "programme-section",
  organisationPath = "/path/to/organisation/programme",
  authorValue = "sjwest",
  programmeTitle = "Programme Section",
  progTemplate = "Programme-Section-Template.Rmd",
  progSummaryTemplate = "Programme-Section-Summary-Template.Rmd"
)
}

}
\seealso{
\itemize{
\item \code{\link{find_org_directory}} to locate the organisation directory.
\item \code{\link{get_settings_yml}} for retrieving configuration settings.
\item \code{\link{write_file}} for writing content to files.
}
}
