% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_sect_index}
\alias{create_sect_index}
\title{Create Programme Section Index Rmd}
\usage{
create_sect_index(
  sectPath,
  sectionTitle,
  parentIndex,
  tempPath,
  sectTemplate,
  authorValue,
  orgPath,
  settings
)
}
\description{
Fill with contents, return contents
}
