% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{create_selection}
\alias{create_selection}
\title{Create Selection}
\usage{
create_selection(filePath, contents, line, settings)
}
\description{
Creates a selection list generated from a filePath, its contents, and an
existing line.  The selection list consists of a `rmdType` named vector that
defines what kind of file is in the current selection: `DOC` `HEAD` `SUB`
`NOTE` or `FALSE`.
}
\details{
If a Project Document (`DOC`) is selected, the returned list contains several
other useful parameters:

From project doc filepath contents and selection line.
}
