% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_settings_yaml_file}
\alias{create_settings_yaml_file}
\title{Create Settings YAML File}
\usage{
create_settings_yaml_file(
  orgPath,
  settings,
  utime,
  orgPaths,
  orgName,
  orgTitle,
  sitePath
)
}
\arguments{
\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings to be saved in the YAML file.}

\item{utime}{Unused.}

\item{orgPaths}{Character vector. List of existing organisation paths.}

\item{orgName}{Character. Name of the organisation.}

\item{orgTitle}{Character. Title of the organisation.}

\item{sitePath}{Character. Path to the site directory.}
}
\value{
Character. Path to the created settings YAML file.
}
\description{
Generates a settings YAML file for the organisation.
}
