% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_status_yaml_content}
\alias{create_status_yaml_content}
\title{Generate Status YAML Content}
\usage{
create_status_yaml_content(
  orgPaths,
  orgPath,
  orgName,
  orgTitle,
  updateTime,
  sitePath
)
}
\arguments{
\item{orgPaths}{Character vector. List of existing organisation paths.}

\item{orgPath}{Character. Path to the new organisation.}

\item{orgName}{Character. Name of the organisation.}

\item{orgTitle}{Character. Title of the organisation.}

\item{updateTime}{Character. Timestamp of the last update.}

\item{sitePath}{Character. Path to the site directory.}
}
\value{
List. A named list containing metadata for the organisation.
}
\description{
Creates a structured list to be written into a status YAML file.
}
