% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_status_yaml_file}
\alias{create_status_yaml_file}
\title{Create Status YAML File}
\usage{
create_status_yaml_file(
  orgPath,
  settings,
  orgPaths,
  orgName,
  orgTitle,
  sitePath
)
}
\arguments{
\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings used to generate the status file.}

\item{orgPaths}{Character vector. List of existing organisation paths.}

\item{orgName}{Character. Name of the organisation.}

\item{orgTitle}{Character. Title of the organisation.}

\item{sitePath}{Character. Path to the site directory.}
}
\value{
Character. Path to the created status YAML file.
}
\description{
Generates a status YAML file for the organisation, storing metadata and
update timestamps.
}
