% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_sub_note}
\alias{create_sub_note}
\title{Create a New Sub Note within a Project Group}
\usage{
create_sub_note(
  subNoteName,
  subNotePath,
  selection,
  authorValue = get_username(),
  subNoteTitle = "",
  subNoteTemplate = "Project-Sub-Note-Template.Rmd",
  headerNoteContentLinkTemplate = "Project-Header-Note-Content-Link-Template.Rmd",
  subNoteContentLinkTemplate = "Project-Sub-Note-Content-Link-Template.Rmd",
  projNoteLinkTemplate = "Project-Note-Link-Template.Rmd",
  projNoteLinkSummaryTemplate = "Project-Note-Link-Summary-Template.Rmd",
  projNoteSummaryTemplate = "Project-Note-Summary-Template.Rmd",
  subNoteSummaryTemplate = "Project-Sub-Note-Summary-Template.Rmd"
)
}
\arguments{
\item{subNoteName}{A string representing the name of the Project Sub Note.
Must not contain spaces.}

\item{subNotePath}{A character string specifying the absolute directory
where the Sub Note will be stored. It must reside within the Project
Group Note directory.}

\item{selection}{A list with metadata about the current selection. The list
must include the following keys:
\itemize{
\item rmdType: A string specifying the file type. Valid values are "DOC",
"HEAD", or "SUB".
\item filePath: The file path of the associated project document (or note).
\item headerNoteLink (if rmdType == "DOC"): A relative link to the header note
from the doc text.
\item goal, deliverable, task, taskLine: GDT info for linking in the doc.
\item addingSubNote (if rmdType == "DOC"): Must be TRUE for a valid sub-note
insertion.
}}

\item{authorValue}{A string representing the author's name for the note.
Defaults to the current system user.}

\item{subNoteTitle}{(Optional) A title for the Sub Note. If blank, the
subNoteName is used with underscores and dashes replaced by spaces.}

\item{subNoteTemplate}{A template file used to generate the Sub Note.
Default is "Project-Sub-Note-Template.Rmd".}

\item{headerNoteContentLinkTemplate}{A template file for creating a content
link from the header note to the Sub Note. Default is
"Project-Header-Note-Content-Link-Template.Rmd".}

\item{subNoteContentLinkTemplate}{A template file for creating a content link
within the Sub Note. Default is "Project-Sub-Note-Content-Link-Template.Rmd".}

\item{projNoteLinkTemplate}{A template file that defines the structure of the
Project Doc Goal/Del/Task link in the Sub Note. Default is
"Project-Note-Link-Template.Rmd".}

\item{projNoteLinkSummaryTemplate}{A template file that adds a summary and
to-do section beneath the Project Doc link in the Sub Note. Default is
"Project-Note-Link-Summary-Template.Rmd".}

\item{projNoteSummaryTemplate}{A template file used to add a Project Note
summary to the Project Doc. Default is "Project-Note-Summary-Template.Rmd".}

\item{subNoteSummaryTemplate}{A template file used to add a Sub Note summary
to the Project Doc. Default is "Project-Sub-Note-Summary-Template.Rmd".}

\item{todoTemplate}{A template file for the to-do list section. Default is
"Todo-Template.Rmd".}
}
\description{
This function creates a new Sub Note within a Project Group. It reads the
necessary template files, creates the required sub-directory and R Markdown
file for the Sub Note, replaces template markup with provided values, and
updates the Project Group Header Note with a link to the new Sub Note. It
then calls link_sub_note_doc() to insert all relevant links and summaries into
the main project document.
}
\details{
This function performs the following steps:
\enumerate{
\item Validates that subNoteName contains no spaces.
\item Checks that the selection list is valid. The selection must include
keys such as rmdType, filePath, headerNoteLink, goal, deliverable, task,
and taskLine; if rmdType is "DOC", addingSubNote must be TRUE.
\item Determines the header note Rmd path (and link note path) based on the
selection.
\item Reads the required template files for the Sub Note and for header note
content links.
\item Creates a sub-directory for the Sub Note and an associated R Markdown file.
\item Replaces the markup in the Sub Note with provided parameters, then inserts
a link to the header note.
\item Updates the header note, inserting a link back to the new Sub Note.
\item Calls link_sub_note_doc() to update the project document with all relevant
Sub Note links and summaries.
}
}
\seealso{
link_sub_note_doc, insert_header_link_subnote,
insert_subnote_link_header, create_note_directory, create_note_rmd_file
}
