% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_templates_path}
\alias{create_templates_path}
\title{Create Templates Directory}
\usage{
create_templates_path(projectmanagrPath, orgPath, settings)
}
\arguments{
\item{projectmanagrPath}{Character. Path to the projectmanagr package.}

\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings containing file structure configurations.}
}
\value{
Character. The path to the created templates directory.
}
\description{
Creates a directory for storing templates and copies template files from the
package.
}
