% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_volumes_file}
\alias{create_volumes_file}
\title{Create Volumes File}
\usage{
create_volumes_file(volumesPath, projectmanagrPath, orgPath, settings)
}
\arguments{
\item{volumesPath}{Character. Path to the volumes directory.}

\item{projectmanagrPath}{Character. Path to the projectmanagr package.}

\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Project settings containing file structure configurations.}
}
\value{
Character. The path to the created volumes file.
}
\description{
Copies the default volumes file template into the organisation's volumes
directory.
}
