% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_volumes_path}
\alias{create_volumes_path}
\title{Create Volumes Directory}
\usage{
create_volumes_path(orgPath, settings)
}
\arguments{
\item{orgPath}{Character. Path to the organisation directory.}

\item{settings}{List. Settings containing the volumes directory structure.}
}
\value{
Character. The path to the created volumes directory.
}
\description{
Creates a directory for storing data volumes.
}
