% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProgramme.R
\name{createProgramme}
\alias{createProgramme}
\title{Create a New Programme}
\usage{
createProgramme(programmeName, programmePrefix, programmeTitle = "",
  fileSystemPath = getwd())
}
\description{
Generates a new Programme at the top level of the Organisation.  If the
fileSystemPath is not at the top of the Organisation, will traverse until
it is found.  User must supply the programmeName and programmePrefix.  The
programmeName must NOT contain a space, an optional programmeTitle (which
if not supplied will default to the programmeName, replacing "_" & "-" with
" ").  The default fileSystemPath is the working directory.
}
