% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectDoc.R
\name{createProjectDoc}
\alias{createProjectDoc}
\title{Create a New Project}
\usage{
createProjectDoc(projectName, projectTitle = "",
  fileSystemPath = getwd(),
  projDocTemplate = "Project-Doc-Template.Rmd", projectIndex = 0)
}
\description{
Generates a new Project inside a Programme in an Organisation.  The fileSystemPath
must be in the Programme directory, which is the sub-Dir to the root ORGANISATION
dir.
}
\details{
The project is automatically numbered, by reading the current projects and determining
the next number in the sequence, as well as deriving the Programme Prefix.

User must supply the project name, which must contain NO SPACES, and optionally a project
title.  If no title is provided, the Title is derived from the project name, replacing any
"_" & "-" with " ". The default fileSystemPath is the working directory.
}
