% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectOrg.R
\name{createProjectOrg}
\alias{createProjectOrg}
\title{Create a New Project Organisation}
\usage{
createProjectOrg(orgName = "00_ORG", orgTitle = "ORGANISATION",
  fileSystemPath = getwd())
}
\description{
Generates the layout for a new Project Organisation in the File System,
including the content, in '00_ORG' directory by default.  The default location is the working directory.
}
\details{
This will also initialise the config/ and templates/ directories for this Organisation.  The config/
directory will contain the status.yml and settings.yml files.  status.yml will contain data on the
Organisation's status (programmes, projects and project work in progress, update times, etc).  settings.yml
will contain the settings needed for various projectmanagr functions.

The templates/ directory contains several default templates for Project Docs and Notes.

A directory is also created to store the compiled html site from the Organisation - site/
}
