% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{createVolume}
\alias{createVolume}
\title{Create Volume}
\usage{
createVolume(sourcePath, volName, orgPath = getwd())
}
\arguments{
\item{sourcePath}{Path to the source directory. If the dir at sourcePath
does not exist, this function will terminate.}

\item{volName}{The naem of the new volume. If a volume called volName already
exists, it is not overwritten, and this function will terminate.}

\item{orgPath}{Optional argument to specify the location of the organisation with
orgPath, which by default is the current working directory.}
}
\description{
Create a new Volume in the volumes/ dir of the projectmanagr
Organisation.  The function Symlinks the dir at sourcePath to a new
Symlink made in the Organisation volumes/ directory, with
name volName.
}
