% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{cursor_selection}
\alias{cursor_selection}
\title{Document Cursor Selection}
\usage{
cursor_selection(
  save_context_fn = .save_context_doc,
  get_context_fn = .get_source_editor_context
)
}
\description{
Identify the selection from the Active Cursor position in current R Studio file.
}
\details{
First identifies if the current Cursor position is on a GROUP (HEADER NOTE or SUBNOTE).
It then searches back through the Active Document to find the first line starting with
"##" - if this is a \code{TASK} line, then the first previous \code{DELIVERABLE} line and
the first previous \code{GOAL} line are also identified.

If TASK, DELIVERABLE and GOAL lines are all successfully found, this method returns
a LIST:

[\link{1}] or [\link{"rmdType"}] - type of file selected: DOC, HEAD, SUB, NOTE, or UNKNOWN

[\link{2}] or [\link{"task"}]

[\link{3}] or [\link{"taskLine"}]

[\link{4}] or [\link{"deliverable"}]

[\link{5}] or [\link{"delLine"}]

[\link{6}] or [\link{"goal"}]

[\link{7}] or [\link{"goalLine"}]

[\link{8}] or [\link{"originalLine"}]

[\link{9}] or [\link{"originalLineNumber"}]

[\link{10}] or [\link{"addingSubNote"}]

[\link{11}] or [\link{"headerNoteLink"}]

[\link{12}] or [\link{"headerNoteRmdPath"}]

[\link{13}] or [\link{"headerNoteName"}]

[\link{14}] or [\link{"headerNoteLineNumber"}]

[\link{15}] or [\link{"filePath"}]

[\link{16}] or [\link{"defaultPath"}]

[\link{1}] or [\link{"task"}] - contains the line that starts "#### TASK" in the Active Document.

[\link{2}] or [\link{"taskLine"}] - the task line number (index) in the Active Document.

[\link{3}] or [\link{"deliverable"}] - contains the line that starts "### DELIVERABLE" in the Active Document.

[\link{4}] or [\link{"deliverableLine"}] - the deliverable line number (index) in the Active Document.

[\link{5}] or [\link{"goal"}] - contains the line that starts "## GOAL" in the Active Document.

[\link{6}] or [\link{"goalLine"}] - the goal line number (index) in the Active Document.

[\link{7}] or [\link{"originalLine"}] - the original line where the Active Cursor was on in the Active Document.

[\link{8}] or [\link{"originalLineNumber"}] - the original line number where the Active Cursor was on in the Active Document.

[\link{9}] or [\link{"addingSubNote"}] - a Boolean value indicating the Cursor was on a GROUP (HEADER NOTE or SUBNOTE).

[\link{10}] or [\link{"headerNoteLink"}] - if addingSubNote is TRUE, contains the link to the headerSubNote, AS IS GIVEN
IN THE PROJECT DO (i.e. the link TITLE and then the link ADDRESS - for example
\strong{\href{../LAB/LAB~001-00~_GARS_SC_Clearing_Labelling.Rmd}{LAB~001-00~ GARS SC Clearing Labelling}}), else is BLANK

[\link{11}] or [\link{"headerNoteLineNumber"}] - if addingSubNote is TRUE, contains the line that contains the headerSubNoteLink, else is BLANK.

[\link{12}] or [\link{"projectDocPath"}] - the path of the Active Doc (a Project Doc, as successfully retrieved Task/Del/Goal)

If this method is unsuccessful (there is an error), it returns a LIST:

[\link{1}] - contains the String "FALSE"
[\link{2}] - contains the errorMessage - a String indicating why the method failed.
}
