% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{cursorSelection}
\alias{cursorSelection}
\title{Document Cursor Selection}
\usage{
cursorSelection()
}
\description{
Identify the selection from the Active Cursor position in current R Studio file.
}
\details{
First identifies if the current Cursor position is on a GROUP (HEADER NOTE or SUBNOTE).
It then searches back through the Active Document to find the first line starting with
"##" - if this is a "#### TASK" line, then the first previous "### DELIVERABLE" line and
the first previous "### GOAL" line is also identified.

If TASK, DELIVERABLE and GOAL lines are all successfully found, this method returns
a LIST:

[[1]] or [["task"]] - contains the line that starts "#### TASK" in the Active Document.

[[2]] or [["taskLine"]] - the task line number (index) in the Active Document.

[[3]] or [["deliverable"]] - contains the line that starts "### DELIVERABLE" in the Active Document.

[[4]] or [["deliverableLine"]] - the deliverable line number (index) in the Active Document.

[[5]] or [["goal"]] - contains the line that starts "## GOAL" in the Active Document.

[[6]] or [["goalLine"]] - the goal line number (index) in the Active Document.

[[7]] or [["originalLine"]] - the original line where the Active Cursor was on in the Active Document.

[[8]] or [["originalLineNumber"]] - the original line number where the Active Cursor was on in the Active Document.

[[9]] or [["addingSubNote"]] - a Boolean value indicating the Cursor was on a GROUP (HEADER NOTE or SUBNOTE).

[[10]] or [["headerNoteLink"]] - if addingSubNote is TRUE, contains the link to the headerSubNote, AS IS GIVEN
IN THE PROJECT DO (i.e. the link TITLE and then the link ADDRESS - for example
**[LAB~001-00~ GARS SC Clearing Labelling](../LAB/LAB~001-00~_GARS_SC_Clearing_Labelling.Rmd)**), else is BLANK

[[11]] or [["headerNoteLineNumber"]] - if addingSubNote is TRUE, contains the line that contains the headerSubNoteLink, else is BLANK.

[[12]] or [["projectDocPath"]] - the path of the Active Doc (a Project Doc, as successfully retrieved Task/Del/Goal)


If this method is unsuccessful (there is an error), it returns a LIST:

[[1]] - contains the String "FALSE"
[[2]] - contains the errorMessage - a String indicating why the method failed.
}
