% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{database_create}
\alias{database_create}
\title{Create a Database}
\usage{
database_create(filename, columns)
}
\arguments{
\item{filename}{the filename of the database to load.  Do NOT add .csv extension!}

\item{columns}{A character vector containing a list of columns to add to this database.}
}
\value{
The loaded TIBBLE with the "path" attribute set to the database filename.
}
\description{
A Database is a tibble (stored on disk as a CSV file) with a predefined
set of columns for storing items.
}
\details{
To create a database, a filename String and a columns character vector
must be provided.  An empty tibble is formed and returned from this
function: it contains the specified columns, and this tibble contains
a 'path' attribute, that indicates where it is stored RELATIVE TO THE
CURRENT RMD FILE, and this can be retrieved using `attr(tibble, "path")`.

The database is stored alongside the Rmd that calls it in the file
system - therefore this function can only be called from an Rmd, and
the Rmd must be saved to disk.
}
