% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{database_edit_row}
\alias{database_edit_row}
\title{Edit row in database}
\usage{
database_edit_row(database, ID, data)
}
\description{
Overwrite the line in database with parsed ID with content of data.
}
