% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{database_load}
\alias{database_load}
\title{Load Database}
\usage{
database_load(relativepath)
}
\arguments{
\item{relativepath}{the relative path from the Rmd to the dataset CSV to load.  Do NOT
add .csv extension!}
}
\value{
The loaded TIBBLE with the "path" attribute set to the sample filename.
}
\description{
Load a database located at the relativepath from the current Rmd.
}
