% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-rmd-functions.R
\name{datatable_add_data_samples_template_rmd}
\alias{datatable_add_data_samples_template_rmd}
\title{Add Data to Sample DataTable from TEMPLATE and insert into Rmd}
\usage{
datatable_add_data_samples_template_rmd(
  rmd_path,
  rmd_line,
  datatable_name = "samples",
  dt_length = 100,
  summarise_reps = FALSE,
  all_reps = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{line in Rmd file inside the TEMPLATE datatable}

\item{datatable_name}{String of data table name: IDs are sourced from this
datatable, and is used as the name of the newly created datatable.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}

\item{all_reps}{Boolean to indicate whether all reps should be summarised in
the datatable.  If TRUE then ALL reps for each ID are summarised with keyword
ALL - which will apply data cols to all the reps for each sample, without the
potentially complex summary strings representing a disparate collection of reps.
This comes at the expense of not being clear from reading this datatable what
reps actually exist.}

\item{IDs}{Character vector of sample IDs}
}
\description{
Creates a Sample DataTable in specified Rmd file from the template that
should be present between specified start & end line.
}
\details{
The template Sample DataTable will contain an initial ID column containing
a default IDs vector (which will be replaced with IDs from selected existing
datatable), and any other columns will have their initial data
value replicated along the length of the datatable.
}
