% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-functions.R
\name{datatable_add_group_rmd}
\alias{datatable_add_group_rmd}
\title{Generate GROUP Data Table and insert into Rmd}
\usage{
datatable_add_group_rmd(
  rmd_path,
  rmd_startline,
  rmd_endline,
  datatable_name,
  group_names,
  groups,
  dt_length = 100,
  summarise_reps = FALSE,
  all_reps = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_startline}{start line in Rmd file to insert table}

\item{rmd_endline}{end line in Rmd file to insert table - same as
\code{rmd_startline} if a single line is selected, otherwise the selected lines
(between start & end lines) will trim all lines from first COMMENT TEXT
line - beginning with \verb{>>>>}.  Used to parse default group names.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.}

\item{group_names}{Character vector of data table column titles to add.
CANNOT BE BLANK!}

\item{groups}{A list of vectors containing the group labels for each group
defined in group_names.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}

\item{all_reps}{Boolean to indicate whether all reps should be summarised in
the datatable.  If TRUE then ALL reps for each ID are summarised with keyword
ALL - which will apply data cols to all the reps for each sample, without the
potentially complex summary strings representing a disparate collection of reps.
This comes at the expense of not being clear from reading this datatable what
reps actually exist.}
}
\description{
Adds Groups to an EXISTING Data Table in specified Rmd file at specified line.
The Groups DataTable will contain an initial ID column containing all
sample IDs, and a series of extra data columns named using the
group_names vector.
}
\details{
If the table exceeds dt_length characters
(default 100), then the table is split into multiple tables, with IDs as
first col, and subsequent group_names given in subsequent tables.
}
