% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_add_group_template}
\alias{datatable_add_group_template}
\title{Add Group to Sample Data Table from Template}
\usage{
datatable_add_group_template(
  IDs,
  template_table,
  datatable_name,
  dt_length = 100
)
}
\arguments{
\item{IDs}{Character vector of sample IDs}

\item{template_table}{A character vector of the template table, as presented
in the Rmd file.}

\item{datatable_name}{String of data table name.}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
Creates add dataSample DataTable vector. The Sample DataTable will contain an
initial ID column containing the IDs vector, and set of extra group
data columns as specified in the data_cols vector.  If the table exceeds
\code{dt_length} characters (default 100), then the table is split into multiple
tables, with IDs as first col, and subsequent group data_cols given in
subsequent tables.
}
