% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-functions.R
\name{datatable_create_rmd}
\alias{datatable_create_rmd}
\title{Generate new CREATE Data Table and insert into Rmd}
\usage{
datatable_create_rmd(
  rmd_path,
  rmd_line,
  datatable_name = "samples",
  data_cols = "",
  IDs = "",
  default_data_vals = list(),
  dt_length = 100,
  expand = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file.}

\item{rmd_line}{line in Rmd file to insert table.}

\item{datatable_name}{String of data table name - default "samples".}

\item{data_cols}{Character vector of data table column titles.}

\item{IDs}{Character vector of sample IDs.}

\item{default_data_vals}{List of default data values to add to data cols &
IDs.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{expand}{Whether to expand any values in vectors set to \code{default_data_vals},
default is FALSE, meaning custom values can be set in \code{default_data_vals} for
each ID. If TRUE, teh vector is expanded by the length of IDs, meaning ALL
values in each vector are set to EACH ID, added as a multi-obs set of rows
for each ID.}
}
\description{
Creates a new Sample DataTable in specified Rmd file at specified line.
The Sample DataTable will contain an initial ID column containing the
IDs vector, and an optional set of extra data columns as specified in the
data_cols vector.  If the table exceeds dt_length characters (default 100),
then the table is split into multiple tables, with IDs as first col, and
subsequent data_cols given in subsequent tables.
}
