% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_create_template}
\alias{datatable_create_template}
\title{Create new Sample Data Table from Template}
\usage{
datatable_create_template(IDs, template_table, datatable_name, dt_length = 100)
}
\arguments{
\item{IDs}{Character vector of sample IDs}

\item{template_table}{A character vector of the template table, as presented
in the Rmd file.}

\item{datatable_name}{String of data table name.}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
DEPRECATED : Using datatable_create to handle both novel & templated
datatable create generation.
}
\details{
Creates a new Sample DataTable vector. The Sample DataTable will contain an
initial ID column containing the IDs vector, and an optional set of extra
data columns as specified in the data_cols vector.  If the table exceeds
`dt_length` characters (default 120), then the table is split into multiple
tables, with IDs as first col, and subsequent data_cols given in subsequent
tables.
}
