% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_datetimes_to_durations}
\alias{datatable_datetimes_to_durations}
\title{Convert datetime (suffix `_dt`) columns to DURATIONS}
\usage{
datatable_datetimes_to_durations(dt, include_now = FALSE, cumulative = FALSE)
}
\arguments{
\item{dt}{a datatable (tibble) that contains datetime columns (suffix `_dt`)}

\item{include_now}{Boolean to indicate whether the last datetime should have
its duration set up to the current time.  The last `_dt` column has its
duration calculated from the current time.  FALSE by default, meaning the
last datetime will not have a duration set.}

\item{cumulative}{Boolean to indicate whether the cumulative duration from the
first datetime should be calculated, or individual durations for each step
calculated.  FALSE by default, meaning the datetimes are returned as
independent durations}
}
\value{
A new tibble with all datetime columns replaced with duration
columns (suffix `_dur`), containing the duration data in`HH:MM` format.
}
\description{
This extracts all columns in dt that end in `_dt` (the conventional suffix
in projectmanagr datatable columns for a DATETIME), and computes the
DURATIONS between these columns.
}
\details{
The durations are returned in a new tibble, along with any non-datetime
columns, using the same naming convention as the original datatable, with
the `_dt` suffix replaced with `_dur`.

The duration data is in the format `HH:MM`, indicating the duration in
hours:minutes of the step.
}
