% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_dispose}
\alias{datatable_dispose}
\title{Dispose of a set of Samples}
\usage{
datatable_dispose(
  contents,
  datatable_name,
  dt_length = 100,
  summarise_reps = TRUE
)
}
\arguments{
\item{contents}{character vector containing current document with
existing data tables.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST IN \code{contents}.  The resampling will be default include all sample IDs
from this datatable.  The user may then delete sample IDs as appropriate.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}
}
\description{
This generates a new datatable that disposes of samples - indicates they no
longer exist.
}
\details{
Datatable comprises the dataframe NAME, sample IDs, reps (if applicable), and
the special column titled \code{dispose}.  In this column the datetime of disposal
(i.e. time of creation of dispose datatable) is written.
}
