% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-rmd-functions.R
\name{datatable_dispose_template_rmd}
\alias{datatable_dispose_template_rmd}
\title{Generate DISPOSE Data Table and insert into Rmd}
\usage{
datatable_dispose_template_rmd(
  rmd_path,
  rmd_line,
  datatable_name = "samples",
  dt_length = 100,
  summarise_reps = FALSE,
  all_reps = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file.}

\item{rmd_line}{line in Rmd file to insert table.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST IN \code{contents}.  The disposing will by default include all sample IDs
from this datatable.  The user may then delete sample IDs as appropriate.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to TRUE.}

\item{all_reps}{Boolean to indicate whether all reps should be summarised in
the datatable.  If TRUE then ALL reps for each ID are summarised with keyword
ALL - which will apply data cols to all the reps for each sample, without the
potentially complex summary strings representing a disparate collection of reps.
This comes at the expense of not being clear from reading this datatable what
reps actually exist.

param \code{cdt} is not passed - as it should be in the TEMPLATE datatable under the
\code{dispose} data column. s Datetime to insert into dispose column. Defaults to the
current datetime string retrieved from \code{get_datetime()}}
}
\description{
This generates a new datatable that disposes samples - indicates they no
longer exist.
}
\details{
Datatable comprises the dataframe NAME, sample IDs, reps (if applicable), and
the special column titled \code{dispose}.  In this column the datetime of disposal
(i.e. time of creation of dispose datatable) is written.
}
