% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-functions.R
\name{datatable_export_rmd}
\alias{datatable_export_rmd}
\title{Generate EXPORT Data Table and insert into Rmd}
\usage{
datatable_export_rmd(
  source_rmd_path,
  source_rmd_line,
  destination_rmd_path,
  settings,
  datatable_name,
  ids_vector = "",
  reps_vector = "",
  dt_length = 100,
  summarise_reps = FALSE,
  exportTemplate = "Datatables-Export-Template.Rmd"
)
}
\arguments{
\item{source_rmd_path}{path to Source Rmd file: the file that contains the
datatables with \code{datatable_name} that is read to extract the IDs (& REPs),
and where the EXPORT datatable is written to.}

\item{source_rmd_line}{line in Rmd file to read datatables up to, and where
the EXPORT datatable will be inserted.}

\item{destination_rmd_path}{path to Destination Rmd file: the file where the
sample IDs will be IMPORTED to.}

\item{settings}{projectmanagr settings yml data - to define prefix in links
and the export header in template.}

\item{datatable_name}{String of data table name - MUST exist in rmd_path!}

\item{ids_vector}{A vector containing the IDs to add: Must be a set of EXISTING
IDs from the selected \code{datatable_name} - cannot be ALL or a set of Group IDs.
Default blank string "" indicates to add data to ALL EXISTING IDs.  Existing
IDs are all IDs that have not been resampled, exported, or disposed.}

\item{reps_vector}{A vector containing the REPs to add: This is combined with
the \code{ids_vector} to specify ID-REP pairs.  Each ID-REP pair must exist (not
be resampled, exported, or disposed) from the selected \code{datatable_name}.
Default blank string "" indicates to add data to ALL EXISTING ID-REPs.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}

\item{exportTemplate}{The Rmd file to use as a template to generate EXPORT
Content in the \code{source_rmd_path}.  Contains parameters for placement of the
Export Header \code{{{DT_EXPORT_HEADER_TEMPLATE}}} and export datatable(s)
\code{{{DT_EXPORT_DATATABLE}}} in the content.  The Export Header is filled with
the \code{DatatableExportHeader} parameter from \code{settings} in Organisation, concatenated
to the \code{datatable_name}.  Any matches to this header that exist in \code{source_rmd_path}
file, causes the header name to be iterated with a number suffix.  This ensures
the EXPORT of samples in a project note are easy to identify in RStudio Rmd outlines
*as \code{DatatableExportHeader} by default is a markdown section header!)
Default is "Datatables-Export--Template.Rmd" from projectmanagr templates, and the
content of this file can be modified to suit the users needs.}

\item{destination_rel_path}{The relative path to the Destination Note where
these samples are EXPORT/IMPORTED to.}
}
\description{
This function generates an EXPORT datatable based on contents of \code{rmd_path}
up to \code{rmd_line}, & datatable contents of \code{datatable_name}. The EXPORT
datatable is inserted into \code{rmd_path} at \code{rmd_line}.
}
\details{
\code{rmd_path} is confirmed to be a PRoject Note or Doc with the \code{settings} yml
list.
}
