% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{datatable_get_delimiter}
\alias{datatable_get_delimiter}
\title{Return the datatable delimiter}
\usage{
datatable_get_delimiter()
}
\description{
String that indicates teh start and end of a datatable.  Using centralised
function to define this structure - for generation & for lookup.
}
