% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_get_sample_data_history}
\alias{datatable_get_sample_data_history}
\title{Get History of samples}
\usage{
datatable_get_sample_data_history(df)
}
\description{
Retrieves all sample data, traversing all divisions (resample, export).
}
\details{
Passes back through all Rmd's through imports, and identifies the CREATE
datatable for each sample.  Then collects all data, moving through all
resampling and exports, and omitting any blank columns.

Returns a list of tibbles that contains all sample data, collating samples
where the data columns match.
}
