% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{datatable_import}
\alias{datatable_import}
\title{Generate IMPORT Datatable}
\usage{
datatable_import(
  source_contents,
  destination_contents,
  datatable_name,
  source_rel_path_link,
  ids_vector = "",
  reps_vector = "",
  dt_length = 100,
  summarise_reps = FALSE
)
}
\arguments{
\item{source_contents}{character vector containing source document with
existing data tables for samples to be imported from.}

\item{destination_contents}{character vector containing destination document,
which may contains existing CREATE or IMPORT datatables, that must be checked
for possible sample ID/REP clashes.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.}

\item{source_rel_path_link}{Markdown link containing relative path to the Source
Note where these samples are impored from.  This is used to fill the IMPORT
datatable, allowing the tracing of imported sample data. Link typically has
structure of \verb{[PREFIX](relative_path)} (name is prefix, relative path declared
in path sportion of the link).}

\item{ids_vector}{A vector containing the IDs to add: Must be a set of EXISTING
IDs from the selected \code{datatable_name} - cannot be ALL or a set of Group IDs.
Default blank string "" indicates to add data to ALL EXISTING IDs.  Existing
IDs are all IDs that have not been resampled, exported, or disposed.}

\item{reps_vector}{A vector containing the REPs to add: This is combined with
the \code{ids_vector} to specify ID-REP pairs.  Each ID-REP pair must exist (not
be resampled, exported, or disposed) from the selected \code{datatable_name}.
Default blank string "" indicates to add data to ALL EXISTING ID-REPs.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}
}
\value{
List of datatables generated.
}
\description{
Imports all samples IDs from \code{datatable_name} in \code{source_contents} from
\code{source_rel_path}.
}
\details{
The generated IMPORT DataTable will contain:
\itemize{
\item an initial ID column containing all EXISTING IDs from the \code{source_contents}
datatable \code{datatable_name}  (IDs that have been resampled or exported will
automatically be EXCLUDED)
\item and an \code{import} column that contains the \code{source_rel_path} combined with
the ID & REP values in the source contents/datatable, for each ID/REP.  These
are declared using special separate syntax defined in \code{get_import_sep()} - by
default this is \code{::}.  This is necessary in case there are clashes in sample
IDs/REPs when IMPORTING data from different notes.
}

A subset of existing IDs or IDs/REPs (if reps exist in selected datatable)
can be set via the \code{ids_vector} and \code{reps_vector} arguments.  If blank, ALL
existing IDs (& REPs) will be imported.  This table will add the rep column
in the format indicated by \code{summarise_reps}.

If the table exceeds \code{dt_length} characters (default 100), then the table is
split into multiple tables, with IDs as first col, and subsequent data_cols
given in subsequent tables.
}
