% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-functions.R
\name{datatable_import_rmd}
\alias{datatable_import_rmd}
\title{Generate IMPORT Data Table and insert into Rmd}
\usage{
datatable_import_rmd(
  source_rmd_path,
  source_rmd_line,
  destination_rmd_path,
  destination_rmd_line,
  settings,
  datatable_name,
  ids_vector = "",
  reps_vector = "",
  dt_length = 100,
  summarise_reps = FALSE,
  exportHeader
)
}
\arguments{
\item{source_rmd_path}{path to Source Rmd file: the file that contains the
datatables with \code{datatable_name} that is read to extract the IDs (& REPs).}

\item{source_rmd_line}{line in Rmd file to read datatables up to.}

\item{destination_rmd_path}{path to Destination Rmd file: the file where the
new IMPORT datatable is written to.}

\item{destination_rmd_line}{line in Rmd file to insert the new IMPORT table.}

\item{settings}{projectmanagr settings yml data - to define prefix in links.}

\item{datatable_name}{String of data table name - MUST exist in rmd_path!}

\item{ids_vector}{A vector containing the IDs to add: Must be a set of EXISTING
IDs from the selected \code{datatable_name} - cannot be ALL or a set of Group IDs.
Default blank string "" indicates to add data to ALL EXISTING IDs.  Existing
IDs are all IDs that have not been resampled, exported, or disposed.}

\item{reps_vector}{A vector containing the REPs to add: This is combined with
the \code{ids_vector} to specify ID-REP pairs.  Each ID-REP pair must exist (not
be resampled, exported, or disposed) from the selected \code{datatable_name}.
Default blank string "" indicates to add data to ALL EXISTING ID-REPs.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}

\item{exportHeader}{String of markdown header written to source note where the export
datatable was written. This is used to link directly to this section from the import
datatable IMPORT col values - the source_re_path_link - using html link syntax.
Can navigate this link in R markdown using the \code{navigate_markdown_link()} function
in \code{projectmanagr}}

\item{destination_rel_path}{The relative path to the Destination Note where
these samples are EXPORT/IMPORTED to.}
}
\description{
This function generates an IMPORT datatable based on contents of \code{source_rmd_path}
up to \code{source_rmd_line}, & datatable contents with \code{datatable_name}. The IMPORT
datatable is inserted into \code{destination_rmd_path} at \code{destination_rmd_line}.
}
\details{
Both \code{source_rmd_path} and \code{destination_rmd_path} are confirmed to be
Project Notes or Docs with the \code{settings} yml list.
}
