% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_insert_from_dataframe}
\alias{datatable_insert_from_dataframe}
\title{Insert a datatable from a dataframe}
\usage{
datatable_insert_from_dataframe(df, dt_function = "CREATE", dt_length = 100)
}
\arguments{
\item{df}{a dataframe/tibble with data to insert into Rmd}

\item{dt_function}{String to declare type of table: MUST be CREATE ADD_DATA
RESAMPLE GROUP EXPORT or IMPORT. Default is CREATE.}

\item{dt_length}{Int that dictates the maximum length of any one inserted datatable.}
}
\description{
Will insert the passed dataframe, using the variable's NAME as the datatable
NAME.  Will CREATE the datatable first, then subsequent cols that dont fit in
the first datatable will be in subsequent ADD_DATA datatables.
}
\details{
Inserts the datatable where the CURSOR IS in the Active Rmd.
}
