% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-rmd-functions.R
\name{datatable_load_data}
\alias{datatable_load_data}
\title{Load all sample data from summary datatable}
\usage{
datatable_load_data(summary_dt)
}
\description{
Retrieves all sample data, traversing all divisions (resample, export) of
all samples in a summary datatable, generated by datatable_find()
}
\details{
Will return all columns that are COMMON amongst all samples in datatable.

Passes back through all Rmd's through imports, and identifies the CREATE
datatable for each sample.  Then collects all data, moving through all
resampling and exports, and omitting any blank columns.

Returns a list of tibbles that contains all sample data, collating samples
where the data columns match.
}
