% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{datatable_read_active_vector}
\alias{datatable_read_active_vector}
\title{Read and return all ACTIVE datatables}
\usage{
datatable_read_active_vector(rmd_contents)
}
\description{
Return all datatables which contain sample IDs that are ACTIVE : they have
not been resampled or exported.
}
\details{
Will exclude any tables that contain the \code{resample} or \code{export} columns only.
It is true they may not be filled for all sample IDs, but in these cases any
further resampling or exporting should be performed in the original table
in order to keep all the data log for resampling or exporting of sample IDs
from one datatable in one place.
}
