% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-rmd-functions.R
\name{datatable_read_rmd}
\alias{datatable_read_rmd}
\title{Read datatables in Rmd}
\usage{
datatable_read_rmd(rmd_path)
}
\arguments{
\item{rmd_path}{Path to RMarkdown Document.}
}
\value{
a NAMED LIST of tibbles containing all data.  Names correspond to
the names of the datatables in the file.  Pivot tables from IMPORT IDs are
generated for each source file - using the file name (or minimal unique
path) as the NAME for this table (which will be present in the \code{import}
column alongsidethe ID in its original table from the first file).
}
\description{
Function to read all projectmanagr datatables declared across one or more
plain text documents (typically an R Markdown doc), and return them as
tibbles.
}
\details{
\enumerate{
\item Extracts all datatables between \code{datatable_get_delimiter()} and checks the
validity of each datatable declared:
\item Datatables have NAMES, and contains samples which must all possess a
unique ID.
\item Datatables have FUNCTIONS - sample IDs are created in CREATE, and have
data added to them in ADD_DATA tables.  IDs from a given datatable can be put
into new groups using a GROUP table, resampled into subsamples in RESAMPLE
tables, disposed in DISPOSE tables, and exported to a new file with EXPORT
tables.
\item Samples may also be generated via IMPORT tables, which links with
corresponding EXPORT tables in linked notes.  This function reads the IMPORT
tables, and traverses back to the source files to extract all existing meta
data for these sample IDs up to the point of EXPORT from the source note.
}
}
