% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-rmd-functions.R
\name{datatable_read_rmd_active_ids}
\alias{datatable_read_rmd_active_ids}
\title{Read and return all ACTIVE IDs in datatables}
\usage{
datatable_read_rmd_active_ids(rmd_contents, rmd_line)
}
\description{
Excludes all IDs that have been RESAMPLED DISPOSED or EXPORTED.
}
