% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{datatable_read_vector}
\alias{datatable_read_vector}
\title{Read all complete Datatables - between to delimiters}
\usage{
datatable_read_vector(rmd_contents, notify = TRUE)
}
\arguments{
\item{rmd_contents}{String vector containing datatables between delimiters}
}
\value{
A named list of tibbles.
}
\description{
Delimiters defined in \code{datatable_get_delimiter()}
}
