% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_split}
\alias{datatable_split}
\title{Split Samples into separate datatables}
\usage{
datatable_split(contents, datatable_name, split_vector, dt_length = 120)
}
\description{
This is NOW DEPRECATED - not using SPLIT as this interferes with the FIND function:
find will summarise sample IDs and REPs from each Rmd into each sub-sample, which
is dictated by the resampling vector chain - which indicates WHAT TISSUES ARE
AVAILABLE!  By corrupting this with SPLIT codes, this becomes harder to use?
}
\details{
Further, the get_history function will ELIMINATE any blank columns when reading
back to collect data on samples/reps so this is not really a problem?

JUST USE GROUP - no need to split!
}
