% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{default_events_list_fn}
\alias{default_events_list_fn}
\title{Retrieve Events from a Google Calendar}
\usage{
default_events_list_fn(
  cal_id,
  tmin,
  tmax,
  gar_api_generator = googleAuthR::gar_api_generator
)
}
\arguments{
\item{cal_id}{A character string specifying the calendar ID.}

\item{tmin}{A POSIXct object representing the start of the time window (UTC).}

\item{tmax}{A POSIXct object representing the end of the time window (UTC).}
}
\value{
A parsed list object (from JSON) representing events returned by the
Google Calendar API. If no events are found, \verb{$items} will be an empty list.
}
\description{
This function defines and executes a default API request to the Google Calendar API
to fetch all events for a specified calendar within a given time window.
}
\details{
It uses \code{googleAuthR::gar_api_generator()} to construct a call to the
\code{events.list} endpoint for a specific calendar.

This function is intended for injection into higher-level functions (e.g.,
\code{get_google_calendar_events()}) and can be mocked for testing.
}
