% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_data_cols}
\alias{define_data_cols}
\title{Define IDs and REPs - trimmed based on current datatable data_cols contents}
\usage{
define_data_cols(
  datatable,
  data_cols,
  IDs,
  REPs,
  rep_exists,
  summarise_reps,
  all_reps
)
}
\value{
named list containing MODIFIEDs IDs, REPs, summarise_reps, rep_exists
}
\description{
Define IDs and REPs - trimmed based on current datatable data_cols contents
}
