% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_data_vals}
\alias{define_data_vals}
\title{Define data_cols + wds + vals - based on data_cols, IDs, & REPs}
\usage{
define_data_vals(
  default_data_vals,
  data_cols,
  IDs,
  REPs,
  rep_exists,
  groups = FALSE,
  expand = TRUE
)
}
\arguments{
\item{groups}{Boolean to indicate id IDs are group IDs.  Stops rep col being added
if IDs are group IDs.}

\item{expand}{Whether to expand default_data_vals by length of IDs.  If default_data_vals
have been set, they are expanded along the length of IDs by default (expand is TRUE). If
expand is set to FALSE, the default_data_vals are not expanded - this allows a list
of vectors that match length of IDs vector to be added, allowing custom values for
each ID.}
}
\value{
named list containing MODIFIED data_cols, data_col_wds, default_data_vals
}
\description{
Define data_cols + wds + vals - based on data_cols, IDs, & REPs
}
