% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_data_vals_groups}
\alias{define_data_vals_groups}
\title{Define data_cols + wds + vals - based on data_cols, IDs, & REPs}
\usage{
define_data_vals_groups(datatable, groups, data_cols, IDs, REPs, rep_exists)
}
\value{
named list containing MODIFIED data_cols, data_col_wds, default_data_vals
}
\description{
Staggers the values in groups along the length of IDs & REPs.
}
