% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_groups}
\alias{define_groups}
\title{Validate & define group names - from datatable, var_names & group_names.}
\usage{
define_groups(dt, var_names, group_names)
}
\value{
group_names vector
}
\description{
Validate & define group names - from datatable, var_names & group_names.
}
