% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_ids}
\alias{define_ids}
\title{Define IDs and REPs - based on current datatable contents}
\usage{
define_ids(ids_vector, datatable, summarise_reps, all_reps)
}
\value{
named list containing IDs, REPs, summarise_reps, rep_exists
}
\description{
Define IDs and REPs - based on current datatable contents
}
