% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_ids_reps}
\alias{define_ids_reps}
\title{Define IDs and REPs - based on current datatable contents}
\usage{
define_ids_reps(ids_vector, reps_vector, datatable, summarise_reps)
}
\value{
named list containing IDs, REPs, summarise_reps, rep_exists
}
\description{
This function fails to parse ALL or <GROUP_NAMES>, and only allows explicit
subsets of IDs & REPs to be defined in the \code{ids_vector} & \code{reps_vector}.
}
