% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{define_import_vals}
\alias{define_import_vals}
\title{Define Import Valus}
\usage{
define_import_vals(dc_list, source_rel_path_link)
}
\description{
Returns list of \code{default_data_vals} for \code{import} col in IMPORT datatable,
consisting of source_rel_path_link separated from IDs with get_import_sep(), and
possibly REPs if they exist.
}
