% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{define_org_path}
\alias{define_org_path}
\title{Define Organisation Path}
\usage{
define_org_path(orgParentPath, orgName)
}
\arguments{
\item{orgParentPath}{Character. The parent directory for the organisation.}

\item{orgName}{Character. The name of the organisation.}
}
\value{
Character. The full path for the organisation.
}
\description{
Constructs the full path for a new organisation and ensures it does not already exist.
}
