% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{define_prog_title}
\alias{define_prog_title}
\title{Check programmeTitle, and if blank, fill with programmeName
replacing all "_" and "-" with spaces}
\usage{
define_prog_title(programmeTitle, programmeName)
}
\description{
Check programmeTitle, and if blank, fill with programmeName
replacing all "_" and "-" with spaces
}
