% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{define_section_title}
\alias{define_section_title}
\title{Check sectionTitle, and if blank, fill with sectionName
replacing all "_" and "-" with spaces}
\usage{
define_section_title(sectionTitle, sectionName)
}
\description{
Check sectionTitle, and if blank, fill with sectionName
replacing all "_" and "-" with spaces
}
