% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-functions.R
\name{define_template_data}
\alias{define_template_data}
\title{define template data}
\usage{
define_template_data(
  rmd_contents,
  rmd_line,
  template_indices,
  datatable_name,
  all_ids,
  expand = TRUE
)
}
\arguments{
\item{all_ids}{Boolean to indicate whether all IDs from \code{source_daatable_name}
should be returned: ignoring any that are DISPOSED EXPORTED or RESAMPLED. When
creating a new CREATE datatable from existing set of IDs, may want to include
all IDs independent of whether DISPOSED/EXPORTED/RESAMPLED.  Setting \code{all_ids}
to TRUE ensures all ID Strings are extracted to be written to the new datatable.}

\item{expand}{Whether to expand any values in vectors set to \code{default_data_vals},
default is FALSE, meaning custom values can be set in \code{default_data_vals} for
each ID. If TRUE, teh vector is expanded by the length of IDs, meaning ALL
values in each vector are set to EACH ID, added as a multi-obs set of rows
for each ID.}
}
\value{
named list : data_cols, IDs, default_data_vals
}
\description{
define template data
}
