% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-functions.R
\name{define_template_data_IDs}
\alias{define_template_data_IDs}
\title{define template data with manual IDs}
\usage{
define_template_data_IDs(
  rmd_contents,
  rmd_line,
  template_indices,
  IDs,
  all_ids = FALSE,
  expand = TRUE
)
}
\arguments{
\item{expand}{Whether to expand any values in vectors set to \code{default_data_vals},
default is FALSE, meaning custom values can be set in \code{default_data_vals} for
each ID. If TRUE, teh vector is expanded by the length of IDs, meaning ALL
values in each vector are set to EACH ID, added as a multi-obs set of rows
for each ID.}
}
\value{
named list : data_cols, IDs, default_data_vals
}
\description{
define template data with manual IDs
}
