% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-functions.R
\name{define_template_ids_cols}
\alias{define_template_ids_cols}
\title{Exract IDS and Cols}
\usage{
define_template_ids_cols(
  template_data,
  IDs,
  rmd_line,
  all_ids,
  expand,
  datatable_name
)
}
\arguments{
\item{all_ids}{Boolean to indicate whether all IDs from \code{source_daatable_name}
should be returned: ignoring any that are DISPOSED EXPORTED or RESAMPLED.}

\item{expand}{Whether to expand any values in vectors set to \code{default_data_vals},
default is FALSE, meaning custom values can be set in \code{default_data_vals} for
each ID. If TRUE, teh vector is expanded by the length of IDs, meaning ALL
values in each vector are set to EACH ID, added as a multi-obs set of rows
for each ID.}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
Exract IDS and Cols
}
