% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{derive_scope_string}
\alias{derive_scope_string}
\title{Derive a short scope label from a file/directory location}
\usage{
derive_scope_string(location)
}
\description{
Derive a short scope label from a file/directory location
}
\keyword{internal}
