% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlink-functions.R
\name{doc_head_link}
\alias{doc_head_link}
\title{create link from Project Doc to Project Sub Note}
\usage{
doc_head_link(headerNoteRmdPath, projectDocPath, settings)
}
\description{
For inserting into a Project Doc under a GDT to form a link between the Note
and Doc GDT.
}
