% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{doc_note_link}
\alias{doc_note_link}
\title{create link from Project Doc to Project Note}
\usage{
doc_note_link(projNoteRmdPath, projectDocPath, settings)
}
\description{
For inserting into a Project Doc under a GDT to form a link between the Note
and Doc GDT.
}
