% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.cleanup_stale_sessions}
\alias{.cleanup_stale_sessions}
\title{Clean up stale RStudio sessions}
\usage{
.cleanup_stale_sessions(session_dirs, matching_session, verbose = TRUE)
}
\arguments{
\item{session_dirs}{A character vector of all session directories.}

\item{matching_session}{A character string path to the active session directory.}

\item{verbose}{Logical. If \code{TRUE}, print message when deleting stale sessions.}
}
\description{
Deletes session directories that were not identified as the active session.
}
\keyword{internal}
