% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.get_active_project_path}
\alias{.get_active_project_path}
\title{Get Active Project Path (wrapper)}
\usage{
.get_active_project_path()
}
\description{
Thin wrapper around rstudioapi::getActiveProject().
Returns the project path or NULL if not in a project.
}
\keyword{internal}
