% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.get_candidate_sources_dirs}
\alias{.get_candidate_sources_dirs}
\title{Find candidate RStudio sources directories}
\usage{
.get_candidate_sources_dirs()
}
\value{
A character vector of "sources" directory paths.
}
\description{
Determines whether the current session is project-based or global,
and returns the appropriate "sources" directories to inspect.
}
\keyword{internal}
