% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.get_open_doc_paths}
\alias{.get_open_doc_paths}
\title{Get open document paths from RStudio source editor}
\usage{
.get_open_doc_paths()
}
\value{
A character vector of absolute file paths.
}
\description{
Attempts to safely retrieve the full paths of all open documents in the RStudio source editor.
}
\keyword{internal}
