% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.is_rstudio_available}
\alias{.is_rstudio_available}
\title{Check if RStudio API is available}
\usage{
.is_rstudio_available()
}
\description{
Check if RStudio API is available
}
\keyword{internal}
