% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.navigate_pane_to_file}
\alias{.navigate_pane_to_file}
\title{Navigate Pane to File (wrapper)}
\usage{
.navigate_pane_to_file(path)
}
\description{
Thin wrapper around rstudioapi::insertText.
}
\keyword{internal}
