% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.select_most_recent_session}
\alias{.select_most_recent_session}
\title{Select the most recent matching session}
\usage{
.select_most_recent_session(matches)
}
\arguments{
\item{matches}{A list of matched sessions.}
}
\value{
A string path to the selected session directory.
}
\description{
If multiple sessions match, selects the session with the most recently modified metadata file.
}
\keyword{internal}
