% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{.set_document_contents}
\alias{.set_document_contents}
\title{Set document contents in RStudio}
\usage{
.set_document_contents(text, id)
}
\description{
Set document contents in RStudio
}
\keyword{internal}
