% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{expand_id_rep}
\alias{expand_id_rep}
\title{Expand IDs and REPs}
\usage{
expand_id_rep(IDs, REPs)
}
\description{
IDs contains each ID ONCE, and in REPs this is accompanied with a summarised
reps string.  This method expands both, and returns them in a NAMED LIST.
}
