% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{expand_id_rep_data}
\alias{expand_id_rep_data}
\title{Expand IDs & REPs Strings}
\usage{
expand_id_rep_data(IDs, REPs, DATA, dt)
}
\description{
Expanding REPs in summarised form to a sequence of integers:
}
\details{
Example: "1:3,5,7:9" summary string of REPS should yield the following
integer sequence:
\verb{1 2 3 5 7 8 9}

Expanding \code{IDs} and \code{DATA} alongside these, and returning as a named list.
}
