% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{expand_rep}
\alias{expand_rep}
\title{Expand Rep Strings}
\usage{
expand_rep(rep_str)
}
\description{
Expanding REPs in summarised form to a sequecne of integers.
}
\details{
Example: "1:3,5,7:9" summary string of REPS should yield the following
integer sequence:
\verb{1 2 3 5 7 8 9}
}
