% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{extract_data_tables}
\alias{extract_data_tables}
\title{Extract Data Tables}
\usage{
extract_data_tables(rmd_contents, indices)
}
\arguments{
\item{rmd_contents}{Character Vector containing Data Tables to read.}

\item{indices}{Line indices of table delineators : \verb{+====}.}
}
\value{
List of character vectors, each containing a data table.
}
\description{
Separate each data table character vector into a new list.
}
