% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{extract_google_calendar_events_markdown}
\alias{extract_google_calendar_events_markdown}
\title{Extract events from Google Calendar for one day}
\usage{
extract_google_calendar_events_markdown(
  day = Sys.Date(),
  location = getwd(),
  settings = list(),
  calEventsTemplate = "Calendar-Events-Template.Rmd",
  extract_google_calendar_events_fun = extract_google_calendar_events
)
}
\arguments{
\item{day}{\code{\link[base:Dates]{Date}} Date to query (default: \code{Sys.Date()}).}

\item{settings}{\code{list} Named options; currently only
\code{googleCalendars = c("Calendar A", "Calendar B")}
to filter the calendar list.}
}
\value{
A \link[tibble:tibble]{tibble} with the columns\
\code{calendar}, \code{summary}, \code{start}, \code{end}, \code{description}, \code{htmlLink},
\code{location}.
}
\description{
This convenience wrapper\
• Returns extracted google calendar events from \code{extract_google_calendar_events()}
as a block of markdown text, for direct insertion into a Rmd file
}
