% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{extract_ids}
\alias{extract_ids}
\title{Extract IDs}
\usage{
extract_ids(data, table_name, table_start_line)
}
\arguments{
\item{data}{A list containing vectors of data, with ID in the first vector.}

\item{table_name}{name of data table.}

\item{table_start_line}{Line data table starts on in character vector.}
}
\value{
the plaintext datatable function from \code{table_vector}
}
\description{
Extract IDs vector from the data list - IDs should be in FIRST VECTOR,
and ID values should be in 2:length of that vector.
}
\details{
Fails gracefully if any IDs are repeated.
}
