% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{extract_objectives_note_GDT}
\alias{extract_objectives_note_GDT}
\title{Extract GDT & summary & TODO from Project Note Objectives vector}
\usage{
extract_objectives_note_GDT(
  linkObjectives,
  linkRmdPath,
  subNoteRmdPath,
  settings,
  orgPath
)
}
\description{
Extract GDT & summary & TODO from Project Note Objectives vector
}
