% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{extract_prog_summ}
\alias{extract_prog_summ}
\title{extract string vector that contains the programme summary - between indices
of settings: \code{ProgSummaryHeader} & \code{ProgSummaryFooter} & after \code{orgLink}}
\usage{
extract_prog_summ(progContents, orgIndex, progIndex, orgPath, settings)
}
\description{
extract string vector that contains the programme summary - between indices
of settings: \code{ProgSummaryHeader} & \code{ProgSummaryFooter} & after \code{orgLink}
}
