% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{extract_sect_summ}
\alias{extract_sect_summ}
\title{extract string vector that contains the programme section summary - between
indices of settings: \code{SectionSummaryHeader} & \code{SectionSummaryFooter} & after
\code{parLink}}
\usage{
extract_sect_summ(sectContents, parentIndex, sectIndex, orgPath, settings)
}
\description{
extract string vector that contains the programme section summary - between
indices of settings: \code{SectionSummaryHeader} & \code{SectionSummaryFooter} & after
\code{parLink}
}
