% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{extract_table_function}
\alias{extract_table_function}
\title{Extract Table Function}
\usage{
extract_table_function(table_vector)
}
\arguments{
\item{table_vector}{A character vector containing a data table.}
}
\value{
the plaintext datatable function from \code{table_vector}
}
\description{
Extract Table Function
}
