% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{extract_table_name}
\alias{extract_table_name}
\title{Extract Table Name}
\usage{
extract_table_name(table_vector)
}
\arguments{
\item{table_vector}{A character vector containing a data table.}
}
\value{
the plaintext datatable name from \code{table_vector}
}
\description{
Extract Table Name
}
