% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{extract_template_named_table}
\alias{extract_template_named_table}
\title{Extract Template Named Table}
\usage{
extract_template_named_table(table_vector)
}
\arguments{
\item{table_vector}{A character vector containing a data table.}
}
\value{
the plaintext datatable function from \code{table_vector}
}
\description{
Returns the name. eg from NAMED TEMPLATE TABLE below:
}
\details{
\code{    TEMPLATE  :  CREATE  :  fix-solution-wts}

Will return "fix-solution-wts" - returns NA if no name exists.
}
