% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{extract_todos}
\alias{extract_todos}
\title{Extract All TODO lists from a Collection of Project Notes}
\usage{
extract_todos(
  location,
  fileNamingConvention = "YYYYMMDD_SCOPE",
  todoGDTExtractionTemplate = "Todo-GDT-Extraction-Template.Rmd",
  todoExtractionTemplate = "Todo-Extraction-Template.Rmd",
  todoProgCollectionTemplate = "Todo-Programme-Collection-Template.Rmd",
  todoCollectionTemplate = "Todo-Collection-Template.Rmd"
)
}
\arguments{
\item{location}{A Project Note file or Directory inside an Organisation
where TODO lists should be extracted from each GDT.  Can be Organisation
root, a programme, any sub-directory that contains project notes, or a
single project note.}
}
\description{
Extracted to Todo-Extraction-Template.Rmd that provides a template to
extract an absolute filesystem link to the project note, summarise each
Doc GDT with link, and copy all incompleted TODOs under this Doc GDT.
}
\details{
Todo-Collection-Template.Rmd that provides a template for organising multiple
Project Notes, listing notes by PROGRAMME in organisation directory tree.

incompleted TODOs can then be reviewed, and prioritised by re-ordering &
deleting the project note / doc GDT / TODOs.

The filled Todo-Extraction-Template.Rmd is saved to the todo directory,
defined in settings.  The TODO extraction is saved using a naming convention
of the date & scope (location directory name or project-note title) :
`YYYMMDD_SCOPE.Rmd`
}
