% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_header_Rmd_path}
\alias{find_header_Rmd_path}
\title{Find Group Note Header Rmd Path}
\usage{
find_header_Rmd_path(subNoteRmdPath, settings)
}
\arguments{
\item{subNoteRmdPath}{Path to sub note Rmd file}

\item{settings}{List from \code{.config/settings.yml} file.}
}
\description{
From the \code{subNoteRmdPath} - an absolute path.
}
