% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_org_path}
\alias{find_org_path}
\title{find organisation root directory from passed path}
\usage{
find_org_path(path)
}
\description{
Used by \code{find_org_directory()} as first pass attempt to identify an
organisation root directory.
}
