% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{find_org_rstudio_editor}
\alias{find_org_rstudio_editor}
\title{find organisation root directory from RStudio Editor active doc path}
\usage{
find_org_rstudio_editor()
}
\description{
Used by \code{find_org_directory()} as second pass attempt to identify an
organisation root directory.
}
