% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_prog_dir}
\alias{find_prog_dir}
\title{Find Programme Dir}
\usage{
find_prog_dir(fileSystemPaths)
}
\arguments{
\item{fileSystemPaths}{a vector of paths to identify programme paths from.}
}
\description{
Recursively searches fileSystemPaths to identify a
Programme directory.  This is identified by identifying the organisation root
directory from each parent direcotry searched.
}
\details{
If Programme paths are identified, they are  returned, otherwise
the function returns BLANK strings "".
}
