% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{find_protocols_in_dir_tree}
\alias{find_protocols_in_dir_tree}
\title{Find Protocols in Project Note}
\usage{
find_protocols_in_dir_tree(dirTree, orgPath, settings)
}
\description{
Searches through dirTree (non-recursively) for Project Notes, and then looks
for Protocols in these based on the protocol sep delimiter (identified in
`config/templates/PROTOCOL_SEP.txt`).
}
\details{
Returns a structured lists that describe each protocol in the project notes:

Each Structured List is comprised
[[3]] : Name of Protocol in the Project Note
[[1]] : ProjectNotePath
[[2]] : Project Note Line Number where Protocol BEGINS (start of first delimiter)
}
