% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{findProjectDocGoalDelTaskLineIndex}
\alias{findProjectDocGoalDelTaskLineIndex}
\title{Find Goal/Del/Task Line Index in Project Doc}
\usage{
findProjectDocGoalDelTaskLineIndex(
  projectDocContents,
  goalNum,
  delNum,
  taskNum
)
}
\description{
Identify the line index of a goal/del/task, given the goalNum, delNum, taskNum.
}
