% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-rmd-addins.R
\name{format_group_declaration_bullets}
\alias{format_group_declaration_bullets}
\title{Format group declaration}
\usage{
format_group_declaration_bullets(content_selection)
}
\arguments{
\item{content_selection}{Selection of GROUP TITLEs and GOUP IDs in bullets}
}
\description{
Convert the bulleted list of GROUP TITLEs and GROUP IDs to space-separated
string, for use in the addin_datatable_add_groups ADDIN.
}
