% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{generate_plaintext_calendar}
\alias{generate_plaintext_calendar}
\title{Generate Plaintext Calendar}
\usage{
generate_plaintext_calendar(
  year,
  month,
  day,
  calendar_header = "# DAILY LOG :"
)
}
\description{
Internal function to generate a plaintext calendar for insertion into
the weekly journal.
}
\details{
The calendar is laid out as a matrix with columns for each day of the week,
starting with SUNDAY, and dates filled into this matrix - from the first date
of the month to the end of the week (ie. the Saturday) BEYOND the last date
of the month.
}
