% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{generate_weekly_rmarkdown}
\alias{generate_weekly_rmarkdown}
\title{Generate Weekly RMarkdown Daily Log}
\usage{
generate_weekly_rmarkdown(year, month, day, separator_lines = "{{SEP02}}")
}
\description{
Creates RMarkdown content for 7 days, from the year-month-day date passed
to this function.
}
