% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_absolute_path}
\alias{get_absolute_path}
\title{Get Absolute Path from full + relative paths}
\usage{
get_absolute_path(fullPath, relativePath)
}
\arguments{
\item{fullPath}{initial full (absolute) path as source.}

\item{relativePath}{A relative path from the fullPath, to collapse.}
}
\description{
Collapses paths to one absolute path
}
