% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_config_dir}
\alias{get_config_dir}
\title{Get config dir}
\usage{
get_config_dir(orgPath)
}
\description{
This is set as \code{.config} GLOBALLY - used to identify
a Project Organisation root directory.
}
