% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_declarations}
\alias{get_content_declarations}
\title{Extract metadata from each declaration of content in source note}
\usage{
get_content_declarations(sourceNoteRmdPath, settings, orgPath)
}
\arguments{
\item{sourceNoteRmdPath}{Path to source project note Rmd}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\value{
List of content declaration containing named parameters: "contentTitle",
"contentDescription", "contentSource", "contentStartLine", "contentEndLine",
"projectNotePath", "projectNoteContentHeader"
}
\description{
Declarations of content identified as existing between \code{settings[["ContentSep"]]}
delimiters.  This function gets all declared content from \code{sourceNoteRmdPath},
and returns a named list of all content identified.
}
