% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_description}
\alias{get_content_description}
\title{Get Insertable Content Description}
\usage{
get_content_description(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the insertable content with startSep at startSepLineIndex.
}
