% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_header}
\alias{get_content_header}
\title{Get Insertable Content Header}
\usage{
get_content_header(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the insertable content with startSep at startSepLineIndex.
}
