% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_source}
\alias{get_content_source}
\title{Get Content Source}
\usage{
get_content_source(contentDeclContents, settings, orgPath)
}
\arguments{
\item{contentDeclContents}{The Content Declaration Contents - character vector that
includes all lines between content separators.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\value{
the Content Source value as string
}
\description{
Extract content Source from the content Contents, using the ContentSourceField
parameter declared in settings.
}
