% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-functions.R
\name{get_css_theme}
\alias{get_css_theme}
\title{Return CSS theme that matches RStudio theme: Light or Dark Mode.}
\usage{
get_css_theme()
}
\description{
Return CSS theme that matches RStudio theme: Light or Dark Mode.
}
