% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_date}
\alias{get_date}
\title{Get Date}
\usage{
get_date(timezone = get_locale(), split = "-")
}
\arguments{
\item{timezone}{Timezone code to use. Default "UTC".}

\item{split}{Chracter to split YYYY & MM & DD. Default "/".}
}
\description{
internal function to get date in YYYY MM DD format
}
