% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{get_delim_indices}
\alias{get_delim_indices}
\title{extract the START of each column delimiter index}
\usage{
get_delim_indices(delim)
}
\description{
String defines column widths: "    ======  =============  ==================  "
}
\details{
This function extracts the indices where "=" starts after spaces.
More generally, it extracts indices where any NON-SPACE char starts after spaces.

EG above returns c(5, 13, 28).

Used for determining which COL a string from a multi-observation belongs to
As in this case, not each col is filled with a value!

May also use for VALIDATING which cols data belongs to on a regular line...?
}
