% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_deliverable_number}
\alias{get_deliverable_number}
\title{get deliverable number}
\usage{
get_deliverable_number(deliverable, settings)
}
\description{
`deliverable` as collected from selection[["deliverable"]] - fitting the format specified
in settings[["ProjectDeliverableHeader"]].
}
