% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_deliverable_title}
\alias{get_deliverable_title}
\title{get deliverable title}
\usage{
get_deliverable_title(deliverable, settings)
}
\description{
\code{deliverable} as collected from selection[\link{"deliverable"}] - fitting the format specified
in settings[\link{"ProjectDeliverableHeader"}].
}
