% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_doc_GDT_link_note_paths}
\alias{get_doc_GDT_link_note_paths}
\title{Get Project Note Paths from specified Doc GDT link}
\usage{
get_doc_GDT_link_note_paths(
  projectDocPath,
  projectDocContents,
  taskLine,
  settings
)
}
\description{
Get project note paths from the doc GDT link at \code{taskLine}. Assumes
\code{projectDocContents} is a character vector from \code{projectDocPath}.
}
