% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_doc_gdt_project_note_paths}
\alias{get_doc_gdt_project_note_paths}
\title{Get Project Note Paths from Doc GDT links}
\usage{
get_doc_gdt_project_note_paths(projectDocPath, projDocContents, settings)
}
\description{
Get project note paths from doc GDT links  Assumes `projDocContents` is a
character vector from a project doc.
}
