% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{get_doc_task_title}
\alias{get_doc_task_title}
\title{get doc task title}
\usage{
get_doc_task_title(DocPrefix, task, settings)
}
\description{
\code{DocPrefix} is prefix to project document that \code{task} is from.
}
\details{
\code{task} as collected from selection[\link{"task"}] - fitting the format specified
in settings[\link{"ProjectTaskHeader"}].

Returns DocTaskTitle Header - DocName plus the TaskTitle, spearated with \code{:}.
}
