% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_doc_task_title_name}
\alias{get_doc_task_title_name}
\title{get doc task title}
\usage{
get_doc_task_title_name(DocPrefix, taskName, settings)
}
\description{
Used for forming the GDT link - see \code{copmute_doc_GDT_link()}
}
\details{
\code{DocPrefix} is prefix to project document that \code{task} is from.

\code{taskName} as collected from \code{get_task_title(selection[["task"]])}, is the
task name string of a task.

Returns DocTaskTitle Header - DocName plus the TaskTitle, spearated with \code{:}.
}
