% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_file_contents_headers}
\alias{get_file_contents_headers}
\title{Get File Contents Headers}
\usage{
get_file_contents_headers(contents)
}
\arguments{
\item{contents}{Character vector containing the contents of the file.}
}
\value{
Character vector of every header line.
}
\description{
Get every Markdown header line from text file - every line that begins with
\verb{#}.  Pre-filter contents to remove all code blocks, that may have lines
beginning with \verb{#} but are not markdown headers.
}
