% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_file_types}
\alias{get_file_types}
\title{Get File Types}
\usage{
get_file_types(filePaths, settings)
}
\arguments{
\item{settings}{List from \code{.config/settings.yml} file.}

\item{filePath}{the Absolute path to the file.}
}
\description{
Get types of a vector of file paths.
}
\details{
Determine if the current file is a PROJECT DOC (it is inside the PROJECTS/ Directory),
a HEADER GROUP NOTE (contains the string "-00~_"), or a SIMPLE or SUB Project Note (any
other file).

Returns the relevant String:  DOC, HEAD, SUB, NOTE, or UNKNOWN for unidentified file.
}
