% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_goal_number}
\alias{get_goal_number}
\title{get goal number}
\usage{
get_goal_number(goal, settings)
}
\description{
\code{goal} as collected from selection[\link{"goal"}] - fitting the format specified
in settings[\link{"ProjectGoalHeader"}].
}
