% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{get_goal_title}
\alias{get_goal_title}
\title{get goal title}
\usage{
get_goal_title(goal, settings)
}
\description{
\code{goal} as collected from selection[\link{"goal"}] - fitting the format specified
in settings[\link{"ProjectGoalHeader"}].
}
