% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_header_hash}
\alias{get_header_hash}
\title{Get Header Hash}
\usage{
get_header_hash(header)
}
\arguments{
\item{header}{The header string}
}
\description{
Returns leading #s in a markdown header.
}
\details{
eg. '### Header : Title # after hash ## after double-hash'

is returned as "###"
}
