% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_header_note_dir_path}
\alias{get_header_note_dir_path}
\title{Get Project Header Note directory path from header note path}
\usage{
get_header_note_dir_path(headerNoteFilePath, settings)
}
\description{
Compute a Project Header Note Directory Path from a fileSystem path.
}
\details{
The fileSystemPath must be a DIRECTORY and sub-dir in a PROGRAMME DIRECTORY
\itemize{
\item either a project Document DIR, or another Directory in a PROGRAMME for
storing Project Notes.
}

The NAME of the selected Directory is used to derive the PREFIX for files within this
Directory - it uses all LETTERS and NUMBERS UP TO the first "_" in the directory name.
If no Underscore exists, will use the whole Directory name as the Prefix for the projectNote.
}
