% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_header_title}
\alias{get_header_title}
\title{Get Header Title}
\usage{
get_header_title(header)
}
\arguments{
\item{header}{The header string}
}
\description{
Removes leading #s in a markdown header.  Ensures any subsequent #s in the title
are RETAINED.
}
\details{
eg. '### Header : Title # after hash ## after double-hash'

is returned as " Header : string # after hash ## after double-hash"
}
