% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{get_import_sep}
\alias{get_import_sep}
\title{Get Import Separator}
\usage{
get_import_sep()
}
\description{
return double colon string - used to separate source relative path, ID,
and REP(s) in an IMPORT datatable.
}
