% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{get_indices_in_col}
\alias{get_indices_in_col}
\title{Return all indices of all matches of all \code{group_names} in
\code{dt[[group_col_names]]} in the order they exist in \code{group_names}.}
\usage{
get_indices_in_col(dt, group_names, group_col_name)
}
\description{
Return all indices of all matches of all \code{group_names} in
\code{dt[[group_col_names]]} in the order they exist in \code{group_names}.
}
