% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_journal_dir}
\alias{get_journal_dir}
\title{Get journal dir}
\usage{
get_journal_dir(orgPath, settings)
}
\description{
Stores journal Rmd files.  This is set as \verb{journal/} by default,
actual dir name stored in settings.
}
