% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{get_last_unique_col_suffix}
\alias{get_last_unique_col_suffix}
\title{Get last unique data col value based on suffix}
\usage{
get_last_unique_col_suffix(dg, col_suffix = "_location")
}
\arguments{
\item{dg}{datatable grouped by ID desired to summarise by.}

\item{col_suffix}{The suffix to filter col names by.}
}
\value{
vector of unique values for each group-by ID from last data col
with specified suffixs
}
\description{
Get last unique data col value based on suffix
}
