% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_name_from_file_path}
\alias{get_name_from_file_path}
\title{Get Name from File Name}
\usage{
get_name_from_file_path(projectFilePath, settings)
}
\description{
fileName is PREFIX~_COMP_TITLE.Rmd - returned is COMP_TITLE
}
