% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_next_header_prefix}
\alias{get_next_header_prefix}
\title{get next header prefix from path}
\usage{
get_next_header_prefix(groupNotePath, settings)
}
\arguments{
\item{groupNotePath}{Path to group Note.}

\item{settings}{List from \code{.config/settings.yml} file.}
}
\description{
get next header prefix from path
}
