% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_next_simple_prefix}
\alias{get_next_simple_prefix}
\title{Get Next Simple Prefix}
\usage{
get_next_simple_prefix(projectNotePath, settings)
}
\arguments{
\item{projectNotePath}{Is the path the new simple project note is to be placed
into.}
}
\description{
This method extracts the Prefix String from projectNotePath (isolated by extracting
all letters and numbers up to the \code{projIndexSep}).  It then checks the
Rmd Files in projectNotePath that begin with prefixString to compute the next SIMPLE
Prefix - +1 from the highest prefix value.
}
\details{
NOTE must use Rmd files and not the Project
Note DIRs, as the DIRs can be symlinks, and if not active are not returned as existing!

projectNotePath - must be a path to a VALID DIR for Project Notes - a SubDir in a PROGRAMME DIR.
}
