% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_next_subnote_prefix}
\alias{get_next_subnote_prefix}
\title{Get Next SubNote Prefix}
\usage{
get_next_subnote_prefix(headerNoteDir, settings)
}
\arguments{
\item{headerNoteDir}{The Group Header Note Directory.}

\item{settings}{List from \code{.config/settings.yml} file.}
}
\description{
This method extracts the Prefix String from headerNoteDir, then checks the
DIRs in headerNoteDir that begin with prefixString to compute the next Group
Prefix - +1 from the highest prefix value.
}
