% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{get_org_paths}
\alias{get_org_paths}
\title{Get Existing Organisation Paths}
\usage{
get_org_paths(newOrgPath)
}
\arguments{
\item{newOrgPath}{Character. The directory path of the new organisation.}
}
\value{
Character vector. List of paths to all identified organisations.
}
\description{
Ensures all new organisations are aware of existing ones, and vice versa,
if they reside in the same parent directory.
}
