% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_path_from_link}
\alias{get_path_from_link}
\title{Get Path from Link}
\usage{
get_path_from_link(linkString, linkIndex = 1)
}
\arguments{
\item{linkString}{String containing at least one link.}

\item{linkIndex}{If the string contains more than one link, which link to
get the path from. Default is the first (\code{1}).}
}
\description{
Extracts the path from a markdown link - content between \verb{](} && \verb{)}.
}
