% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_prefix}
\alias{get_prefix}
\title{get prefix from path}
\usage{
get_prefix(filePath, settings)
}
\arguments{
\item{filePath}{Path to Project Note or Doc Rmd file.}

\item{settings}{List from \code{.config/settings.yml} file.}
}
\description{
get prefix from path
}
