% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{get_proj_note_storage}
\alias{get_proj_note_storage}
\title{Get Storage Vector from Contents}
\usage{
get_proj_note_storage(project_note_contents)
}
\description{
returned storage vector is a list of vectors: Each vector is named the directory name,
and the vector contents are each location the directory has been moved/copied to.
}
\details{
Returns a blank list if no content in DATA STORAGE
}
