% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_project_doc_dir_path}
\alias{get_project_doc_dir_path}
\title{Get Project Doc Dir Path}
\usage{
get_project_doc_dir_path(projectDocPath, settings)
}
\description{
Retrieves the projectDoc DIR from the projectDoc path:
}
\details{
\itemize{
\item Removes the projectDocTitle and EXT from the path, and returns
the ProjectDoc Dir - the path plus the projectDoc PREFIX.
}
}
