% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_project_note_dir_path}
\alias{get_project_note_dir_path}
\title{Get Project Note Dir Path}
\usage{
get_project_note_dir_path(projectNoteFilePath, settings)
}
\description{
Retrieves the project Note DIR from the project Note path:
}
