% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_project_note_path}
\alias{get_project_note_path}
\title{Get Project Note Path}
\usage{
get_project_note_path(fileSystemPath, projectNoteName, settings)
}
\description{
Compute a Project Note Path from a fileSystem path and a projectNote Name.
}
\details{
The fileSystemPath must be a DIRECTORY and sub-dir in a PROGRAMME DIRECTORY
\itemize{
\item either a project Document DIR, or another Directory in a PROGRAMME for
storing Project Notes.
}

The NAME of the selected Directory is used to derive the PREFIX for files within this
Directory - it uses all LETTERS and NUMBERS UP TO the first "_" in the directory name.
If no Underscore exists, will use the whole Directory name as the Prefix for the projectNote.
}
