% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_project_note_pathFromDir}
\alias{get_project_note_pathFromDir}
\title{Return the Project Note Rmd Path from a Project DIR Path}
\usage{
get_project_note_pathFromDir(projectNoteDir)
}
\description{
Return the Project Note Rmd Path from a Project DIR Path
}
