% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{get_project_note_paths}
\alias{get_project_note_paths}
\title{convert samples_paths to project note paths}
\usage{
get_project_note_paths(samples_paths)
}
\description{
samples_paths are __sample CSV files inside the Project DIR directly,
so just return the projectnote path from the DIR path - which is the parent
of the samples_apths.
}
