% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_project_prefix_from_name}
\alias{get_project_prefix_from_name}
\title{Get Project Prefix From Name}
\usage{
get_project_prefix_from_name(projectFileName, settings)
}
\description{
Extracts the Project Prefix from the projectName - all characters before "~_"
}
