% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_protocol_description}
\alias{get_protocol_description}
\title{Get Protocol Description}
\usage{
get_protocol_description(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the protocol with startSep at startSepLineIndex.
}
