% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_protocol_header}
\alias{get_protocol_header}
\title{Get Protocol Header}
\usage{
get_protocol_header(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the protocol with startSep at startSepLineIndex.
}
