% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_protocol_title}
\alias{get_protocol_title}
\title{Get Protocol Title}
\usage{
get_protocol_title(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the protocol with startSep at startSepLineIndex.
}
