% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_relative_path_org}
\alias{get_relative_path_org}
\title{From an absolute path compute the relative path from the Organisation Root.}
\usage{
get_relative_path_org(path)
}
\description{
From an absolute path compute the relative path from the Organisation Root.
}
