% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{get_reordered_rstudio_doc_paths}
\alias{get_reordered_rstudio_doc_paths}
\title{Reorder open RStudio documents with active path used as a pivot}
\usage{
get_reordered_rstudio_doc_paths(
  active_path,
  get_doc_ids = get_rstudio_open_doc_IDs
)
}
\arguments{
\item{active_path}{The full path of the active document}

\item{get_doc_ids}{A function that returns the open document IDs/paths (injectable for testing)}
}
\value{
A character vector of reordered file paths
}
\description{
Returns the list of file paths in order: right of active doc (L to R), then left.
}
