% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_config_dir}
\alias{get_rstudio_config_dir}
\title{Get RStudio Config Dir}
\usage{
get_rstudio_config_dir()
}
\description{
Returns this from the settings or the default depending on OS.type (unix or Windows)
}
\details{
Returns config directory depending on OS.type (unix or Windows) & rstudio
version (1.2 and prior OR 1.3+)
\itemize{
\item 1.3+ unix : ~/.config/rstudio/keybindings/rstudio_bindings.json
}

This is where key configuration files are kept - eg. keybindings

See :

https://support.posit.co/hc/en-us/articles/206382178
'Customizing Keyboard Shortcuts in the RStudio IDE'
}
