% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_doc_list_filtered_reordered}
\alias{get_rstudio_doc_list_filtered_reordered}
\title{RStudio Docs filtered and Reordered}
\usage{
get_rstudio_doc_list_filtered_reordered(filePath, settings)
}
\description{
Filtered to ensure only files that are in the \code{filePath} orgPath and are
project notes (not unknown files or project docs).
}
\details{
Reordered to list all files from the right of the selected file (\code{filePath}).
}
