% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{get_rstudio_docs_list_numbered}
\alias{get_rstudio_docs_list_numbered}
\title{Get numbered list of open docs (for selectInput)}
\usage{
get_rstudio_docs_list_numbered(active_path)
}
\arguments{
\item{active_path}{Full path of currently active document}
}
\value{
A named list of integer indices, names are file basenames
}
\description{
Get numbered list of open docs (for selectInput)
}
