% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{get_rstudio_docs_list_reordered}
\alias{get_rstudio_docs_list_reordered}
\title{Get reordered list of open doc paths (excluding active)}
\usage{
get_rstudio_docs_list_reordered(active_path)
}
\arguments{
\item{active_path}{Full path of currently active document}
}
\value{
A character vector of reordered paths (right of active, then left)
}
\description{
Get reordered list of open doc paths (excluding active)
}
