% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_internal_state_dir}
\alias{get_rstudio_internal_state_dir}
\title{Get RStudio Internal State Dir}
\usage{
get_rstudio_internal_state_dir()
}
\description{
Returns this from the settings or the default depending on OS.type (unix or Windows)
}
\details{
https://support.posit.co/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
'Resetting RStudio Desktop's State'

Accessing the RStudio-Desktop Directory (Internal State)

RStudio Desktop stores its internal state in a hidden directory: includes
information about open documents, log files, and other state information
}
