% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_internal_state_dir2}
\alias{get_rstudio_internal_state_dir2}
\title{Get RStudio Internal State Dir}
\usage{
get_rstudio_internal_state_dir2(path)
}
\description{
Returns this from the settings or the default depending on OS.type (unix or Windows)
}
