% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_open_doc_IDs}
\alias{get_rstudio_open_doc_IDs}
\title{Extract IDs and paths of currently open RStudio documents}
\usage{
get_rstudio_open_doc_IDs()
}
\value{
A list of character vectors where each entry contains the file ID and
its associated path (or "null" if unsaved).
}
\description{
Uses internal RStudio session state files to extract a list of open documents.
This function is useful for inspecting session metadata for all open files.
}
