% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_settings_yml_file}
\alias{get_settings_yml_file}
\title{Get settings yml file}
\usage{
get_settings_yml_file(orgPath)
}
\description{
This is set as \code{.config/settings.yml} GLOBALLY - stores all settings for
projectmanagr organisation.
}
