% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_site_dir}
\alias{get_site_dir}
\title{Get site dir}
\usage{
get_site_dir(orgPath, settings)
}
\description{
Stores html site of org.  This is set as \verb{site/} by default, stored in
settings.
}
