% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_status_yml}
\alias{get_status_yml}
\title{Get status yml}
\usage{
get_status_yml(orgPath, settings)
}
\description{
This is stored in \code{.config} GLOBALLY - filename is derived from settings
in \code{ConfigStatusYamlFile} parameter
}
