% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_sub_note_path}
\alias{get_sub_note_path}
\title{Get Project Sub Note Path}
\usage{
get_sub_note_path(
  headerNoteDir,
  projectSubNoteName,
  projectHeaderNoteName,
  settings
)
}
\description{
Compute a Project Sub Note Path from a fileSystem path and a projectNote Name.
}
\details{
The headerNoteDir must be the directory of a header note.

The NAME of the selected Directory is used to derive the PREFIX for files within this
Directory - it uses all LETTERS and NUMBERS UP TO the first "_" in the directory name.
If no Underscore exists, will use the whole Directory name as the Prefix for the projectNote.
}
