% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_task_number}
\alias{get_task_number}
\title{get task number}
\usage{
get_task_number(task, settings)
}
\description{
\code{task} as collected from selection[\link{"task"}] - fitting the format specified
in settings[\link{"ProjectTaskHeader"}].
}
