% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{get_task_title}
\alias{get_task_title}
\title{get task title}
\usage{
get_task_title(task, settings)
}
\description{
\code{task} as collected from selection[\link{"task"}] - fitting the format specified
in settings[\link{"ProjectTaskHeader"}].
}
