% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_template_dir}
\alias{get_template_dir}
\title{Get template dir}
\usage{
get_template_dir(orgPath)
}
\description{
This is set as \code{.config/templates} GLOBALLY - used to identify
a Project Organisation root directory.
}
