% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_tz_gmt_offset}
\alias{get_tz_gmt_offset}
\title{get timezone offset from GMT}
\usage{
get_tz_gmt_offset(tz = get_locale())
}
\value{
timezone offset from GMT. eg. if in BST will return \code{+1}
}
\description{
get timezone offset from GMT
}
