% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_tz_info}
\alias{get_tz_info}
\title{get timezone info}
\usage{
get_tz_info(date, tz = get_locale())
}
\value{
the timezone abbreviation plus offset from GMT. eg. BST (GMT+1)
}
\description{
Using a \code{tz} (or locale, such as \code{Europe/London}), get the actual timezone
info for a given date.  This will return the actual timezone, including
taking into account any daylight savings time, for a given locale and date.
}
\details{
\code{Sys.timezone()} can be used to retrieve the current system timezone/locale,
an internal projectmanagr function returns this - so its can be mocked
during testing!
}
