% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_volumes_dir}
\alias{get_volumes_dir}
\title{Get volumes dir}
\usage{
get_volumes_dir(orgPath, settings)
}
\description{
Stores volume mounts. This is set as \verb{volumes/} by default, stored in
settings.
}
