% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_weekly_journal_dir}
\alias{get_weekly_journal_dir}
\title{Get weekly journal dir}
\usage{
get_weekly_journal_dir(orgPath, settings)
}
\description{
Stores weekly journal Rmd files.  This is set as \verb{weekly-journal/} by default,
stored in settings.
}
