% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileType.R
\name{getFileType}
\alias{getFileType}
\title{Get File Type}
\usage{
getFileType(fileSystemPath)
}
\arguments{
\item{fileSystemPath}{the Absolute path to the file.}
}
\description{
Determine if the current file is a PROJECT DOC (it is inside the PROJECTS/ Directory),
a HEADER GROUP NOTE (contains the string "-00~_"), or a SIMPLE or SUB Project Note (any
other file).
}
\details{
Returns the relevant String:  DOC, HEAD, NOTE
}
