% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getHeaderNotePath}
\alias{getHeaderNotePath}
\title{Get Project Header Note Path}
\usage{
getHeaderNotePath(fileSystemPath, projectHeaderNoteName)
}
\description{
Compute a Project Header Note Path from a fileSystem path and a projectNote Name.
}
\details{
The fileSystemPath must be a DIRECTORY and sub-dir in a PROGRAMME DIRECTORY
- either a project Document DIR, or another Directory in a PROGRAMME for
storing Project Notes.

The NAME of the selected Directory is used to derive the PREFIX for files within this
Directory - it uses all LETTERS and NUMBERS UP TO the first "_" in the directory name.
If no Underscore exists, will use the whole Directory name as the Prefix for the projectNote.
}
