% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getNameFromFileName}
\alias{getNameFromFileName}
\title{Get Project TiNametle From File Name}
\usage{
getNameFromFileName(fileName)
}
\description{
fileName is PREFIX~_COMP_TITLE.Rmd - returned is COMP_TITLE
}
