% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNextGroupPrefix.R
\name{getNextGroupPrefix}
\alias{getNextGroupPrefix}
\title{Get Next Group Prefix}
\usage{
getNextGroupPrefix(headerNotePath)
}
\description{
This method extracts the Prefix String from headerNotePath, then checks the
DIRs in headerNotePath that begin with prefixString to compute the next Group
Prefix - +1 from the highest prefix value.
}
\details{
headerNotePath - must be the path to the headerNote - DIR or RMD file.
}
