% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getPrefixFromFileName}
\alias{getPrefixFromFileName}
\title{Get Prefix From File Name}
\usage{
getPrefixFromFileName(fileName)
}
\description{
fileName is PREFIX~_COMP_TITLE.Rmd - returned is PREFIX
}
