% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getProjCompTitle}
\alias{getProjCompTitle}
\title{Get Project Comp title from Project Comp contents}
\usage{
getProjCompTitle(contents)
}
\description{
Get Project Comp title from Project Comp contents
}
