% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getProjectDocDirPath}
\alias{getProjectDocDirPath}
\title{Get Project Doc Dir Path}
\usage{
getProjectDocDirPath(projectDocPath)
}
\description{
Retrieves the projectDoc DIR from the projectDoc path:
}
\details{
- Removes the projectDocTitle and EXT from the path, and returns
the ProjectDoc Dir - the path plus the projectDoc PREFIX.
}
