% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getProjectNoteDirPath}
\alias{getProjectNoteDirPath}
\title{Get Project Note Dir Path}
\usage{
getProjectNoteDirPath(projectDocPath)
}
\description{
Retrieves the project Note DIR from the project Note path:
}
