% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectNotePath.R
\name{getProjectNotePath}
\alias{getProjectNotePath}
\title{Get Project Note Path}
\usage{
getProjectNotePath(fileSystemPath, projectNoteName)
}
\description{
Assumes fileSystemPath is the projectDocPath, a Sub-Dir to a PROGRAMME DIRECTORY.
}
