% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getProjectNotePathFromDir}
\alias{getProjectNotePathFromDir}
\title{Return the Project Note Rmd Path from a Project DIR Path}
\usage{
getProjectNotePathFromDir(projectNoteDir)
}
\description{
Return the Project Note Rmd Path from a Project DIR Path
}
