% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateProjectOrg.R
\name{getProjectNoteSummaryTodoList}
\alias{getProjectNoteSummaryTodoList}
\title{Get Project Note Summary and Todo Link List}
\usage{
getProjectNoteSummaryTodoList(
  contents,
  projectNotePath,
  todoHeader,
  goalHeader,
  delHeader,
  taskHeader
)
}
\description{
Returns a list of VECTORS:
}
\details{
list[[i]] : A VECTOR that includes the Project Doc ABSOLUTE Link, GOAL Num, DEL Num, TASK Num.

 list[[i]][[1]] : Project Doc ABSOLUTE Link

 list[[i]][[2]] : Project Doc GOAL - number plus title

 list[[i]][[3]] : Project Doc DELIVERABLE - number plus title

 list[[i]][[4]] : Project Doc TASK - number plus title

 list[[i]][[5]] : Vector containing all summary information in Project Note for this
                  Project Doc GOAL/DEL/TASK
 list[[i]][[6]] : Vector containing all TODO information in Project Note for this
                  Project Doc GOAL/DEL/TASK

 length( list ) returns the number of Project Doc links in the returned list.

 @param projectNoteContents Character vector containing the contents of a Project Note, which includes links to
 at least one ProjectDoc, separated by "----", and ended with "------".

 @param projectNotePath the FULL PATH to the project note, from which the contents is derived.

 @param todoHeader The header of any TODO section that may be defined under the project note summary.  This Header
 and remaining content of a summary should be saved SEPARATELY to the summaryVector.
}
