% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getProjectPath}
\alias{getProjectPath}
\title{Get Project Path}
\usage{
getProjectPath(fileSystemPath, projectName)
}
\description{
Assumes fileSystemPath is a PROGRAMME DIRECTORY, containing a 'PROJECTS/' Directory.
}
