% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getProjectPrefixFromPath}
\alias{getProjectPrefixFromPath}
\title{Get Project Prefix From Path}
\usage{
getProjectPrefixFromPath(projectPath)
}
\description{
Extracts the Project Prefix from the projectPath - all characters before "~_"
}
