% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getProtocolSummary}
\alias{getProtocolSummary}
\title{Get Protocol Summary}
\usage{
getProtocolSummary(projNotePath, protocolName)
}
\arguments{
\item{projNotePath}{a path to a valid Project Note.}

\item{protocolName}{the name of a valid protocol in SOP/ in programme relative to projNotePath}
}
\description{
Extracts the Protocol Summary from protocolName in SOP/ relative to projNotePath.
}
