% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{getRStudioInternalStateDir}
\alias{getRStudioInternalStateDir}
\title{Get RStudio Internal State Dir}
\usage{
getRStudioInternalStateDir()
}
\description{
Returns this from the settings or the default depending on OS.type (unix or Windows)
}
