% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{getRStudioInternalStateDir2}
\alias{getRStudioInternalStateDir2}
\title{Get RStudio Internal State Dir}
\usage{
getRStudioInternalStateDir2(path)
}
\description{
Returns this from the settings or the default depending on OS.type (unix or Windows)
}
