% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{getSubNotePathFromDoc}
\alias{getSubNotePathFromDoc}
\title{Get Sub Note Path form Project Doc}
\usage{
getSubNotePathFromDoc(projectDocPath, headerNoteRelPath, subNoteName)
}
\description{
Assumes projectDocPath is the path to the Project Doc.
}
\details{
The headerNoteRelPath is assumed to be the RELATIVE path from projectDocPath
to the headerNote.
}
