% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getTaskSectionHeader}
\alias{getTaskSectionHeader}
\title{Get Task Section Header}
\usage{
getTaskSectionHeader(
  orgPath,
  taskSectionHeaderTemplate = "Task-Section-Header.Rmd"
)
}
\description{
Retrieves the Task Section Header from template: Task-Section-Header.Rmd
}
