% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{grep_line_index_from}
\alias{grep_line_index_from}
\title{Grep Line Index from initialIndex to END of contents}
\usage{
grep_line_index_from(line, contents, initialIndex)
}
\arguments{
\item{line}{a String to be found}

\item{contents}{a vector of strings, the index of the first instance of line in
this vector is returned.}
}
\description{
Returns the index of the first PARTIAL MATCH (using grep) of line in contents,
from initialIndex, if not found returns -1.
}
