% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{grepLineIndex}
\alias{grepLineIndex}
\title{Grep Line Index}
\usage{
grepLineIndex(line, contents)
}
\arguments{
\item{line}{a String to be found}

\item{contents}{a vector of strings, the index of the first instance of line in
this vector is returned.}
}
\description{
Returns the index of the first PARTIAL MATCH (using grep) of line in contents,
if not found returns -1.
}
