% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{grepLineIndexFrom}
\alias{grepLineIndexFrom}
\title{Grep Line Index from initialIndex}
\usage{
grepLineIndexFrom(line, contents, initialIndex)
}
\arguments{
\item{line}{a String to be found}

\item{contents}{a vector of strings, the index of the first instance of line in
this vector is returned.}
}
\description{
Returns the index of the first PARTIAL MATCH (using grep) of line in contents,
from initialIndex, if not found returns -1.
}
