% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{id_rep_error_ass}
\alias{id_rep_error_ass}
\title{Check Errors in IDs + REPs}
\usage{
id_rep_error_ass(dt, col_name, IDsREPs, errorString, sep = ".-.")
}
\description{
\enumerate{
\item Ensure assigned data in dt:col_name does not clash with IDREPs
}
}
