% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{id_rep_error_dup_div}
\alias{id_rep_error_dup_div}
\title{Check Errors in IDs + REPs}
\usage{
id_rep_error_dup_div(IDsREPs, sep, dt, table_name, table_start_line)
}
\description{
\enumerate{
\item Ensure no duplicated ID+REP identifier.
}
}
\details{
\enumerate{
\item Ensure no ID+REP identifier has been previously resampled, disposed, exported.
}
}
