% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-utils-functions.R
\name{insert_at_indices}
\alias{insert_at_indices}
\title{Insert vector at indices in contents}
\usage{
insert_at_indices(templateContents, indices, replacementVector)
}
\arguments{
\item{templateContents}{Character vector of lines}

\item{indices}{Integer vector of line numbers at which to insert}

\item{replacementVector}{Character vector to insert}
}
\value{
Modified character vector with insertions
}
\description{
Inserts \code{replacementVector} into \code{templateContents} \emph{before} each line number
in \code{indices}.
Line numbers must be between 1 and \code{length(templateContents) + 1}.
}
