% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_doc_deliverable_section}
\alias{insert_doc_deliverable_section}
\title{Append a Deliverable Section in Project Doc}
\usage{
insert_doc_deliverable_section(
  selection,
  deliverableSectionTemplate = "Project-Doc-Section-Deliverable-Template.Rmd"
)
}
\arguments{
\item{selection}{Selection object from the Project Doc file where the
deliverable section is to be inserted.  Use
`projectmanagr::cursor_selection()` or `projectmanagr::user_selection()` to
create this object.}

\item{deliverableSectionTemplate}{Template file that contains boilerplate
content for deliverable section insertion into project doc.}
}
\description{
Appends a new Deliverable Section in a Project Doc based on the cursor
position.  Identifies the current Goal segment as selected by the cursor
position, and inserts a new deliverable section template into this segment,
incremented from the last Deliverable Section number.
}
\details{
This function will only work if the selection is in a Project Doc.  The
deliverable section will be placed in correction position and with correct
spacing according to the `deliverableSectionTemplate`.
}
