% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_doc_goal_section}
\alias{insert_doc_goal_section}
\title{Append a Goal Section to Project Doc}
\usage{
insert_doc_goal_section(
  selection,
  goalSectionTemplate = "Project-Doc-Section-Goal-Template.Rmd"
)
}
\arguments{
\item{selection}{Selection object from the Project Doc file where the goal
section is to be inserted.  Use \code{projectmanagr::cursor_selection()} or
\code{projectmanagr::user_selection()} to create this object.}

\item{goalSectionTemplate}{Template file that contains boilerplate content
for goal section insertion into project doc}
}
\description{
Appends a new Goal Section to a Project Doc.  Searches through the Project
Doc contents to identify the last Goal Section, and inserts a new goal
section template, incremented from the last Goal Section number.
}
\details{
This function will only work if the selection is in a project doc.  The goal
section will be placed in correction position and with correct spacing
according to the \code{goalSectionTemplate}.
}
