% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_doc_task_section}
\alias{insert_doc_task_section}
\title{Append a Task Section in Project Doc}
\usage{
insert_doc_task_section(
  selection,
  taskSectionTemplate = "Project-Doc-Section-Task-Template.Rmd"
)
}
\arguments{
\item{selection}{Selection object from the Project Doc file where the
task section is to be inserted.  Use
\code{projectmanagr::cursor_selection()} or \code{projectmanagr::user_selection()} to
create this object.}

\item{taskSectionTemplate}{Template file that contains boilerplate
content for task section insertion into project doc.}
}
\description{
Appends a new Task Section in a Project Doc based on the cursor
position.  Identifies the current Goal segment as selected by the cursor
position, and inserts a new Task section template into this segment,
incremented from the last Deliverable Section number.
}
\details{
This function will only work if the selection is in a Project Doc.  The
task section will be placed in correction position and with correct
spacing according to the \code{taskSectionTemplate}.
}
