% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-functions.R
\name{insert_dt_to_contents}
\alias{insert_dt_to_contents}
\title{Insert data_tables vector into rmd_contents, between rmd_startline & rmd_endline}
\usage{
insert_dt_to_contents(data_tables, rmd_contents, rmd_startline, rmd_endline)
}
\description{
Removes content between \code{rmd_startline} and \code{rmd_endline}
}
