% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_header_link_subnote}
\alias{insert_header_link_subnote}
\title{Insert header link into subnote}
\usage{
insert_header_link_subnote(
  subNoteContents,
  headerNoteFileName,
  headerNoteRmdPath,
  subNoteRmdPath,
  headerNoteContentLinkContents,
  settings,
  orgPath
)
}
\description{
Insert header link into subnote
}
