% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_lines}
\alias{insert_lines}
\title{Insert lines into a Project Note}
\usage{
insert_lines(selection, lines)
}
\arguments{
\item{selection}{Selection object from Project Note.}

\item{lines}{character vector of lines to insert.}
}
\description{
This Function adds `lines` to a Project Note.
}
