% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_protocol}
\alias{insert_protocol}
\title{Insert a Protocol into a Project Note}
\usage{
insert_protocol(
  protocolName,
  selection,
  protocolInsertionTemplate = "Protocol-Insertion-Template.Rmd"
)
}
\arguments{
\item{protocolName}{The name of the Protocol Rmd file - WITHOUT extension.}

\item{selection}{Selection object from the Project Note file where the protocol
is to be inserted.  Use `projectmanagr::cursor_selection()` or
`projectmanagr::user_selection()` to create this object.}

\item{protocolInsertionTemplate}{Template file that contains boilerplate content
for protocol insertion into project note.}
}
\description{
This Function adds a Protocol to a Project Note - protocols are found in the
 PROGRAMMES' protocols/ Dir.
}
\details{
Inserted protocols add each Procedure defined under the '# PROTOCOL' section
 of the document. Notes sections (FALSE by default) and the Equipment section
(TRUE by default) can be optionally included in the inserted protocol.

All links in the Protocol are UPDATED to work from the destination Project
 Note.

All graphics included in a knitr::include_graphics() r code chunk are
 transferred to the new Project Note DIR and linked appropriately.
}
