% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{insert_subnote_header_link}
\alias{insert_subnote_header_link}
\title{Insert header link content into subnote}
\usage{
insert_subnote_header_link(
  subNoteContents,
  headerNoteFileName,
  headerNoteRmdPath,
  subNoteRmdPath,
  headerNoteContentLinkContents,
  settings,
  orgPath
)
}
\description{
Insert header link content into subnote
}
