% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_subnote_link_header}
\alias{insert_subnote_link_header}
\title{Insert header link into subnote}
\usage{
insert_subnote_link_header(
  headerNoteRmdContents,
  subNoteFileName,
  subNoteRmdPath,
  headerNoteRmdPath,
  subNoteContentLinkContents,
  settings,
  orgPath
)
}
\description{
Insert header link into subnote
}
