% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertProtocol.R
\name{insertProtocol}
\alias{insertProtocol}
\title{Insert a Protocol into a Project Note}
\usage{
insertProtocol(projectNotePath, projectNoteRow, protocolName,
  includeNotes = FALSE, includeEquip = TRUE)
}
\arguments{
\item{projectNotePath}{The ABSOLUTE path of the Project Note.}

\item{projectNoteRow}{Index to insert the protocol into project notes}

\item{protocolName}{The name of the Protocol Rmd file}

\item{includeNotes}{Boolean to indicate whether all Notes bullets are included or not.}

\item{includeEquip}{Boolean to indicate whether the initial equipment and material section should be included.}
}
\description{
This Function adds a new Protocol to a Project Note - saved in the PROGRAMMES' SOP/ Dir.
The Protocol is formed using a Rmd template, and uses the tufte::tufte_handout format to
generate a PDF of the protocol.
}
\details{
Protocols are stored in the SOP/ directory inside the PROGRAMME Directory.  Each Protocol
exists in its own directory, to keep its compiled files together.

The Protocol source Rmd will link to its creating Project Note, and the Project Note will link to the
compiled PDF of the Protocol.
}
