% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{interpret_todo_flags}
\alias{interpret_todo_flags}
\title{Interpret schedule/deadline => set flags (passed, due, etc.)}
\usage{
interpret_todo_flags(schedule, deadline, today)
}
\arguments{
\item{schedule}{A string in the form \code{"YYYY-MM-DD:HHmm"} if scheduling is set.}

\item{deadline}{A string in the form \code{"YYYY-MM-DD"} if there's a deadline.}

\item{today}{A string representing "today," e.g. "2025-04-01."}
}
\value{
A list with \code{isScheduledToday}, \code{isSchedulePassed},
\code{isDueToday}, \code{isDeadlinePassed}.
}
\description{
Interpret schedule/deadline => set flags (passed, due, etc.)
}
\keyword{internal}
