% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{link_add_section}
\alias{link_add_section}
\title{Add Section to Link}
\usage{
link_add_section(NoteLink, toFileSection)
}
\description{
\code{NoteLink} is just a path to a file to link to.
\code{toFileSection} is the Markdown Header of a section in NoteLink
}
\details{
This function adds the Markdown Header to end of NoteLink, using html link
syntax : using a \verb{#} then all words from markdown header separate by \code{-}

This link to a section of R markdown file can be navigaed in markdown using the
\code{navigate_markdown_link()} function in projectmanagr
}
