% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{link_doc_group_note}
\alias{link_doc_group_note}
\title{Link Project Document GDT to Project Note Group (Header plus all Sub Notes)}
\usage{
link_doc_group_note(
  selection,
  headerNoteRmdPath,
  projNoteLinkTemplate = "Project-Note-Link-Template.Rmd",
  projNoteLinkSummaryTemplate = "Project-Note-Link-Summary-Template.Rmd",
  todoTemplate = "Todo-Template.Rmd",
  projNoteSummaryTemplate = "Project-Note-Summary-Template.Rmd",
  subNoteSummaryTemplate = "Project-Sub-Note-Summary-Template.Rmd"
)
}
\arguments{
\item{selection}{List containing the Goal, Del, Task selected from the Project
Doc, as well as other useful information - lines of Task/Del/Goal, projectDoc
path content of selection line.  See `cursor_selection()` or `user_selection()`.}

\item{headerNoteRmdPath}{Path to existing Group Header Note Rmd, in same Org
as the Project Doc `selection` object.}

\item{projNoteLinkTemplate}{Template with structure to add the Project Doc
Goal/Del/Task link in all Project Notes.  Includes Project Doc title, link,
and then links to Goal / Del / Task, plus a Summary Info section (filled
with the content from `taskSectionHeaderTemplate`).}

\item{projNoteLinkSummaryTemplate}{Template with structure to add underneath the
Project Doc Goal/Del/Task link in all Sub Note.  Includes a 'summary' section
and a 'todo' section by default, linked to `NoteObjectivesSummarySectionHeader`
& `NoteObjectivesTodoSectionHeader` in `config/settings.yml`}

\item{projNoteSummaryTemplate}{Template with structure to add Sub Note
summaries to Project Doc under Goal/Del/Task.}

\item{subNoteSummaryTemplate}{Template with structure to add Project Sub Note
summary to Project Doc under Goal/Del/Task.}
}
\description{
Add a cross-reference between a Project Doc at the Goal-Del-Task as shown in
`selection` with the group project note, with header note at
`headerNoteRmdPath`, using the provided templates.
}
