% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{link_group_note_doc}
\alias{link_group_note_doc}
\title{Link Group Note to Project Doc GDT}
\usage{
link_group_note_doc(
  selection,
  settings,
  headerNoteRmdPath,
  headerNoteContents,
  headerNoteLinkContents,
  projNoteLinkContents,
  projNoteLinkSummaryContents,
  projNoteSummaryContents,
  subNoteSummaryContents,
  projDocContents,
  orgPath
)
}
\description{
Single internal function to perform the linking between a project doc at GDT
and a group note set.
}
