% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{link_group_note_in_doc}
\alias{link_group_note_in_doc}
\title{Link a Newly Created Group Note (Header + Sub) into the Project Doc}
\usage{
link_group_note_in_doc(
  selection,
  settings,
  headerRmdPath,
  headerRmdContents,
  subRmdPath,
  subRmdContents,
  templates,
  docPath,
  docContents,
  orgPath,
  addObjToHeader = TRUE
)
}
\arguments{
\item{selection}{A list describing the user’s selection in the doc:
must include \code{rmdType} = "DOC", \code{filePath}, \code{taskLine}, etc.}

\item{settings}{A list of settings from \code{.config/settings.yml}.}

\item{headerRmdPath}{Absolute file path to the newly created \emph{header} Rmd.}

\item{headerRmdContents}{The lines of that newly created header note.}

\item{subRmdPath}{Absolute file path to the newly created \emph{sub-note} Rmd.}

\item{subRmdContents}{The lines of that newly created sub-note.}

\item{templates}{A list of template strings for linking (e.g.
\code{headerNoteLinkContents}, \code{subNoteLinkContents}, \code{projNoteLinkContents}, etc.).}

\item{docPath}{The absolute file path to the \strong{project doc} (the G/D/T Rmd).}

\item{docContents}{The lines of the project doc, already read in.}

\item{orgPath}{The root org directory, used if you need to re-read
or do \code{sub_template_param} calls.}

\item{addObjToHeader}{Logical. If \code{TRUE}, link to the header note only;
otherwise link to the sub-note only.}
}
\value{
A logical, \code{TRUE} if linking succeeded, or \code{FALSE} if it failed.
}
\description{
This function wraps the old branching logic:
\itemize{
\item If \code{addObjToHeader = TRUE}, we attach the project doc to the \emph{header note}
(the “header” Rmd). The old code used \code{link_group_note_doc()} for that.
\item Otherwise, we link the doc to the \emph{sub-note} Rmd only
(the old code used \code{link_project_note_doc()}).
}
}
\details{
We assume these two linking functions (\code{link_group_note_doc} and
\code{link_project_note_doc}) already exist in your codebase and return a
boolean: \code{TRUE} if the link was formed successfully, \code{FALSE} if
linking failed (because the link already exists, etc.).
}
