% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{link_prog_to_org_index}
\alias{link_prog_to_org_index}
\title{Link Programme Index to Project Org Index}
\usage{
link_prog_to_org_index(
  progPath,
  programmeName,
  progContents,
  tempPath,
  progSummaryTemplate,
  orgPath,
  settings
)
}
\description{
Fill with default Programme Summary - updated with projectmanagr::update()
}
