% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{link_project_note_doc}
\alias{link_project_note_doc}
\title{Link Project Note to Project Doc GDT}
\usage{
link_project_note_doc(
  selection,
  settings,
  projNoteRmdPath,
  projNoteRmdContents,
  projNoteLinkContents,
  projNoteLinkSummaryContents,
  todoContents,
  projNoteSummaryContents,
  projDocContents,
  orgPath
)
}
\description{
Single internal function to perform the linking between a project doc at GDT
and a simple project note.
}
