% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{link_sect_to_parent_index}
\alias{link_sect_to_parent_index}
\title{Link Programme Section Index to Parent Index}
\usage{
link_sect_to_parent_index(
  sectPath,
  sectionName,
  sectContents,
  parentIndex,
  tempPath,
  sectSummaryTemplate,
  orgPath,
  settings
)
}
\description{
Fill with default Programme Section Summary - updated with
\code{projectmanagr::update()}
}
