% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{list_open_documents}
\alias{list_open_documents}
\title{List the open documents as a tibble}
\usage{
list_open_documents(get_open = get_rstudio_open_doc_IDs)
}
\description{
\emph{Assumes} you already have (or can write) \code{get_rstudio_open_doc_IDs()}
returning \verb{list(list(id, path), …)}.
}
\keyword{internal}
