% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSubNotePaths.R
\name{listSubNotePaths}
\alias{listSubNotePaths}
\title{List all SubNotes that are below the Header Note.}
\usage{
listSubNotePaths(headerNotePath)
}
\arguments{
\item{headerNotePath}{The path to the Header Note.  This MUST be in a sub-directory or lower inside a
PROGRAMME Directory.  This MUST be an ABSOLUTE PATH.}
}
\description{
This Function returns a CHARACTER VECTOR of absolute paths to each subnote under the header note at headerNotePath.
}
