% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{load_param_vector}
\alias{load_param_vector}
\title{Load `paramContents` (a param from settings.yml) correctly}
\usage{
load_param_vector(paramContents, orgPath)
}
\description{
Some params start with keyword `template:` which indicates the relative path
from the templatesDir to the template file to be loaded.
}
\details{
Else `paramContents` is returned unchanged.
}
