% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{load_yml_settings_default}
\alias{load_yml_settings_default}
\title{Load YAML Settings with Defaults}
\usage{
load_yml_settings_default(settingsYamlPath, projectmanagrPath)
}
\arguments{
\item{settingsYamlPath}{Character. Path to a custom YAML settings file. If empty,
the default projectmanagr settings are used.}

\item{projectmanagrPath}{Character. The file path of the projectmanagr package.}
}
\value{
List. The parsed settings from the YAML file.
}
\description{
Loads a YAML settings file, either from a provided path or from the default
settings in the projectmanagr package.
}
