% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{make_abs}
\alias{make_abs}
\title{Make paths absolute relative to organisation root}
\usage{
make_abs(paths, orgPath)
}
\arguments{
\item{paths}{Character vector of paths.}

\item{orgPath}{Organisation root.}
}
\value{
Absolute, normalised paths.
}
\description{
Make paths absolute relative to organisation root
}
\keyword{internal}
