% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{match_line_index}
\alias{match_line_index}
\title{return index of line that matches line in contents}
\usage{
match_line_index(line, contents)
}
\description{
return index of line that matches line in contents
}
