% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{match_vector}
\alias{match_vector}
\title{Match vector in parent vector}
\usage{
match_vector(vector, parent, nomatch = 0L)
}
\description{
Identifies each initial index of `vector` in `parent` where all elements in
`vector` match elements in `parent` in order.
}
