% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{match_vector_sub}
\alias{match_vector_sub}
\title{Match vector is subset in parent}
\usage{
match_vector_sub(vector, parent, nomatch = 0L)
}
\description{
Identifies each initial index where each consecutive element of \code{vector} is
present in a consecutive set of elements in \code{parent} - where all elements in
\code{vector} match parts of elements in \code{parent} in order.
}
