% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{matchLineIndex}
\alias{matchLineIndex}
\title{Match Line Index}
\usage{
matchLineIndex(line, contents)
}
\arguments{
\item{line}{a String to be found - any leading/trailing whitespace is TRIMMED.}

\item{contents}{a vector of strings, the index of the first instance of line in
this vector is returned - any leading/trailing whitespace is TRIMMED from each line.}
}
\description{
Returns the index of the first COMPLETE MATCH of line in contents,
if not found returns -1.
}
