% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{modify_timezone}
\alias{modify_timezone}
\title{Modify Timezone of a Datetime String}
\usage{
modify_timezone(datetimeString, newTimeZone)
}
\arguments{
\item{datetimeString}{Datetime string in format 'YYYY-MM-DD:hhmmZ'.}

\item{newTimeZone}{Single-letter code of the new timezone.}
}
\value{
Modified datetime string with new timezone code.
}
\description{
Takes a datetime string in format 'YYYY-MM-DD:hhmmZ' and converts it to a
new timezone using a single-letter timezone code.
}
