% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-functions.R
\name{move_dir}
\alias{move_dir}
\title{Move a Directory to new location}
\usage{
move_dir(dirPath, destinationDir)
}
\arguments{
\item{dirPath}{Path to directory to be moved.  Must not be the ORG root or
a PROGRAMME directory.}

\item{destinationDir}{Path to destination directory.  Must be in same
Organisation.  If the directory does not exist, it will be created.}
}
\description{
This Function moves a directory containing Project Note(s) \link{simple, group}
or Project Doc(s) to a new location.
}
\details{
Updates all GDT links and hyperlinks across organisation.
}
