% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-functions.R
\name{move_file}
\alias{move_file}
\title{Move a File to new Directory}
\usage{
move_file(filePath, destinationDir)
}
\arguments{
\item{filePath}{Path to file to be moved.  Must be a Project Note
\link{simple, group} or Project Doc.}

\item{destinationDir}{Path to destination directory.  Must be in same
Organisation.  If the directory does not exist, it will be created.}
}
\description{
This Function moves a File (Project Note \link{simple, group} or Project Doc) to
a new directory in the same organisation.
}
\details{
\itemize{
\item Moves the Dir and all contents to new location
\item Updates all GDT links and hyperlinks
\itemize{
\item The moved project note/doc
\item Any project notes contained in the Dir
}
}
}
