% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{moveToVolume}
\alias{moveToVolume}
\title{Move Project Note DIR to Volume}
\usage{
moveToVolume(volName, projectNotePath)
}
\description{
Creates the Project Note DIR in new volume, copies all contents from
the old Project Note DIR location into the new Project Note DIR location,
deletes the old Project Note DIR, and symlinks to the new Project Note DIR.
}
\details{
Finally, changes the location of the volume in the Project Note itself.

 NO LONGER export - deprecated
}
