% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{note_link_todo_params}
\alias{note_link_todo_params}
\title{Substitute Note Link Summary in Contents}
\usage{
note_link_todo_params(projNoteLinkSummaryContents, settings, orgPath)
}
\description{
Substitute Note Link Summary in Contents
}
