% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{note_summary_headers_quote_in}
\alias{note_summary_headers_quote_in}
\title{quote out all Markdown Headers}
\usage{
note_summary_headers_quote_in(summary)
}
\description{
Converts every line beginning with \verb{>#} in \code{summary} to \verb{#}
}
\details{
This is used to quote-in the markdown headers in Project Note Summary Section,
so it forms html headers when rendered to html.

When the Project Note Rmd content is converted to html via \code{render()}, these
headers are treated specially - to ensure they are available as sub-headers
under a task when clicked on in the html outline (TODO!)
}
