% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{note_summary_headers_quote_out}
\alias{note_summary_headers_quote_out}
\title{quote out all Markdown Headers}
\usage{
note_summary_headers_quote_out(summary)
}
\description{
Converts every line beginning with \verb{#} in \code{summary} to \verb{>#}
}
\details{
This is used to quote-out the markdown headers in Project Note Summary Section,
so when it is inserted into the Project Doc, the Markdown Headers will not show
in the Project Note Outline in RStudio.  This function is used by \code{update()}
to ensure Project Note summary Markdown Headers are not present in the Doc GDT
summary for each Note.

When the Project Note Rmd content is converted to html via \code{render()}, these
headers are treated specially - to ensure they are available as sub-headers
under a task when clicked on in the html outline (TODO!)
}
