% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_data_type}
\alias{parse_data_type}
\title{Parse data type}
\usage{
parse_data_type(data, expanded_data, indices)
}
\description{
Ensure data added to an existing column (expanded_data) at indices is of the same
type as any existing data.
}
