% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{parse_datatable_add_data}
\alias{parse_datatable_add_data}
\title{ADD DATA to a tibble from a datatable vector}
\usage{
parse_datatable_add_data(dt_vector, dt, table_name)
}
\description{
This datatable has the FUNCTION : ADD_DATA
}
\details{
It can have IDs, variables, or timetable in the first column.

It will have subsequent columns with further data DEPENDING on the first col:

* ID: subsequent cols are variables - can be added directly to dt.


* variables:


* timetable:

  dt_vector <- tables[[i]]
  dt <- datatables[[ table_name ]]
}
