% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_add_data_ids}
\alias{parse_datatable_add_data_ids}
\title{Parse Datatable Add Data IDs}
\usage{
parse_datatable_add_data_ids(
  data,
  headers,
  datatables,
  table_name,
  table_start_line
)
}
\description{
PARSE ADD_DATA datatable with IDs as column names.  The first column header is
set to \code{variables} to indicate this layout.
}
\details{
This function rearranges the extracted \code{data} and \code{headers}, and then parses
this through \code{parse_datatable_add_data_variables()}
}
