% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_add_data_timetable}
\alias{parse_datatable_add_data_timetable}
\title{Parse Datatable Add Data TIMETABLE}
\usage{
parse_datatable_add_data_timetable(
  data,
  headers,
  datatables,
  table_name,
  table_start_line
)
}
\description{
PARSE ADD_DATA datatable in Timetable Format: The first column header is set
to \code{timetable}, which has planned times for condition changes.  Subsequent
columns are \code{group} names that must already exist in the \code{datatables} list
under \code{table_name.}  The final column is \verb{<col_name>_dt}, and datetimes are
inserted here to match conditions declared in the \code{group} names columns.
}
\details{
This function rearranges the extracted \code{data} and \code{headers}, and then parses
this through \code{parse_datatable_add_data_variables()}
}
