% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_add_data_variables}
\alias{parse_datatable_add_data_variables}
\title{Parse ADD_DATA variables layout}
\usage{
parse_datatable_add_data_variables(
  data,
  headers,
  datatables,
  table_name,
  table_start_line
)
}
\description{
SAMPLE-FIRST datatable : IDs in first col, new data in remaining cols
}
