% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_create}
\alias{parse_datatable_create}
\title{CREATE a tibble from a datatable Rmd vector}
\usage{
parse_datatable_create(dt_vector, datatables, table_name, table_start_line)
}
\arguments{
\item{dt_vector}{String vector containing the datatable to parse : should be
of FUNCTION CREATE.}

\item{datatables}{List of existing datatables extracted before the datatable
in \code{dt_vector}.}

\item{table_name}{Name of datatable in \code{dt_vector}.}

\item{table_start_line}{The line in RMD where the table starts - to locate it
if any errors are generated.}
}
\description{
The \code{dt_vector} has the FUNCTION : CREATE
}
\details{
MUST have IDs in the first column.  Can have subsequent columns with
further data.
}
