% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_resample}
\alias{parse_datatable_resample}
\title{RESAMPLE a tibble from a datatable Rmd vector}
\usage{
parse_datatable_resample(dt_vector, datatables, table_name)
}
\description{
This dt_vector has the FUNCTION : CREATE
}
\details{
It MUST have IDs in the first column, then \code{resample} and \code{reps} cols.

It will add these resample and reps cols to the PARENT datatable, THEN will
create NEW DATATABLES - one for each resample element - which are added to
the \code{datatables} list and returned.
}
