% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{parse_datatable_resample_single}
\alias{parse_datatable_resample_single}
\title{Parse a Resample datatable with single SOURCE sample IDs}
\usage{
parse_datatable_resample_single(dt_vector, datatables, table_name)
}
\description{
This will generate the new datatables from a resample dt_vector that contains
single SOURCE sample IDS - it may still generate multiple REPS of subsamples!
}
