% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{parse_deadline_datetime}
\alias{parse_deadline_datetime}
\title{Parse a deadline string (YYYY-MM-DD) => interpret as 00:00 that day}
\usage{
parse_deadline_datetime(deadlineStr)
}
\description{
Parse a deadline string (YYYY-MM-DD) => interpret as 00:00 that day
}
\keyword{internal}
