% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{parse_schedule_datetime}
\alias{parse_schedule_datetime}
\title{Parse a schedule string (YYYY-MM-DD:HHmm) into a POSIXct}
\usage{
parse_schedule_datetime(scheduleStr)
}
\description{
Parse a schedule string (YYYY-MM-DD:HHmm) into a POSIXct
}
\keyword{internal}
