% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectmanagr-package.R
\docType{package}
\name{projectmanagr-package}
\alias{projectmanagr}
\alias{projectmanagr-package}
\title{projectmanagr: Project Mangement Functions for Creating, Editing, Linking and Updating Project Management R Markdown Documents}
\description{
Provides functions for creating new Project Notes linked to a Project Document Goal, Deliverable, Task, as well as new SubNotes under a Compound Project Note, the ability to edit Project Document titles or locations whilst updating the linked documents, and the ability to link new Project Docs and Notes, as well as updating Project documentation summary content.
}
\author{
\strong{Maintainer}: Steven West \email{stevenjonwest@gmail.com} (\href{https://orcid.org/0000-0002-2413-0999}{ORCID})

}
\keyword{internal}
