% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{projectNoteSelection}
\alias{projectNoteSelection}
\title{Document Cursor Selection}
\usage{
projectNoteSelection()
}
\description{
Identify the selection from the Active Cursor position in current R Studio file.
}
\details{
First identifies if the current Cursor position is on a GROUP (HEADER NOTE or SUBNOTE).
It then searches back through the Active Document to find the first line starting with
"##" - if this is a "#### TASK" line, then the first previous "### DELIVERABLE" line and
the first previous "### GOAL" line is also identified.

If TASK, DELIVERABLE and GOAL lines are all successfully found, this method returns
a LIST:

[[1]] or [["line"]] - line index selected in Project Note.

[[2]] or [["column"]] - column index selected in Project Note.

[[3]] or [["projectNotePath"]] - the path of the Active Doc (a Project Doc, as successfully retrieved Task/Del/Goal)

If this method is unsuccessful (there is an error), it returns a LIST:

[[1]] - contains the String "FALSE"
[[2]] - contains the errorMessage - a String indicating why the method failed.
}
