% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{read_data_tables}
\alias{read_data_tables}
\title{Read Data Tables}
\usage{
read_data_tables(rmd_contents, indices)
}
\arguments{
\item{rmd_contents}{Character Vector containing Data Tables to read.}

\item{indices}{Line indices of table delimiters : from \code{datatable_get_delimiter()}}
}
\value{
A named list of tibbles.
}
\description{
Read plaintext data tables into tibbles.
}
