% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{read_file}
\alias{read_file}
\title{Read File}
\usage{
read_file(filePath)
}
\arguments{
\item{filPath}{path to file to read}
}
\description{
Returns a string vector of text file at \code{filePath}.
}
