% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-functions.R
\name{rename_project_doc_deliverable}
\alias{rename_project_doc_deliverable}
\title{Rename a Project Doc Deliverable}
\usage{
rename_project_doc_deliverable(deliverableSelection, newDeliverableName)
}
\arguments{
\item{deliverableSelection}{Selection from projectmanagr Project Doc file,
made on the Deliverable to be renamed.}

\item{newDeliverableName}{The new name to use for the selected Deliverable.}
}
\description{
Rename a Project Doc Deliverable
}
