% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-functions.R
\name{rename_project_doc_task}
\alias{rename_project_doc_task}
\title{Rename a Project Doc Task}
\usage{
rename_project_doc_task(taskSelection, newtaskName)
}
\arguments{
\item{taskSelection}{Selection from projectmanagr Project Doc file, made on
the Task to be renamed.}

\item{newtaskName}{The new name to use for the selected Task}
}
\description{
Rename a Project Doc Task
}
