% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-functions.R
\name{rename_project_note}
\alias{rename_project_note}
\title{Rename a Project Note}
\usage{
rename_project_note(
  projectNotePath,
  newProjectNoteName,
  newProjectNoteTitle = "",
  replaceLinksFileExtensions = list("Rmd")
)
}
\arguments{
\item{projectNotePath}{defines the path to the Project Note - should point
to the Project Note Rmd file}

\item{newProjectNoteName}{defines the NEW Project Note File name}

\item{newProjectNoteTitle}{defines the NEW Project Note Title - written to
its Rmd file.  By default this is the projectNotePath, with spaces replacing
"-" and "_".}
}
\description{
Renames a Project Note at projectNotePath with file name newProjectNoteName,
and replaces title with newProjectNoteTitle if set.
}
\details{
This also modifies all LINKS to this file throughout the Organisation -
updating them to use the new project Note name and title.

NOTE - this DOES NOT modify the PREFIX for a Project Note.
}
