% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectComp.R
\name{renameProjectComp}
\alias{renameProjectComp}
\title{Rename a Project Component}
\usage{
renameProjectComp(
  oldProjectCompPath,
  newProjectCompName,
  newProjectCompTitle = ""
)
}
\arguments{
\item{oldProjectCompPath}{defines the path to the Project Comp - should point to the Project Comp Rmd file}

\item{newProjectCompName}{defines the NEW Project Comp File name}

\item{newProjectCompTitle}{defines the NEW Project Comp Title - written to its Rmd file.  By default this
is the oldProjectCompPath, with spaces replacing "-" and "_".}
}
\description{
Renames a Project Component at oldProjectCompPath with file name newProjectCompName, and newProjectCompTitle.
}
\details{
This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
project Comp name and title.

NOTE - this DOES NOT modify the PREFIX for a Project Comp (set by the PROGRAMME)!
}
