% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectDocGDT.R
\name{renameProjectDocGDT}
\alias{renameProjectDocGDT}
\title{Rename a Project Doc Goal/Del/Task}
\usage{
renameProjectDocGDT(
  oldProjectCompPath,
  newProjectCompName,
  newProjectCompTitle = ""
)
}
\arguments{
\item{oldProjectCompPath}{defines the path to the Project Comp - should point to the Project Comp Rmd file}

\item{newProjectCompName}{defines the NEW Project Comp File name}

\item{newProjectCompTitle}{defines the NEW Project Comp Title - written to its Rmd file.  By default this
is the oldProjectCompPath, with spaces replacing "-" and "_".}
}
\description{
Renames a Project Doc Goal/Del/Task and updates this in all linked Project
Notes.
}
