% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectPrefix.R
\name{renameProjectDocPrefix}
\alias{renameProjectDocPrefix}
\title{Rename a Project Document Prefix}
\usage{
renameProjectDocPrefix(projectDocPath, newProjectDocPrefix, openRmd = TRUE)
}
\arguments{
\item{projectDocPath}{defines the path to the Project Doc - should point to the Project Doc Rmd file}

\item{openRmd}{Whether to open the newlt renamed Rmd in rstudio - TRUE by default.}
}
\description{
Renames a Project Doc Prefix at projectDocPath with new prefix newProjectDocPrefix
}
\details{
This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
project Comp name and title.
}
