% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectPrefix.R
\name{renameProjectNotePrefix}
\alias{renameProjectNotePrefix}
\title{Rename a Project Note Prefix}
\usage{
renameProjectNotePrefix(projectNotePath, newProjectNotePrefix, openRmd = TRUE)
}
\arguments{
\item{projectNotePath}{defines the path to the Project Note - should point to the Project Note Rmd file}

\item{newProjectNotePrefix}{defines the new prefix - this must be consistent with the current project note prefix
and also be available!}

\item{openRmd}{Whether to open the newlt renamed Rmd in rstudio - TRUE by default.}
}
\description{
Renames a Project Note Prefix at projectNotePath with new prefix newProjectNotePrefix.
}
\details{
Project note prefix is either a Simple Project Note (eg. SJW01~001) or a HEADER NOTE (eg. SJW01~002-00).

The renaming here pertains to the Project Note NUMBERING - after the ~ and before any - (eg. SJW01~002-00
to SJW01~003-00).

Renaming of SubNote prefixes inside a HEADER NOTE Group is conducted with renameSubNotePrefix function.

This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
project Note name, and renames any subnotes to new Header Note Prefix if required.
}
