% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectPrefix.R
\name{renameSubNotePrefix}
\alias{renameSubNotePrefix}
\title{Rename a Project Note Prefix}
\usage{
renameSubNotePrefix(subNotePath, newSubNotePrefix, openRmd = TRUE)
}
\arguments{
\item{subNotePath}{defines the path to the Project Note - should point to the Project Note Rmd file}

\item{newSubNotePrefix}{defines the new prefix - this must be consistent with the current project note prefix
and also be available!}

\item{openRmd}{Whether to open the newlt renamed Rmd in rstudio - TRUE by default.}

\item{initialRename}{Set to FALSE if this method is called recursively.  Controls which file is opened when renaming
has finished - only the initial file that is renamed will be opened in rstudio. For internal recursive function calls}
}
\description{
Renames a Project Note Prefix at projectNotePath with new prefix newProjectNotePrefix.
}
\details{
Project note prefix is either a Simple Project Note (eg. SJW01~001) or a HEADER NOTE (eg. SJW01~002-00).

The renaming here pertains to the Project Note NUMBERING - after the ~ and before any -.

Renaming of SubNote prefixes inside a HEADER NOTE Group is conducted with renameSubNotePrefix function.

This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
project Note name, and renames any subnotes to new Header Note Prefix if required.
}
