% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-utils-functions.R
\name{replace_at_indices}
\alias{replace_at_indices}
\title{Replace lines at indices with replacement vector}
\usage{
replace_at_indices(templateContents, indices, replacementVector)
}
\arguments{
\item{templateContents}{Character vector of lines}

\item{indices}{Integer vector of line numbers to replace}

\item{replacementVector}{Character vector to insert in place of each index}
}
\value{
Modified character vector
}
\description{
Replaces each index in \code{indices} with the contents of \code{replacementVector}.
The line at each index is completely replaced. All indices must be between
1 and \code{length(templateContents)} (inclusive).
}
