% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{replace_hyper_links}
\alias{replace_hyper_links}
\title{Replace Hyper Links}
\usage{
replace_hyper_links(
  contentContents,
  sourceFilePath,
  destinationFilePath,
  hyperLinkID = "](",
  hyperLinkEnd = ")"
)
}
\arguments{
\item{contentContents}{String vector of contents that may contain hyperlinks,
identified by presence of \verb{](} in any line.}

\item{sourceFilePath}{The path to the source file - where contentContents was
read from}

\item{destinationFilePath}{The path where the content will be moved to, used
to compute the relative path of all files.}

\item{hyperLinkID}{The identified to search for, set to \verb{](} by default.}

\item{hyperLinkEnd}{The end of the hyper link, set to \verb{)} by default.}
}
\description{
Replace the path inside any markdown hyperlinks with a relative
path to \code{sourceFilePath} from \code{destinationFilePath}.  A hyperlink is identified
with the \verb{)[} string by default, denoting delimiter between the link title and
path.
}
