% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{replace_knitr_include_graphics_link}
\alias{replace_knitr_include_graphics_link}
\title{Replace Knitr include_graphics link}
\usage{
replace_knitr_include_graphics_link(
  contentContents,
  sourceFilePath,
  destinationFilePath,
  knitrGraphics = "knitr::include_graphics(",
  knitrGraphicsEnd = ")"
)
}
\arguments{
\item{contentContents}{String vector of contents that may contain calls to
Knitr::include_graphics()}

\item{sourceFilePath}{The path to the source file - where contentContents was
read from}

\item{destinationFilePath}{The path where the content will be moved to, used
to compute the relative path of all files.}

\item{knitrGraphics}{The function call to knitr grpahics to search for, set
to \verb{knitr::include_graphics(} by default.}

\item{knitrGraphicsEnd}{The end of the function call, set to \verb{)} by default.}
}
\description{
Replace the path inside knitr::include_graphics() functions with a relative
path to sourceFilePath from destinationFilePath.  Deals with different quotes
and whitespace in function call.
}
