% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceAndInsertVector.R
\name{replaceAndInsertVector}
\alias{replaceAndInsertVector}
\title{Replace and insert a vector}
\usage{
replaceAndInsertVector(pattern, values, x)
}
\description{
This function will find the FIRST INSTANCE of the exact match pattern in the passed vector, x,
and replace this element with all elements in the vector, values.
}
