% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{resample_create_datatables_id}
\alias{resample_create_datatables_id}
\title{Create new subsample datatables}
\usage{
resample_create_datatables_id(datatables, table_name, IDs, dataADD)
}
\description{
Using syntax : \code{table_name}-\verb{<parent-rep-num>}_\verb{<subsample-string>}
}
\details{
Tables are also filled with subsample IDs and REPs - expanded by the rep number.
}
