% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{resample_reps_in_col_rep}
\alias{resample_reps_in_col_rep}
\title{Determine whether single or multiple reps across IDs for each Reample String.}
\usage{
resample_reps_in_col_rep(dataADD, REPs)
}
\description{
Take into account the IDs & REPs of existing samples!
}
\details{
Returns result as a named boolean list - each boolean indices whether the
Resample String has more than one rep in AT LEAST one ID.
}
