% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{safe_as_date}
\alias{safe_as_date}
\title{Safely convert a string like "2025-04-01" to Date}
\usage{
safe_as_date(x)
}
\description{
Safely convert a string like "2025-04-01" to Date
}
\keyword{internal}
