% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_add_data}
\alias{sample_add_data}
\title{Add Data to Samples}
\usage{
sample_add_data(
  sampleDataFrame,
  ...,
  IDs = "all",
  rep_select = "all",
  overwrite = FALSE,
  writeToDisk = TRUE
)
}
\arguments{
\item{sampleDataFrame}{a tibble opened with sample_load, imported with sample_import, or created
with sample_create.  The tibble MUST have a 'path' attribute assigned to it, and this MUST point
to the tibbles location relative to the Project Note DIR.  Use sample_load() to load Single tibbles.

#' @param ... one or more vectors to insert into the dataframe. This data must be supplied in the
form 'col-name = data-vector'. The function will fail if the col-name already exists.}

\item{IDs}{Character vector (`c("MS01", "MS02")`) indicating which subset of IDs to add data to.
By default data is added to "all".  The function will fail if any of the selected IDs (and
replicates) have been exported or re-sampled (as they no longer exist!).}

\item{rep_select}{List of numeric vectors (`list(c(1:3), c(2:4)`) to represent the selection
 of replicates from the sampleDataFrame. This vector is used to select SECTIONS from a
 sampleDataFrame which has previously been re-sampled with replicate subsamples (see rep_create
 in sample_resample function).  The list can always be of length 1, in which case the same
 replicates are selected for ALL IDs. Otherwise it must be the same length as ID from the
 sampleDataFrame, in which case each vector in the list selects a unique set of replicates
 from each ID in the sampleDataFrame; or if the IDs parameter has been set, a list the same
 length as IDs can be passed, with each vector in the list selecting a unique set of replicates
 from the set of IDs passed.  The function will fail if any of the selected replicates have been
 re-sampled or exported already. Uses the KEY WORD "all" to select all replicates by default.

 @param overwrite Set to TRUE to overwrite the contents of a previous data column with new input.
 Otherwise, this method will fail when trying to add the same data column(s) again.}

\item{writeToDisk}{Boolean to indicate whether the edits made to sampleDataFrame and the newly formed
DataFrames are saved to disk.  TRUE by default - set to FALSE to test run this function.}
}
\value{
A newly created and saved TIBBLE : saved to the Project Note DIR, showing the new
data associated with the Sample data.
}
\description{
This function receives a sample DataFrame (a Tibble which contains a 'path' attribute - this
is the name of the file in the Project Note's Data Dir), and a series of vectors (...).  After checking
the validity, each vector is inserted into the sample DataFrame and a new copy returned.
}
\details{
- Each data vector must be supplied in the form 'col-name = data-vector'.

By default, all samples in sample DataFrame have data added - the data-vector must be of length 1, or
of the length of all samples.

However, subsets of IDs and replicates can be selected by setting the IDs and rep_select ARGs:

- IDs - can be set to a Character Vector containing the subset of IDs to add data to  Any
  remaining IDs will have their data columns columns left BLANK.

- rep_select - can be set to a List of numeric vectors to represent the selection of
  replicates to add data to.  Each numeric vector in the list represents the SELECTION
  INDICES for each ID (in order of IDs in the Data Frame if all IDs, otherwise using
  the IDs ordering from the IDs ARG above).

To check the results of adding data, its possible to set the writeToDisk boolean to FALSE: in
this case the data additions are made to a copy of the original sampleDataFrame, and it is
returned for inspection.  Once the inputs have been checked, the actual process (with writing
of CSVs to Disk) can be performed with writeToDisk set to TRUE - its default value.

The function is assumed to be run inside an R Chunk in an Rmd file that is a Project Note from the
projectmanagr package.  It assumes the document is saved to disk, it contains a PREFIX, separated from
the Note title with the '~_' separator, and that an adjacent Data DIR exists next to the Note Rmd file,
titled with the PREFIX - this is where the sampleDataFrame is assumed to exist, and where the new Data
Frames will be written.
}
