% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_add_data_by_group}
\alias{sample_add_data_by_group}
\title{Add Data to Samples by Group Values}
\usage{
sample_add_data_by_group(
  sampleDataFrame,
  ...,
  by_group = c("group_col_name", "group_col_value"),
  overwrite = FALSE,
  writeToDisk = TRUE
)
}
\arguments{
\item{sampleDataFrame}{a tibble opened with sample_load or created with sample_create.  The tibble MUST
have a 'path' attribute assigned to it, and this MUST point to the tibbles location relative to the
Project Note DIR.  Use sample_load() to load Single tibbles.

#' @param ... one or more vectors to insert into the dataframe. This data must be supplied in the
form 'name = vector'.}

\item{by_group}{Character vector of length 2 that specifies the group column name followed by
the group character ID/value.

 @param overwrite Set to TRUE to overwrite the contents of a previous data column with new input.
 Otherwise, this method will fail when trying to add the same data column(s) again.}

\item{writeToDisk}{Boolean to indicate whether the edits made to sampleDataFrame and the newly formed
DataFrames are saved to disk.  TRUE by default - set to FALSE to test run this function.}
}
\value{
A newly created and saved TIBBLE : saved to the Project Note DIR, showing the new
data associated with the Sample data.
}
\description{
This function receives a sample DataFrame (a Tibble which contains a 'path' attribute - this
is the name of the file in the Project Note's Data Dir), and a series of vectors (...).  After checking
the validity, each vector is inserted into the sample DataFrame and a new copy returned.
}
\details{
- Each data vector must be supplied in the form 'col-name = data-vector'.

This function supports the addition of data by group:

- `by_group` must be set to a character vector of length TWO: the group column name and group column
value, respectively.

- Data is then added to the columns only where group_col_name's value is group_col_value.

To check the results of adding data, its possible to set the writeToDisk boolean to FALSE: in
this case the data additions are made to a copy of the original sampleDataFrame, and it is
returned for inspection.  Once the inputs have been checked, the actual process (with writing
of CSVs to Disk) can be performed with writeToDisk set to TRUE - its default value.

The function is assumed to be run inside an R Chunk in an Rmd file that is a Project Note from the
projectmanagr package.  It assumes the document is saved to disk, it contains a PREFIX, separated from
the Note title with the '~_' separator, and that an adjacent Data DIR exists next to the Note Rmd file,
titled with the PREFIX - this is where the sampleDataFrame is assumed to exist, and where the new Data
Frames will be written.
}
