% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_add_data_old}
\alias{sample_add_data_old}
\title{Add Sample Data}
\usage{
sample_add_data_old(sampleDataFrame, ...)
}
\arguments{
\item{sampleDataFrame}{a tibble opened with sample_load or created with sample_create.  The tibble MUST
have a 'path' attribute assigned to it, and this MUST point to the tibbles location relative to the
Project Note DIR.  Use sample_load() to load Single tibble for adding data.}

\item{...}{one or more vectors to insert into the dataframe. This data must be supplied in the
form 'name = vector'.}
}
\value{
The newly created and saved TIBBLE with correct 'path' attribute.
}
\description{
This function receives a sampleDataFrame (a special Tibble which contains a 'path' attribute - this
is the name of the file in the Project Note's Data Dir), and a series of vectors (...).  After checking
the validity, each vector is inserted into the sampleDataFrame and a new copy returned.
}
