% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_data_collate}
\alias{sample_data_collate}
\title{Collate Sample Data}
\usage{
sample_data_collate(sampleDataFrame)
}
\description{
This will return a composite dataframe that contains every data entry for a given sample,
extending back through all resampling and exporting, therefore providing access to all
conditions and metadata associated with the samples in a given sample TIBBLE.
}
