% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_export}
\alias{sample_export}
\title{Export Sample Data}
\usage{
sample_export(
  sampleDataFrame,
  export_dest,
  export_dt,
  IDs = "all",
  rep_select = "all",
  writeToDisk = TRUE
)
}
\arguments{
\item{sampleDataFrame}{a tibble opened with sample_load or created with sample_create.  The tibble MUST
have a 'path' attribute assigned to it, and this MUST point to the tibbles location relative to the
Project Note DIR.  Use sample_load() to load Single tibble for exporting}

\item{export_dest}{Character String containing the relative path of the destination Project Note.
This can be the relative path to the Project Note Rmd file, or its DATA_DIR (just the Prefix!)}

\item{export_dt}{a Datetime of the export event, can be a single string (for all IDs), or a vector
with a new datetime for each ID.}

\item{IDs}{Character vector indicating which subset of IDs to apply the export to. By default
the export is applied to "all".  The function will fail if any of the selected IDs (and
replicates) have been exported or re-sampled already.}

\item{rep_select}{List of integer/numeric vectors to represent the selection of replicates from the
sampleDataFrame. This vector is used to select SECTIONS from a sampleDataFrame which has previously
been re-sampled with replicate subsamples (see rep_create in sample_resample function).  The list can
always be of length 1, in which case the same replicates are selected for ALL IDs. Otherwise it must
be the same length as ID from the sampleDataFrame, in which case each vector in the list selects a
unique set of replicates from each ID in the sampleDataFrame; or if the IDs parameter has been set,
a list the same length as IDs can be passed, with each vector in the list selecting a unique set of
replicates from the set of IDs passed.  The function will fail if any of the selected replicates
have been re-sampled or exported already. Uses the KEY WORD "all" to select all replicates by default.}

\item{writeToDisk}{Boolean to indicate whether the edits made to sampleDataFrame and the newly formed
DataFrames are saved to disk.  TRUE by default - set to FALSE to test run the export.}
}
\value{
A newly created and saved list of TIBBLES : the ORIGINAL Tibble containing the export data, and
the NEW TIBBLE saved to the destination Project Note, showing the newly imported Sample data.
}
\description{
This function receives a sampleDataFrame (a NAMED list that contains a Single NAMED Tibble - the name
is the name of the file in the Project Note's Data Dir), an export_dest that contains a character
vector - to identify which Project Note to export to, and an export_dt - a datetime string vector.
}
\details{
- Obtain the export_dest by using the projectmanagr ADDIN: ProjectManagr: Add Hyper Link: This will paste
the relative link to any other open Rmd in RStudio the user selects.  Use this result to set export_dest.

From this input data, the selected IDs from the sampleDataFrame are exported - new columns to indicate
the export are created (export, export_dt), and filled (with the DESTINATION Project Note RELATIVE PATH,
 and the export_dt, respectively).

By default, all samples in sampleDataFrame will be exported. However, subsets of IDs and replicates can
be selected by setting the IDs and rep_select ARGs:

- IDs - can be set to a Character Vector containing the subset of IDs to export.  Any
  remaining IDs will have their export/export_dt columns left BLANK.

- rep_select - can be set to a List of Integer/numeric vectors to represent the selection of replicates
  to apply the export to.  Each vector in the list represents each ID (in order of IDs in the Data
  Frame if all IDs, otherwise using the IDs ordering from the IDs ARG).

When exporting, the export_dest contains the destination Project Note (and its DIR).  This should be the
RELATIVE PATH from the current Project Note to the destination Project Note.  This will be saved in the
export column in the current Sample DataFrame, and the Source Project Note relative path will be computed
and saved to the new sample DataFrame fromed in the Exported Proejct Note's DIR.

The export process will create a NEW DATA FRAME in the destination Project Note DIR:  This dataframe
will be filled with the ID, all previous subsampling and import columns, and a new import column
(import_INDEX) that contains the source Project Note relative path for this import.  For each rep_select
made, the relevant previous data is filled into the new Data Frame.

To check the results of a given export, its possible to set the writeToDisk boolean to FALSE: in this
case the computations are performed, the edits made to a copy of the original sampleDataFrame, and copies
of the old and new DataFrames are all returned for inspection.  Once the inputs are known to be correct,
the actual exporting (with writing of CSVs to Disk) can be performed with writeToDisk set to TRUE - its
default value.

The function is assumed to be run inside an R Chunk in an Rmd file that is a Project Note from the
projectmanagr package.  It assumes the document is saved to disk, it contains a PREFIX, separated from
the Note title with the '~_' separator, and that an adjacent Data DIR exists next to the Note Rmd file,
titled with the PREFIX - this is where the sampleDataFrame is assumed to exist, and where the new Data
Frames will be written.
}
