% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_get_available}
\alias{sample_get_available}
\title{Filter DataFrame to removed Re-Sampled and Exported Samples}
\usage{
sample_get_available(sampleDataFrame)
}
\arguments{
\item{sampleDataFrame}{A named list containing a SINGLE sample dataframe as a tibble.  If more than one
dataframe exists in the list, this method will fail (to guarantee the dataframe being processed is the
intended one).  Use sample_load() to load Single dataframe list for resampling.}
}
\description{
This returns a new copy of sampleDataFrame with all samples that have been resampled or exported
removed.  This is a convenience function for sample_add_data, as data can only be added to a
Sample DataFrame for samples that still exist within it!
}
\details{
Not sure this is necessary for sample_add_data?  As could count number of rows which are blank in both
resample and export columns, if exist, and use this to check against input data vector lengths.  If
they match, can just create a new blank column, and fill rows with blank resample & export with the
correct values.

However, this function may be useful in viewing AVAILABLE SAMPLES that remain in a DataFrame!
}
