% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_get_available_all}
\alias{sample_get_available_all}
\title{get all available samples from dataframes in dirtree}
\usage{
sample_get_available_all(dirtree)
}
\description{
get all available samples from dataframes in dirtree
}
