% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_import}
\alias{sample_import}
\title{Import Samples from SOURCE Project Note to DESTINATION Project Note}
\usage{
sample_import(samples_summary)
}
\description{
Imports all samples listed in the `samples_summary` tibble into the current note from their
source notes.  The `samples_summary` tibble must be a tibble derived from the `sample_find`
function, but with an additional column added, REPS, to indicate how many REPS of each sample
to import.
}
\details{
NOTE: After initial import, the function creates a special samples file - __IMPORT_samples.csv
If this file exists already, this function just returns this file
this ensures the method will run with no errors repeatedly
With all future calls to projectmanagr::sample_ functions, the data will be saved as __samples.csv
To import another set of samples, use 'reimport = TRUE' in sample_import function
}
