% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_import_2}
\alias{sample_import_2}
\title{Import a Sample Dataset}
\usage{
sample_import_2(filename = "IMPORT_samples.csv")
}
\arguments{
\item{filename}{the filename of the dataset to load.  The dataset filename, if exported
from a Project Note in projectmanagr with sample_export() WILL be Prefixed with
`IMPORT_`.  The default value `IMPORT_samples.csv` is likely its name!}
}
\value{
The loaded TIBBLE with the "path" attribute set to imported sample filename:
this is the fielname with the `IMPORT_` prefix removed.
}
\description{
When a Sample is EXPORTED from a Project Note into a New Project Note,
a CSV file is copied into the new Project Note's dir that contains the
base import data.
}
\details{
To maintain the functional paradigm (where no state is stored, and the
sample data files can be writtne to without reference to its state), any
exported data is imported into the new Project Note in a special CSV file
prefixed with `IMPORT_`.

This method will open the `IMPORT_` file, and return the tibble, but the
path attribute is set to the filename OMITTING the `IMPORT_` prefix.

- This is to allow the manipulation of the samples data (adding data etc)
but to also allow all the data processing to be run continuously, by
starting from the raw IMPORT data when running all Rmd code chunks.
}
