% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_load_csv}
\alias{sample_load_csv}
\title{Load a Sample Dataset}
\usage{
sample_load_csv(filename = "__samples.csv")

sample_load_csv(filename = "__samples.csv")
}
\arguments{
\item{filename}{the filename of the dataset to load.}

\item{sampledir}{the path to the Dir containing the samples.}
}
\value{
The loaded TIBBLE with the "path" attribute set to the sample filename.

The loaded TIBBLE with the "path" attribute set to the sample filename.
}
\description{
Load a Sample Dataset from inside a Project Note from ProjectManagr:
The data is automatically loaded from the associated DATA_DIR of the
projectmanagr Project Note - defined as the Directory named as the
PREFIX to the Project Note (the string before the ~_ separator).

Load a Sample Dataset from inside a Project Note from ProjectManagr:
The data is automatically loaded from the associated DATA_DIR of the
projectmanagr Project Note - defined as the Directory named as the
PREFIX to the Project Note (the string before the ~_ separator).
}
\details{
The loaded dataset is a tibble inside a Named List, which ensures
the tibble remains NAMED - all other sample_ functions use this named
tibble to save the data to the correct file when saving data, so it
is important to use this function for loading a dataset.

The loaded dataset is a tibble inside a Named List, which ensures
the tibble remains NAMED - all other sample_ functions use this named
tibble to save the data to the correct file when saving data, so it
is important to use this function for loading a dataset.
}
