% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_load_dir}
\alias{sample_load_dir}
\title{Load Sample Data from DIR}
\usage{
sample_load_dir(sampledir)
}
\description{
Loads sample data from directory `sampledir`.  The loaded datasets will contain all data
for each EXISTING subsample in the DIR.  Separate sub-samples will be returned in separate tibbles,
as different sub-samples may have different data columns.
}
