% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_load_from_summary}
\alias{sample_load_from_summary}
\title{Load Sample Data from Summary Tibble}
\usage{
sample_load_from_summary(samples_summary)
}
\description{
Loads all data for the samples in the Summary Tibble.
}
\details{
This needs completing: sample_load_dir() call bug? Filter rows that match samples_summary?
}
