% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-functions.R
\name{sample_save}
\alias{sample_save}
\title{Save a Sample Dataset}
\usage{
sample_save(sampleDataFrame)
}
\arguments{
\item{sampleDataFrame}{a tibble opened with sample_load or created with sample_create.}
}
\description{
Saves teh sample dataset, using the "path" attribute set to the sampleDataFrame during
sample_load or sample_create.
}
