% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_dir_tree}
\alias{search_dir_tree}
\title{Search Dir Tree}
\usage{
search_dir_tree(path, searchTerm, updateProgress = NULL, settings, orgPath)
}
\description{
Searches all project docs/notes in a Dir Tree within an Organisation,
returning a datatable with FILENAME, CONTEXT, LOCATION LINE columns.
}
