% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{set_proj_note_storage}
\alias{set_proj_note_storage}
\title{Set Storage Vector from Contents}
\usage{
set_proj_note_storage(project_note_contents, project_note_storage)
}
\description{
Set Storage Vector from Contents
}
