% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-addins.R
\name{set_selection_next_horizontal_rule}
\alias{set_selection_next_horizontal_rule}
\title{Set Selection to Next Horizontal Rule}
\usage{
set_selection_next_horizontal_rule()
}
\description{
Set Selection from current line to next line with same number of dashes as
current line.
}
