% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{setSelectionNextHorizontalRule}
\alias{setSelectionNextHorizontalRule}
\title{Set Selection to Next Horizontal Rule}
\usage{
setSelectionNextHorizontalRule()
}
\description{
Set Selection from current line to next line with same number of dashes as
current line.
}
