% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{spaceSummaryBulletPoints}
\alias{spaceSummaryBulletPoints}
\title{Space Summary Bullets}
\usage{
spaceSummaryBulletPoints(summaryBullets)
}
\description{
Summary Bullets - acquired from getProjectNoteDocLinkList() (in returned linkList[[5]]) -
can be re-spaced for insertion into Rmd file.  2 blank lines between primary bullet ("*"), 1
blank line between all other lines, and ends with three blank lines.
}
