% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-read-functions.R
\name{split_multi_obs_data}
\alias{split_multi_obs_data}
\title{Split Multi-Observtional Data}
\usage{
split_multi_obs_data(data)
}
\description{
\code{datatable_extract} returns a list of data, where multi-observational data
is returned in the same list index (column) and vector index (ID), but the
data has been concatenated into one string - space-separated.
}
\details{
This function splits multi-observational data.  Now the FIRST LIST index will
indicate the column, and the SECOND LIST index will indicate the ID.  The
vector index will indicate the multi-observational data for a given column & ID
}
