% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{split_project_notes_by_programme}
\alias{split_project_notes_by_programme}
\title{Split project notes vector by programme}
\usage{
split_project_notes_by_programme(projectNotePaths, settings)
}
\description{
Returns a list where first element of each vector is the programme path, and
subsequent elements are project notes within that programme.
}
\details{
Number of list elements is number of programmes, and number of elements in
vector is number of project notes in that programme -1.
}
