% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{str_in_vec}
\alias{str_in_vec}
\title{returns true is string is identified anywhere in vector.  String must be a
character vecgor of length 1.}
\usage{
str_in_vec(string, vector)
}
\description{
returns true is string is identified anywhere in vector.  String must be a
character vecgor of length 1.
}
