% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{sub_note_link_params}
\alias{sub_note_link_params}
\title{Substitute Note Link in Contents}
\usage{
sub_note_link_params(
  noteLinkContents,
  settings,
  DocGDTList,
  projNoteLinkSummaryContents,
  orgPath
)
}
\description{
Substitute Note Link in Contents
}
