% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{sub_subnote_params}
\alias{sub_subnote_params}
\title{Substitute Sub Note params in contents}
\usage{
sub_subnote_params(
  subNoteContents,
  subNotePrefix,
  subNoteTitle,
  authorValue,
  settings,
  orgPath
)
}
\description{
Substitute Sub Note params in contents
}
