% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-utils-functions.R
\name{sub_template_param}
\alias{sub_template_param}
\title{Substitute a templates' parameter for content}
\usage{
sub_template_param(templateContents, templateParam, paramContents, orgPath)
}
\description{
This function deals with the cases where the paramContents points a multi-lined
vector, or where paramContents points to to a template file in templates/ that
specifies multi-lined content to insert into the templateContents vector.
}
