% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{summarise_id_rep}
\alias{summarise_id_rep}
\title{Convert reps integer vector to R integer string}
\usage{
summarise_id_rep(IDs, REPs, all_reps = FALSE)
}
\description{
Convert reps integer vector to R integer string
}
