% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{summarise_rep_v_str}
\alias{summarise_rep_v_str}
\title{convert a sorted vector of integers to R integer string:}
\usage{
summarise_rep_v_str(REPv)
}
\description{
eg. 1,2,3,5,6,7,9,10,11,12,14 -> 1:3,5:7,9:12,14
}
