% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{todo_data_str}
\alias{todo_data_str}
\title{Empty TODO data frame with the correct columns}
\usage{
todo_data_str(
  file = character(),
  fileMtime = as.POSIXct(character()),
  indexInFile = integer(),
  heading = character(),
  title = character(),
  text = character(),
  deadline = character(),
  schedule = character(),
  isScheduledToday = integer(),
  isSchedulePassed = integer(),
  isDueToday = integer(),
  isDeadlinePassed = integer(),
  itemDateTime = numeric(),
  pTagPriority = integer(),
  groups = character()
)
}
\value{
Zero‑row data frame.
}
\description{
Helper that always returns a data.frame with the  \emph{full} schema the
extractor uses, now including \strong{\code{pTagPriority}} and \strong{\code{groups}}.
}
\keyword{internal}
