% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{trigram_index}
\alias{trigram_index}
\title{Trigram index utilities  (internal)}
\description{
Functions for building, caching and querying a 3‑character inverted index
of an organisation’s R Markdown files.  The main user entry point is
\code{\link{search_dir_tree}}.
}
\keyword{internal}
