% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{trigram_tokenizer}
\alias{trigram_tokenizer}
\title{Tokenise text into lower‑case trigrams}
\usage{
trigram_tokenizer(txt)
}
\arguments{
\item{txt}{Character vector (a document; may contain many lines).}
}
\value{
Character vector of 3‑character n‑grams.
}
\description{
Tokenise text into lower‑case trigrams
}
\keyword{internal}
