% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{trim_deliverable_hash}
\alias{trim_deliverable_hash}
\title{Trim Deliverable Hash}
\usage{
trim_deliverable_hash(del, settings)
}
\arguments{
\item{del}{Deliverable Header String - including markdown header hashes.}

\item{settings}{projectmanagr settings list.}
}
\value{
Deliverable String - without header hashes
}
\description{
Trim default hash from DELIVERABLE, as specified in settings yml:
\code{ProjectDeliverableHeader} & \code{ProjectDeliverableTitle}
}
