% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{trim_goal_hash}
\alias{trim_goal_hash}
\title{Trim Goal Hash}
\usage{
trim_goal_hash(goal, settings)
}
\arguments{
\item{goal}{Goal Header String - including markdown header hashes.}

\item{settings}{projectmanagr settings list.}
}
\value{
Task String - without header hashes
}
\description{
Trim default hash from GOAL, as specified in settings yml:
\code{ProjectGoalHeader} & \code{ProjectGoalTitle}
}
