% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdt‑links-functions.R
\name{trim_task_hash}
\alias{trim_task_hash}
\title{Trim Task Hash}
\usage{
trim_task_hash(task, settings)
}
\arguments{
\item{task}{Task Header String - including markdown header hashes.}

\item{settings}{projectmanagr settings list.}
}
\value{
Task String - without header hashes
}
\description{
Trim default hash from TASK, as specified in settings yml:
\code{ProjectTaskHeader} & \code{ProjectTaskTitle}
}
