% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update}
\alias{update}
\title{Update ProjectManagr Files}
\usage{
update(dirTree = "")
}
\arguments{
\item{dirTree}{OPTIONAL argument to set the directory tree in which an update
will apply.  Will only update the summary information for ORG index, PROG
index, Project Doc files that are contained within dirTree if set.  Default is
a BLANK STRING, which means the current working directory is used, or the
Source Editor context file path is used, to identify the orgPath to update
the whole organisation.}
}
\description{
Updates all summary information in ORG index, PROG index, & Project Docs
with linked PROG index, Project Docs, & Project Notes, respectively.
}
