% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_contents_in_list}
\alias{update_contents_in_list}
\title{Update Insertable Contents in List}
\usage{
update_contents_in_list(contentsStatus, dirPath, orgPath, settings)
}
\description{
Updates the contents found
}
\details{
Searches through dirTree (non-recursively) for Project Notes, and then looks
for Contents in these based on the content sep delimiter (identified in
`config/templates/` & `settings[["ContentSep"]]`).

Returns a structured lists that describe each content in the project notes:

Each Structured List contains: is a character vector comprised
name() : ProjectNotePath && Project Note Line Number where Insertable
Contents BEGINS (start of first delimiter), separated with `:::`
[["contentTitle"]] : Title of Insertable Contents in the Project Note
}
