% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_contents_org_tree}
\alias{update_contents_org_tree}
\title{Update Insertable Contents in Project Notes in ORG Tree}
\usage{
update_contents_org_tree(contentsCache, orgPath, settings)
}
\arguments{
\item{contentsCache}{Named List of contents previous cached - comprising the
contentRetrievalDateTime, and then contents - list of paths to insertable
contents files.  First level in list is the SOURCE DIR - Org Path - shown with
function call \code{names(contentsCache)}.  Second level is
"contentRetrievalDateTime" & "contents" in Org Path - shown with call to
\code{names(contentsCahce[[1]])}, and the actual contents declaration titles are accessed
with a call to \code{names(contentsCache[[1]][['contents']])}}

\item{orgPath}{Organisation path - also used to search for content declatations
within project notes.}

\item{settings}{settings.yml in org config.}
}
\value{
List of content declaration containing named parameters: "contentTitle",
"contentDescription", "contentSource", "contentStartLine", "contentEndLine",
"projectNotePath", "projectNoteContentHeader"
}
\description{
Updates existing cache of insertable contents from Project Notes in ORG.
}
\details{
This is faster than running \code{find_contents_org_tree} as will only read project
notes that have been modified since the last \code{contentRetrievalDateTime}.

Looks through Organisation recursively for Project Notes, and then looks
for Insertable Contents in these based on the content sep delimiter - defined
in \code{CONTENT_SEP.txt}.

Eliminates directories in ORG ROOT which should note be searched: volumes/ &
.config/ & site/ & weekly-journal

Returns structured lists that describe each contents in the project notes:

Each Structured List contains: is a character vector comprised
name() : ProjectNotePath && Project Note Line Number where Insertable
Contents BEGINS (start of first delimiter), separated with \code{:::}

[\link{"contentTitle"}] : Title of Insertable Contents in the Project Note

[\link{"contentDescription"}] : Description of Insertable Contents in the Project Note

[\link{"contentSource"}] : Absolute Path to Contents declared in the Project Note

[\link{"contentStartLine"}] : Start line of the contents declaration in Project Note

[\link{"contentEndLine"}] : End line of the contents declaration in Project Note

[\link{"projectNotePath"}] : Path to Project Note

[\link{"projectNoteContentHeader"}] : Identified markdwon header section that the
Contents Declaration is in.
}
