% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_GDT_link_task}
\alias{update_GDT_link_task}
\title{update GDT links in all project notes with new Task Title that are linked
to a given GDT in Project Doc.}
\usage{
update_GDT_link_task(
  projectDocPath,
  projectDocContents,
  taskSelection,
  newTaskName,
  settings
)
}
\description{
Updating the GDT Link TITLE - which contains the Task String.
}
\details{
Updating the GDT TASK LINK - which is formed from the Task String.
}
