% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_headers}
\alias{update_headers}
\title{Update Headers}
\usage{
update_headers(
  oldHeader,
  newHeader,
  dirTree,
  settings,
  oldHeaderString = "",
  newHeaderString = "",
  fileExtensions = list("Rmd")
)
}
\arguments{
\item{oldHeader}{Old header string to be updated - must start with \verb{#} to
denote a markdown header.}

\item{newHeader}{New header string to replace \code{oldHeader} - must start with
\verb{#} to denote a markdown header.}

\item{dirTree}{Directory tree to search for files for replacing links in.}

\item{settings}{projectmanagr settings list.}

\item{fileExtensions}{List of file extensions indicating what types of files
should be searched for links and have links replaced.  Must be plaintext file
type.  Default is Rmd files only.}
}
\description{
Updates every header in all Project Notes within \code{dirTree}, replacing
\code{oldHeader} with \code{newHeader}.
}
