% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_links}
\alias{update_links}
\title{Update links}
\usage{
update_links(
  oldLinkSuffix,
  newLinkSuffix,
  dirTree,
  settings,
  oldLinkString = "",
  newLinkString = "",
  fileExtensions = list("Rmd")
)
}
\arguments{
\item{oldLinkSuffix}{Old link string to be updated.}

\item{newLinkSuffix}{New link string to replace \code{oldLinkSuffix}.}

\item{dirTree}{Directory tree to search for files for replacing links in.}

\item{settings}{projectmanagr settings list.}

\item{fileExtensions}{List of file extensions indicating what types of files
should be searched for links and have links replaced.  Must be plaintext file
type.  Default is Rmd files only.}
}
\description{
Updates every hyperlink in all Project Notes within \code{dirTree}, replacing
\code{oldLinkSuffix} (the path of the file) with \code{newLinkSuffix} (the new path).
}
