% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_project_note}
\alias{update_project_note}
\title{update project note, putting GDT summary updates for projectDoc that exists at
projectDocPath into projectDocContents.}
\usage{
update_project_note(projectNotePath, projectDocPath, projectDocContents)
}
\description{
update project note, putting GDT summary updates for projectDoc that exists at
projectDocPath into projectDocContents.
}
