% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_protocols_in_list}
\alias{update_protocols_in_list}
\title{Update Protocols in List}
\usage{
update_protocols_in_list(protocolsStatus, dirPath, orgPath, settings)
}
\description{
Updates the protocols found
Searches through dirTree (non-recursively) for Project Notes, and then looks
for Protocols in these based on the protocol sep delimiter (identified in
`config/templates/PROTOCOL_SEP.txt`).
}
\details{
Returns a structured lists that describe each protocol in the project notes:

Each Structured List is comprised
[[3]] : Name of Protocol in the Project Note
[[1]] : ProjectNotePath
[[2]] : Project Note Line Number where Protocol BEGINS (start of first delimiter)
}
