% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectPrefix.R
\name{updateProjectNotePrefix}
\alias{updateProjectNotePrefix}
\title{Update a Project Note Prefix}
\usage{
updateProjectNotePrefix(projectNotePath, newProjectNotePrefix)
}
\arguments{
\item{projectNotePath}{defines the path to the Project Note - should point to the Project Note Rmd file}

\item{newProjectNotePrefix}{defines the new prefix - this must be consistent with the current project note prefix
and also be available!}
}
\description{
Renames a Project Note Prefix at projectNotePath with new prefix newProjectNotePrefix consistent with
Project Doc Prefix.
}
\details{
This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
project Note name, and renames any subnotes to new Header Note Prefix if required.
}
