% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateProjectOrg.R
\name{updateProjectOrg}
\alias{updateProjectOrg}
\title{Update Project Organisation}
\usage{
updateProjectOrg(fileSystemPath = getwd())
}
\arguments{
\item{fileSystemPath}{an absolute path in the filesystem - should be within an Organisation.}
}
\description{
Compares all Project Notes to the Last Update date-time (from config/status.yml).  If a
Project Note has been edited since, reads it summary information for each Project Doc Link.
If there is content, navigates to the Project Doc and compares its content with the Project
Note summary content.  If different, the Project Doc content is updated with the new Project
Note summary content.
}
\details{
The summary information contains an executive summary of progress on the current project
notes work on the Project Doc TASK it is linked to.  It may ALSO contain a TODO section
(by default denoted by the delimiter **TODO**), which this method will cache in `config/todo-cache`
in the orgPath.  This is then retrieved when calling the summariseToDo() method.
}
