% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameProjectPrefix.R
\name{updateSubNotePrefix}
\alias{updateSubNotePrefix}
\title{Update a Sub Note Prefix}
\usage{
updateSubNotePrefix(subNotePath, newSubNotePrefix)
}
\arguments{
\item{subNotePath}{defines the path to the Sub Note - should point to the Sub Note Rmd file}

\item{newSubNotePrefix}{defines the new prefix - this must be consistent with the current sub note prefix
and also be available!}
}
\description{
Renames a Project Sub Note Prefix at subNotePath with new prefix newSubNotePrefix consistent with
Project Header Note Prefix.
}
\details{
This also modifies all LINKS to this file throughout the Organisation - updating them to use the new
sub Note name.
}
