% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{user_selection}
\alias{user_selection}
\title{Document User Selection}
\usage{
user_selection(filePath, line)
}
\description{
Create a selection list from a projectDocPath and line number.
}
\details{
projectDocPath - must be an ABSOLUTE PATH if using the output for addProjectNote() or similar
methods.

line - the line number to generate the selection list from.

This function first identifies if the line in projectDocPath is on a GROUP (HEADER NOTE or SUBNOTE).
It then searches back through the Active Document to find the first line starting with
"##" - if this is a "#### TASK" line, then the first previous "### DELIVERABLE" line and
the first previous "### GOAL" line is also identified.

If TASK, DELIVERABLE and GOAL lines are all successfully found, this method returns
a LIST:

\verb{[[1]]} or \verb{[["task"]]} - contains the line that starts "#### TASK" in the Active Document.

\verb{[[2]]} or \verb{[["taskLine"]]} - the task line number (index) in the Active Document.

\verb{[[3]]} or \verb{[["deliverable"]]} - contains the line that starts "### DELIVERABLE" in the Active Document.

\verb{[[4]]} or \verb{[["deliverableLine"]]} - the deliverable line number (index) in the Active Document.

\verb{[[5]]} or \verb{[["goal"]]} - contains the line that starts "## GOAL" in the Active Document.

\verb{[[6]]} or \verb{[["goalLine"]]} - the goal line number (index) in the Active Document.

\verb{[[7]]} or \verb{[["originalLine"]]} - the original line where the Active Cursor was on in the Active Document.

\verb{[[8]]} or \verb{[["originalLineNumber"]]} - the original line number where the Active Cursor was on in the Active Document.

\verb{[[9]]} or \verb{[["addingSubNote"]]} - a Boolean value indicating the Cursor was on a GROUP (HEADER NOTE or SUBNOTE).

\verb{[[10]]} or \verb{[["headerNoteLink"]]} - if addingSubNote is TRUE, contains the link to the headerSubNote, AS IS GIVEN
IN THE PROJECT DO (i.e. the link TITLE and then the link ADDRESS - for example
\verb{**[LAB~001-00~ GARS SC Clearing Labelling](../LAB/LAB~001-00~_GARS_SC_Clearing_Labelling.Rmd)**}), else is BLANK

\verb{[[11]]} or \verb{[["headerNoteLineNumber"]]} - if addingSubNote is TRUE, contains the line that contains the headerSubNoteLink, else is BLANK.

\verb{[[12]]} or \verb{[["projectDocPath"]]} - the projectDocPath

If this method is unsuccessful (there is an error), it returns a LIST that contains:

\verb{[[1]]} - contains the String "FALSE"
\verb{[[2]]} - contains the errorMessage - a String indicating why the method failed.
}
