% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{volume_remove_dir}
\alias{volume_remove_dir}
\title{Remove directory from Volume and copy back to local filesystem}
\usage{
volume_remove_dir()
}
\description{
Remove directory from Volume and copy back to local filesystem
}
