% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{volume_update_copies}
\alias{volume_update_copies}
\title{Update data copies between volumes}
\usage{
volume_update_copies(project_note_path, from_dir)
}
\description{
This updates any copies of data between volumes so they are all in sync with the latest data - ASSUMED to be
the PRIMARY COPY (i.e. the one which is in or symlinked to the project note dir!)
}
\details{
This function ASSUMES that data copies are IMMUTABLE - data is not edited or deleted, only new data files/paths
are ADDED to a data copy.
}
