% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{volumesAddProgramme}
\alias{volumesAddProgramme}
\title{Volumes Add Programme}
\usage{
volumesAddProgramme(orgPath, programmeName)
}
\description{
Add Programme DIR named programmeName to each data dir in volumes/ in the given orgPath
}
