% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{write_file}
\alias{write_file}
\title{Write File}
\usage{
write_file(contents, filePath)
}
\arguments{
\item{contents}{Character vector to write to filePath}

\item{filPath}{path to file to write}
}
\description{
Writes \code{contents} to \code{filePath}.
}
