% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{write_insertable_contents_cache}
\alias{write_insertable_contents_cache}
\title{Write Insertable Contents Cache}
\usage{
write_insertable_contents_cache(
  contentRetrievalDateTime,
  contents,
  orgPath,
  status,
  statusFile
)
}
\arguments{
\item{contentRetrievalDateTime}{Datetime of retrieval}

\item{contents}{A list indicating the source project note, location, description,
start and end lines of each insertable content file.}

\item{orgPath}{Organisation path.}

\item{status}{The current status file content as list.}

\item{statusFile}{Path to current status file.}
}
\description{
Once retrieved write to status yml file.
}
