% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{write_prog_to_status}
\alias{write_prog_to_status}
\title{Write Programme to Status yml file}
\usage{
write_prog_to_status(programmeTitle, programmeName, status, statusFile)
}
\description{
Fill with datetime of programme creation && programme title
}
