% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{write_yaml_status}
\alias{write_yaml_status}
\title{Write Status YAML to File}
\usage{
write_yaml_status(status, statusFile)
}
\arguments{
\item{status}{List. The status data to be written.}

\item{statusFile}{Character. The path to the status YAML file.}
}
\description{
Saves the status of an organisation as a YAML file.
}
