% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLinkProjectNote.R
\name{writeLinkProjectNote}
\alias{writeLinkProjectNote}
\title{Write a New Link from a Project Note to a Project Doc}
\usage{
writeLinkProjectNote(projectNotePath, selection)
}
\arguments{
\item{projectNotePath}{The path to the Project Note.  This MUST be a sub-directory or lower inside a
PROGRAMME Directory.}

\item{selection}{List containing the Goal, Del, Task selected from the Project Doc, as well as other useful
information - lines of Task/Del/Goal, projectDoc path, content of selection line.  See cursorSelection()
or userSelection().}
}
\description{
This Function adds a new Link from an Existing (SINGLE or SUBNOTE) Project Note to a Project Doc.  The link
is made to a Project Doc GOAL/DEL/TASK, but this GOAL/DEL/TASK is NOT added to the Project Notes' OBJECTIVES.
}
