% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-addins.R
\name{addin_datatable_add_data}
\alias{addin_datatable_add_data}
\title{Add Datatable to existing samples}
\usage{
addin_datatable_add_data()
}
\description{
Generates a Shiny Gadget for adding data to existing samples in a new
datatable.
}
\details{
User selects a datatable name : list of existing datatables
in the current Rmd.

User selects datatable layout : samples, variables, timetable.

User selects the ID format to add : ALL, all-IDs, or a group set.

User defines the data table column names : using the suffix `_dt` denotes
that the column is a DATETIME and it appropriately spaced to hold a datetime.
Otherwise columns are spaced according to the width of the column title.

The datatable with function ADD_DATA is added with appropriate layout.

If the cursor is already INSIDE a TEMPLATE ADD_DATA, which has a valid
ADD_DATA layout (and is already sample, variable or timetable), then
user just needs to select the existing samples datatable to insert data from,
and the IDs to add : ALL, all-IDs, or a group set.

The generic ADD_DATA table will be updated to contain the correct IDs
from the selected existing samples.
}
