% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-addins.R
\name{addin_datatable_create_from_file}
\alias{addin_datatable_create_from_file}
\title{Create Datatable from File}
\usage{
addin_datatable_create_from_file()
}
\description{
Generates a Shiny Gadget for initialising
a new datatable from file.  User selects a csv file from
the file system, and this is converted to datatable format and inserted
in current Rmd at the current line.
}
