% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-functions.R
\name{addin_error_path}
\alias{addin_error_path}
\title{Present Error Message as Addin}
\usage{
addin_error_path(errorTitle, errorMessage, path, settings)
}
\arguments{
\item{errorTitle}{Title for gadgetTitleBar and dialogViewer.}

\item{errorMessage}{Message presented on addin gadget.}
}
\description{
Present Error Message as Addin
}
