% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-addins.R
\name{addin_extract_todos}
\alias{addin_extract_todos}
\title{Extract TODOs Addin}
\usage{
addin_extract_todos()
}
\description{
Currently a non-interactive addin to insert all TODOs from the ORG that
the currently active file is in.
}
