% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-addins.R
\name{addin_insert_todo}
\alias{addin_insert_todo}
\title{Insert Todo Addin}
\usage{
addin_insert_todo()
}
\description{
Inserts a todo block into the current selection in RStudio Source
Editor. Uses format specified in \code{config/templates/todo-block.txt}
}
