% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinDatatableCreate.R
\name{addinDatatableCreate}
\alias{addinDatatableCreate}
\title{Create Datatable from File}
\usage{
addinDatatableCreate()
}
\description{
Generates a Shiny Gadget for initialising
a new datatable from file  User selects a csv file from
the file system, and this is converted to datatable format and inserted
in current Rmd at the current line.
}
