% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProtocol.R
\name{addProtocol}
\alias{addProtocol}
\title{Add a New Protocol}
\usage{
addProtocol(projectNotePath, protocolName, protocolTitle = "",
  protocolTemplate = "Protocol-Template-Tufte.Rmd")
}
\arguments{
\item{projectNotePath}{The ABSOLUTE path of the Project Note.}

\item{protocolName}{The name of the Protocol, a Title with all SPACES replaced
with - or _.
#' @param protocolTitle The title of the Protocol, by default the name with all - and _ replaced
with SPACES.}

\item{protocolTemplate}{Template to use, as found in the `config/templates/` directory.  Default is
"Protocol-Template-Tufte.Rmd"}
}
\description{
This Function adds a new Protocol to a Project Note - saved in the PROGRAMMES' SOP/ Dir.
The Protocol is formed using a Rmd template, and uses the tufte::tufte_handout format to
generate a PDF of the protocol.
}
\details{
Protocols are stored in the SOP/ directory inside the PROGRAMME Directory.  Each Protocol
exists in its own directory, to keep its compiled files together.

The Protocol source Rmd will link to its creating Project Note, and the Project Note will link to the
compiled PDF of the Protocol.
}
