% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeNextHeaderLine}
\alias{computeNextHeaderLine}
\title{Compute next Header line}
\usage{
computeNextHeaderLine(headerContents)
}
\description{
Returns the first blank line after the last content in Header Note contents.
}
