% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{computeNextLine}
\alias{computeNextLine}
\title{Compute Next Line}
\usage{
computeNextLine(line, projDocContents)
}
\description{
Returns the first blank line after the last content under a Task
in a Project Document,
}
