% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNextLinkLine.R
\name{computeNextLinkLine}
\alias{computeNextLinkLine}
\title{Returns the next line in contents, from line, that contains a Rmd link (the next
line that contains the string "](" ), else if no link is found, returns a BLANK
STRING.}
\usage{
computeNextLinkLine(line, contents)
}
\description{
Returns the next line in contents, from line, that contains a Rmd link (the next
line that contains the string "](" ), else if no link is found, returns a BLANK
STRING.
}
