% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNextSubNoteLine.R
\name{computeNextSubNoteLine}
\alias{computeNextSubNoteLine}
\title{Returns the first blank line after the last content under a Header Note Link
in a Project Document,}
\usage{
computeNextSubNoteLine(line, projDocContents)
}
\description{
Returns the first blank line after the last content under a Header Note Link
in a Project Document,
}
