% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePath.R
\name{computePath}
\alias{computePath}
\title{Compute Path from the fullPath combined with the relPath.}
\usage{
computePath(fullPath, relPath)
}
\description{
Relative Path is used to adjust the fullPath, and returns a fullPath
that is directed to the relativePath location.
}
\details{
Example

fullPath <- "/Users/user/00_ORG/01-PROGRAMME/PROJECTS/P01/P01~001~_Project_Doc.Rmd"

relPath <- "../../LAB/LAB01-00~_Exp_01.Rmd"

Output: /Users/user/00_ORG/01-PROGRAMME/LAB/LAB01-00~_Exp_01.Rmd
}
