% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{confirm_dir}
\alias{confirm_dir}
\title{Confirm a Single Directory Exists}
\usage{
confirm_dir(dirPath, errMsg)
}
\arguments{
\item{dirPath}{Character. Path to the directory to check.}

\item{errMsg}{Character. Error message to display if the directory does not
exist.}
}
\description{
Checks if a given directory exists. If not, raises an error.
}
