% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{confirm_file}
\alias{confirm_file}
\title{Confirm a File Exists}
\usage{
confirm_file(filePath, errMsg)
}
\arguments{
\item{filePath}{Character. Path to the file to check.}

\item{errMsg}{Character. Error message to display if the file does not exist.}
}
\description{
Checks if a given file exists. If not, raises an error.
}
