% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_content}
\alias{create_content}
\title{Create Boilerplate Content in a Project Note}
\usage{
create_content(
  selection,
  contentName,
  contentDescription,
  contentTemplate = "Content-Template.Rmd"
)
}
\arguments{
\item{selection}{projectmanagr selection object indicating the type of file
currently selects.  The current file must be a Project Note.}

\item{contentName}{The name of the Protocol.  RECOMMEND to use a VERB-DRIVEN
Naming Convention && use the common verb words first. eg. Fix Perfuse Mouse,
Fix Fog Drosophila.}

\item{contentDescription}{Description of the protocol - shown
to users when choosing a protocol to insert.}

\item{contentTemplate}{Template to use, as found in the `config/templates/`
directory.  Default is "Protocol-Template.Rmd"}
}
\description{
This Function create new boilerplate Content, based on the `contentTemplate`,
which is inserted into selected Project Note.  The base content template is
located in the organisation : `config/templates/Content-Template.Rmd` &
provides suggested layout for a Protocol.
}
\details{
Content is any templated insertable text for boilerplate content.  Typically
in ProjectManage it is used to define a set of Standard Operating
Procedures (SOPs) & LOG Sections that record Protocol execution.  The
Content is defined between specific delimiters, defined in
`CONTENT_SEP.txt`: by default the delimiter is a series of `====`.

Content defined in a source Project Note can be inserted into new Project
Notes, using the `insert_content()` function.
}
