% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_group_note}
\alias{create_group_note}
\title{Add a New Group of Project Notes}
\usage{
create_group_note(
  groupNoteName,
  groupNotePath,
  selection,
  subNoteName,
  addObjToHeader = TRUE,
  groupNoteTitle = "",
  subNoteTitle = "",
  projNoteTemplate = "Project-Header-Note-Template.Rmd",
  subNoteTemplate = "Project-Sub-Note-Template.Rmd",
  headerNoteContentLinkTemplate = "Project-Header-Note-Content-Link-Template.Rmd",
  subNoteContentLinkTemplate = "Project-Sub-Note-Content-Link-Template.Rmd",
  projNoteLinkTemplate = "Project-Note-Link-Template.Rmd",
  projNoteLinkSummaryTemplate = "Project-Note-Link-Summary-Template.Rmd",
  todoTemplate = "Todo-Template.Rmd",
  projNoteSummaryTemplate = "Project-Note-Summary-Template.Rmd",
  subNoteSummaryTemplate = "Project-Sub-Note-Summary-Template.Rmd"
)
}
\arguments{
\item{groupNoteName}{The name of the Header Project Note, a String that should
not contain any SPACES.}

\item{groupNotePath}{The directory where the Header Note will be stored.
This may be a Project Directory, or another Directory specified by the User.
MUST be a sub-directory or lower inside a PROGRAMME Directory.}

\item{selection}{List containing the Goal, Del, Task selected from the Project
Doc, as well as other useful information - lines of Task/Del/Goal, projectDoc
path content of selection line.  See \code{cursor_selection()} or \code{user_selection()}.}

\item{subNoteName}{The First SubNote name, added to the Project Note Group
in the Header Note DIR.}

\item{addObjToHeader}{Boolean to indicate whether the Objective from the
Project Doc is set in the Header Note. True by default.  If False, no Project
Doc Goal/Del/Task is inserted into the Header Note, and the first subnote
is linked in the ProjectDoc G/D/T as a simple note.}

\item{subNoteTitle}{OPTIONAL title for the Project Sub Note.  Default is to
use subNoteName and replace all _ and - with SPACES.}

\item{projNoteTemplate}{Template to use, as found in the \verb{.config/templates/}
directory.  Default is "Project-Header-Note-Template.Rmd"}

\item{subNoteTemplate}{Template to use, as found in the \verb{.config/templates/}
directory.  Default is "Project-Sub-Note-Template.Rmd"}

\item{projNoteLinkTemplate}{Template with structure to add the Project Doc
Goal/Del/Task link in the Project Note.  Includes Project Doc title, link,
and then links to Goal / Del / Task, plus a Summary Info section (filled
with the content from \code{taskSectionHeaderTemplate}).}

\item{projNoteLinkSummaryTemplate}{Template with structure to add underneath the
Project Doc Goal/Del/Task link in the Project Note.  Includes a 'summary' section
and a 'todo' section by default, linked to \code{NoteObjectivesSummarySectionHeader}
& \code{NoteObjectivesTodoSectionHeader} in \code{.config/settings.yml}}

\item{projNoteSummaryTemplate}{Template with structure to add Project Note
summary to Project Doc under Goal/Del/Task.}

\item{subNoteSummaryTemplate}{Template with structure to add Project Sub Note
summary to Project Doc under Goal/Del/Task.}

\item{projectNoteTitle}{OPTIONAL title for the Project HEADER Note.  Default
is to use groupNoteName and replaceall _ and - with SPACES.}
}
\description{
This Function adds a Project Note Group - consisting of one HEADER
Note, and one SUBNOTE inside the Header Note Dir, named with a computed
\code{groupNotePrefix}  including identifier and Major Numbering, separated by
\code{GroupNotePrefixSep} & \code{HeaderNotePrefix} for header note or first subNote
index for subNote, as specified in \code{settings.yml}.
}
\details{
This Note Group can be expanded by adding further SubNotes - useful for
Optimisation and Experimental Repeats - see \code{create_sub_note()}.
}
