% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{create_hyperlink}
\alias{create_hyperlink}
\title{Create Hyperlink}
\usage{
create_hyperlink(toFileName, toFilePath, fromFilePath)
}
\description{
creates string for hyperlink in Rmd, using `toFileName` as the hyperlink text,
and generating a RELATIVE LINK to `toFilePath` from `fromFilePath`.
}
