% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{create_hyperlink_no_ext}
\alias{create_hyperlink_no_ext}
\title{generate a hyperlink between two files, using \code{toFilePath} name without
extension as the link text.}
\usage{
create_hyperlink_no_ext(toFilePath, fromFilePath)
}
\description{
generate a hyperlink between two files, using \code{toFilePath} name without
extension as the link text.
}
