% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_programme}
\alias{create_programme}
\title{Create a New Programme}
\usage{
create_programme(
  programmeName,
  organisationPath,
  programmeTitle = "",
  progTemplate = "Programme-Template.Rmd",
  progSummaryTemplate = "Programme-Summary-Template.Rmd",
  ctime = ""
)
}
\arguments{
\item{programmeName}{Name of Progamme - must NOT contain a space.}

\item{organisationPath}{Path to insert the PROGRAMME into.  If this is not an
Organisation Directory, will search up the directory tree to attempt to find
one.  If none found, the method will end without making a PROGRAMME.}

\item{programmeTitle}{Title of programme - typically the programmeName with
"_" & "-" replaced with spaces.}

\item{progTemplate}{The Rmd file to use as a template to create the Programme.
Default is "Programme-Template.Rmd" from projectmanagr templates.}

\item{progSummaryTemplate}{The Rmd file to use as a template to create the
Programme Summary. Default is "Programme-Summary-Template.Rmd" from
projectmanagr templates.}

\item{ctime}{Add creation time to write to status.yml - for automated testing.}
}
\value{
progIndexPath The path to the programme index Rmd file.
}
\description{
Generates a new Programme in an Organisation.  Generates a directory for the
programme, and an initial \verb{_index_} .Rmd file that defines the programme
homepage.
}
\details{
If \code{organisationPath} is not at the top of the Organisation, will traverse until
it is found.  This function will halt if no Organisation is found - the
organisation root is defined by the presence of .config/ and .config/templates
dirs at root. An optional programmeTitle (which if not supplied will default
to the programmeName, replacing "_" & "-" with " ").
}
