% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_project_doc}
\alias{create_project_doc}
\title{Create a New Project Document}
\usage{
create_project_doc(
  projectPrefix,
  projectName,
  programmePath,
  projectTitle = "",
  projDocTemplate = "Project-Doc-Template.Rmd",
  projDocSummaryTemplate = "Project-Doc-Summary-Template.Rmd",
  projectIndex = 0
)
}
\arguments{
\item{projectPrefix}{Prefix of Project - must contain only alphanumeric chars & be unique to Programme.}

\item{projectName}{Name of Project - must NOT contain a space.}

\item{programmePath}{Path to insert the Project into.  This must be a Programme dir, one level below the
organisation, and containing a PROJECTS/ directory.  The Project will be placed into the PROJECTS/ directory.
If none found, the method will end without making a Project}

\item{projectTitle}{Title of project - by default set to `projectName` with "_" & "-" replaced with spaces.}

\item{projDocTemplate}{Rmd template used to create the Project Doc - default is "Project-Doc-Template.Rmd"}

\item{projectIndex}{If 0, the function will determine the projectIndex by searching the PROJECTS/ directory.
Otherwise, the projectIndex is used to number the Project in its Prefix.}
}
\description{
Generates a new Project inside a Programme in an Organisation.  The programmePath
must be in the Programme directory, which is the sub-Dir to the root ORGANISATION
dir.
}
\details{
The project is automatically numbered, by reading the current projects and determining
the next number in the sequence, as well as deriving the Programme Prefix.

User must supply the project name, which must contain NO SPACES, and optionally a project
title.  If no title is provided, the Title is derived from the project name, replacing any
"_" & "-" with " ". The default programmePath is the working directory.
}
