% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{create_protocol}
\alias{create_protocol}
\title{Create a Protocol in a Project Note}
\usage{
create_protocol(
  selection,
  protocolName,
  protocolDescription,
  protocolTemplate = "Protocol-Template.Rmd"
)
}
\arguments{
\item{selection}{projectmanagr selection object indicating the type of file
currently selects.  The current file must be a Project Note.}

\item{protocolName}{The name of the Protocol.  RECOMMEND to use a VERB-DRIVEN
Naming Convention && use the common verb words first. eg. Fix Perfuse Mouse,
Fix Fog Drosophila.}

\item{protocolDescription}{Description of the protocol - shown
to users when choosing a protocol to insert.}

\item{protocolTemplate}{Template to use, as found in the `config/templates/`
directory.  Default is "Protocol-Template.Rmd"}
}
\description{
This Function create a new Protocol, based on the `protocolTemplate`, which
is inserted into a Project Note.  The base protocol template is located in
the organisation : `config/templates/Protocol-Template.Rmd` & provides
suggested layout for a Protocol.
}
\details{
Protocols are insertable text that define a set of Standard Operating
Procedures (SOPs) & LOG Sections that record Protocol execution.  The
Protocol is defined between specific delimiters, defined in
`PROTOCOL_SEP.txt`: by default the delimiter is a series of `====`.

Protocols defined in a source Project Note can be inserted into new Project
Notes, using the `insert_protocol()` function.
}
