% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectOrg.R
\name{createProjectOrg}
\alias{createProjectOrg}
\title{Create a New Project Organisation}
\usage{
createProjectOrg(
  orgName,
  orgTitle,
  fileSystemPath = getwd(),
  orgTemplate = "Org-Template.Rmd"
)
}
\arguments{
\item{orgName}{defines the directory name of the Organisation. Default is "00_ORG",}

\item{orgTitle}{defines the title of the Organisation, in its document.  Default is "ORGANISATION",}

\item{fileSystemPath}{defines where the organisation directory is written, Defaultis the working directory}

\item{orgTemplate}{defines the template to use from the projectmanagr package.  Default is "Org-Template.Rmd".}
}
\description{
Generates the layout for a new Project Organisation in the File System.
}
\details{
The function generates the organisation directory and its initial contents:  config/ directory with the
config files and templates/ directory containing component templates.  The volumes/ directory, that will
hold all volumes for managing data storage. The docs/ directory, that will hold documentation for a
projectmanagr usage.   A 00_site/ directory is created next to org, which will contain the generated html
site.
}
