% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectOrg.R
\name{createProjectOrg}
\alias{createProjectOrg}
\title{Create a New Project Organisation}
\usage{
createProjectOrg(authorValue, orgName = "00_ORG",
  orgTitle = "00 ORGANISATION", fileSystemPath = getwd(),
  orgTemplate = "Org-Template.Rmd")
}
\arguments{
\item{authorValue}{name of author of this organisation - written into the yaml header of each Rmd file.}

\item{orgName}{defines the directory name of the Organisation. Default is "00_ORG",}

\item{orgTitle}{defines the title of the Organisation, in its document.  Default is "ORGANISATION",}

\item{fileSystemPath}{defines where the organisation directory is written, Defaultis the working directory}

\item{orgTemplate}{defines the template to use from the projectmanagr package.  Default is "Org-Template.Rmd".}
}
\description{
Generates the layout for a new Project Organisation in the File System.
}
\details{
The function generates the organisation directory and its initial contents:  config/ direcotry with the
config files and templates/ directory containing component templates.  The site/ directory, which will
contain the generated html site.  The volumes/ directory, with an initial local/ directory, that will
hold all volumes for managing data storage.
}
