% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_add_data_samples_rmd}
\alias{datatable_add_data_samples_rmd}
\title{Create new Sample Data Table and insert into Rmd}
\usage{
datatable_add_data_samples_rmd(
  rmd_path,
  rmd_line,
  data_cols,
  datatable_name,
  ids_vector = "",
  dt_length = 120,
  summarise_reps = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{line in Rmd file to insert table
#'}

\item{data_cols}{Character vector of data table column titles}

\item{datatable_name}{String of data table name - MUST exist in rmd_path!}

\item{ids_vector}{A vector containing the IDs to add: Can be "ALL", a set of GROUP
IDs (which must already be declared in datatable name under a 'group-' data
column), or (a subset of) IDs from datatable.}

\item{dt_length}{Int of data table max length in characters - default 120.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}
}
\description{
Creates a new Sample DataTable in specified Rmd file at specified line.
The Sample DataTable will contain an initial ID column containing the
IDs vector, and an optional set of extra data columns as specified in the
data_cols vector.  If the table exceeds dt_length characters (default 120),
then the table is split into multiple tables, with IDs as first col, and
subsequent data_cols given in subsequent tables.
}
