% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{datatable_add_data_variables}
\alias{datatable_add_data_variables}
\title{Add a variable-first data table}
\usage{
datatable_add_data_variables(
  contents,
  var_names,
  datatable_name,
  group_names,
  dt_length = 100
)
}
\arguments{
\item{contents}{character vector containing current document with
existing data tables.}

\item{var_names}{Names of variables to be added to first column.}

\item{datatable_name}{The EXISTING datatable from which the IDs must be drawn.}

\item{group_names}{Vector of sample IDs or GROUP NAMES from a group set, which will
constitute the remaining column headers.  These will typically be group
names.}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
for procedure data
}
\details{
First column contains a list of variables, subsequent column titles are
typically group IDs, and data values are in cells.
}
