% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_add_data_variables_rmd}
\alias{datatable_add_data_variables_rmd}
\title{Create new Sample Data Table and insert into Rmd}
\usage{
datatable_add_data_variables_rmd(
  rmd_path,
  rmd_line,
  var_names,
  datatable_name,
  group_names,
  settings,
  dt_length = 100
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{line in Rmd file to insert table
#'}

\item{var_names}{Names of variables to be added to first column.}

\item{datatable_name}{The EXISTING datatable from which the IDs must be drawn.}

\item{group_names}{Vector of sample IDs or GROUP NAMES from a group set, which will
constitute the remaining column headers.  These will typically be group
names.}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
Creates a new Sample DataTable in specified Rmd file at specified line.
The Sample DataTable will contain an initial ID column containing the
IDs vector, and an optional set of extra data columns as specified in the
data_cols vector.  If the table exceeds dt_length characters (default 120),
then the table is split into multiple tables, with IDs as first col, and
subsequent data_cols given in subsequent tables.
}
