% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_add_group}
\alias{datatable_add_group}
\title{Add Groups to Data Table}
\usage{
datatable_add_group(
  contents,
  group_names,
  datatable_name,
  groups = "",
  dt_length = 120
)
}
\arguments{
\item{contents}{character vector containing current document with
existing data tables.}

\item{group_names}{Character vector of data table column titles to add.
CANNOT BE BLANK!  MUST start all group names with `group-`.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.}

\item{groups}{A list of vectors containing the group labels for each group
defined in group_names}

\item{dt_length}{Int of data table max length in characters - default 120.}
}
\description{
Adds Groups to an EXISTING Data Table in the character vector `contents`.
The Groups DataTable will contain an initial ID column containing all
sample IDs, and a series of extra data columns named using the
group_names vector.
}
\details{
If the table exceeds dt_length characters
(default 120), then the table is split into multiple tables, with IDs as
first col, and subsequent group_names given in subsequent tables.
}
