% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_add_group_rmd}
\alias{datatable_add_group_rmd}
\title{Add Groups to Data Table and insert into Rmd}
\usage{
datatable_add_group_rmd(
  rmd_path,
  rmd_line,
  group_names,
  datatable_name,
  groups,
  dt_length = 120,
  summarise_reps = FALSE
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{line in Rmd file to insert table}

\item{group_names}{Character vector of data table column titles to add.
CANNOT BE BLANK!}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.}

\item{groups}{A list of vectors containing the group labels for each group
defined in group_names.}

\item{dt_length}{Int of data table max length in characters - default 120.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the datatable.  If FALSE each ID/rep is on a separate line in the new datatable,
otherwise if TRUE, all reps are summarised using r vector index syntax on
one line in the new datatable.  i.e. each ID is listed ONCE and the reps are
indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}
}
\description{
Adds Groups to an EXISTING Data Table in specified Rmd file at specified line.
The Groups DataTable will contain an initial ID column containing all
sample IDs, and a series of extra data columns named using the
group_names vector.
}
\details{
If the table exceeds dt_length characters
(default 120), then the table is split into multiple tables, with IDs as
first col, and subsequent group_names given in subsequent tables.
}
