% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_create}
\alias{datatable_create}
\title{Create new Sample Data Table}
\usage{
datatable_create(
  IDs = "",
  data_cols = "",
  datatable_name = "samples",
  default_data_vals = list(),
  dt_length = 100
)
}
\arguments{
\item{IDs}{Character vector of sample IDs}

\item{data_cols}{Character vector of data table column titles}

\item{datatable_name}{String of data table name - default "samples".}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
Creates a new Sample DataTable vector. The Sample DataTable will contain an
initial ID column containing the IDs vector, and an optional set of extra
data columns as specified in the data_cols vector.  If the table exceeds
`dt_length` characters (default 120), then the table is split into multiple
tables, with IDs as first col, and subsequent data_cols given in subsequent
tables.
}
