% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{datatable_create}
\alias{datatable_create}
\title{Create new Plaintext Data Table}
\usage{
datatable_create(
  IDs = "",
  data_cols = "",
  datatable_name = "samples",
  default_data_vals = list(),
  dt_length = 100
)
}
\arguments{
\item{IDs}{Character vector of sample IDs}

\item{data_cols}{Character vector of data table column titles}

\item{datatable_name}{String of data table name - default "samples".}

\item{dt_length}{Int of data table max length in characters - default 100.}
}
\description{
Creates a new Plaintext DataTable vector. The DataTable will contain an
initial ID column containing the \code{IDs} vector, and an optional set of extra
data columns as specified in the \code{data_cols} vector.  If the table exceeds
\code{dt_length} characters (default 100), then the table is split into multiple
tables, with IDs as first col, and subsequent data_cols given in subsequent
tables.
}
