% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_create_rmd}
\alias{datatable_create_rmd}
\title{Create new Sample Data Table and insert into Rmd}
\usage{
datatable_create_rmd(
  rmd_path,
  rmd_line,
  IDs = "",
  data_cols = "",
  datatable_name = "samples",
  dt_length = 120
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{line in Rmd file to insert table}

\item{IDs}{Character vector of sample IDs}

\item{data_cols}{Character vector of data table column titles}

\item{datatable_name}{String of data table name - default "samples".}

\item{dt_length}{Int of data table max length in characters - default 120.}
}
\description{
Creates a new Sample DataTable in specified Rmd file at specified line.
The Sample DataTable will contain an initial ID column containing the
IDs vector, and an optional set of extra data columns as specified in the
data_cols vector.  If the table exceeds dt_length characters (default 120),
then the table is split into multiple tables, with IDs as first col, and
subsequent data_cols given in subsequent tables.
}
