% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-write-template-functions.R
\name{datatable_create_template_rmd}
\alias{datatable_create_template_rmd}
\title{Create new Sample DataTable from TEMPLATE and insert into Rmd}
\usage{
datatable_create_template_rmd(
  rmd_path,
  rmd_line,
  datatable_name,
  template_datatable_name,
  IDs = NULL,
  all_ids = FALSE,
  dt_length = 100
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_line}{Selected line in Rmd file within TEMPLATE datatable.}

\item{datatable_name}{Name of an EXISTING datatable to source IDs from: these
will replace the \verb{<<IDS>>} String in the TEMPLATE datatable with all IDs.}

\item{template_datatable_name}{Name the template dtaatable will take when created.}

\item{all_ids}{Boolean to indicate whether all IDs from \code{source_daatable_name}
should be returned: ignoring any that are DISPOSED EXPORTED or RESAMPLED.
FALSE by default - only returning IDs that are NOT DISPOSED EXPORTED or RESAMPLED.}

\item{dt_length}{Int of data table max length in characters - default 100.

Missing param \code{expand} not required - never want to expand the values in the
template datatable, as purpose of template is to specify all values from one
line indicated by \verb{<<IDS>>} variable!
Variable \code{expand} : Whether to expand any values in vectors set to \code{default_data_vals},
default is FALSE, meaning custom values can be set in \code{default_data_vals} for
each ID. If TRUE, teh vector is expanded by the length of IDs, meaning ALL
values in each vector are set to EACH ID, added as a multi-obs set of rows
for each ID.}
}
\description{
Creates a Sample DataTable in specified Rmd file from the template that
should be present between specified start & end line.
}
\details{
\enumerate{
\item The template Sample DataTable will contain an initial ID column containing
the default \verb{<<IDS>>} vector - which will be replaced with IDs from selected
existing datatable.  Can use this syntax to create a new Datatable using IDS
defined in existing table, with different PREFIX or SUFFIX strings.
\item All other columns will have their initial data value replicated along the
length of the datatable - length of IDs.
}
}
