% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_create_template_rmd}
\alias{datatable_create_template_rmd}
\title{Create new Sample Data Table and insert into Rmd}
\usage{
datatable_create_template_rmd(
  rmd_path,
  rmd_startline,
  rmd_endline,
  IDs = "",
  datatable_name = "samples",
  dt_length = 120
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_startline}{line in Rmd file where TEMPLATE datatable starts}

\item{rmd_endline}{line in Rmd file where TEMPLATE datatable ends}

\item{IDs}{Character vector of sample IDs}

\item{datatable_name}{String of data table name - default "samples".}

\item{dt_length}{Int of data table max length in characters - default 120.}
}
\description{
Creates a Sample DataTable in specified Rmd file from the template that
should be present between specified start & end line.
}
\details{
The template Sample DataTable will contain an initial ID column containing
the IDs vector, and any other columns will have their initial data value
replicated along the length of the datatable.
}
