% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-gen-functions.R
\name{datatable_export}
\alias{datatable_export}
\title{Generate EXPORT Datatable}
\usage{
datatable_export(
  contents,
  datatable_name,
  destination_rel_path_link,
  ids_vector = "",
  reps_vector = "",
  dt_length = 100,
  summarise_reps = FALSE
)
}
\arguments{
\item{contents}{character vector containing current document with
existing data tables.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.}

\item{destination_rel_path_link}{Markdown link containing the relative path
to the Destination Note where these samples are EXPORTED to.}

\item{ids_vector}{A vector containing the IDs to add: Must be a set of EXISTING
IDs from the selected \code{datatable_name} - cannot be ALL or a set of Group IDs.
Default blank string "" indicates to add data to ALL EXISTING IDs.  Existing
IDs are all IDs that have not been resampled, exported, or disposed.}

\item{reps_vector}{A vector containing the REPs to add: This is combined with
the \code{ids_vector} to specify ID-REP pairs.  Each ID-REP pair must exist (not
be resampled, exported, or disposed) from the selected \code{datatable_name}.
Default blank string "" indicates to add data to ALL EXISTING ID-REPs.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{summarise_reps}{Boolean to indicate whether reps should be summarised in
the generated datatable.  If FALSE each ID/rep is on a separate line in the
new datatable, otherwise if TRUE, all reps are summarised using r vector index
syntax on one line in the new datatable.  i.e. each ID is listed ONCE and the
reps are indicated as: 1:3,5,6:10,12,14:25 etc.  Default to FALSE.}
}
\value{
List of datatables generated.
}
\description{
Exports all samples IDs from \code{datatable_name} in \code{contents} to
\code{destination_rel_path_link}.
}
\details{
The generated EXPORT DataTable will contain an initial ID column containing
all EXISTING IDs (IDs that have been resampled or exported will automatically
be EXCLUDED), and an \code{export} column that contains the \code{destination_rel_path}.

Explicit IDs or ID-REPs (if reps exist in selected datatable) can be set via
the \code{ids_vector} and \code{reps_vector} arguments.  If blank, all eisting IDs (&
REPs) will be exported.  This table will add the rep column in the format
indicated by \code{summarise_reps}.

If the table exceeds \code{dt_length} characters (default 100), then the table is
split into multiple tables, with IDs as first col, and subsequent data_cols
given in subsequent tables.
}
