% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_export}
\alias{datatable_export}
\title{Export a set of samples from summary datatable}
\usage{
datatable_export()
}
\description{
This function will write an EXPORT table to the end of
}
