% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_find}
\alias{datatable_find}
\title{Find EXISTING Samples in Datatables}
\usage{
datatable_find(path, datatable_name_prefix = "samples")
}
\value{
a tibble containing summary of all EXISTING samples
}
\description{
Search through path recursively to find all Rmd files that contain datatables.
Summarise all EXISTING samples that are declared in them (not resampled,
exported, disposed).
}
\details{
Summary tibble includes following columns:

* ID - sample ID

* SAMPLE: Composite of all subsampling - what the sample is

* COUNT: Number of reps of the sample available (sections)

* EXP: The experiment PREFIX ID and EXPERIMENT TITLE - title can help identify
where the tissue comes from/its properties - as using functional names for
experiments eg. `NRF~005-001~_SLEIGH_CNS_GARS_01_SC-LUM` know tissue is from
James Sleigh, and its GARS mutant or WT control.

* PATH: The absolute path to the Experiment DIR ??

* LOCATION : values of `_location` col IF LAST COL DECLARED - where is the
sample currently?  Only defined last if currently in storage!

* CONDITION : value of last `_condition` col - what is sample in?

If further metadata is needed - can use `datatable_get_sample_data_history`
and filter for relevant information.
}
