% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_import}
\alias{datatable_import}
\title{Import a set of samples and reps from a summary datatable}
\usage{
datatable_import()
}
\description{
Import a set of samples and reps from a summary datatable
}
