% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_read_rmd}
\alias{datatable_read_rmd}
\title{Read datatables in Rmd}
\usage{
datatable_read_rmd(rmd_path)
}
\arguments{
\item{rmd_path}{Path to RMarkdown Document.}
}
\description{
Function to read all projectmanagr datatables declared in a plain text
document (typically an R Markdown doc), and return them as tibbles. Extracts
all datatables between \verb{+===} markers, and checks the validity of each
datatable declared.
}
\details{
Each datatable is named, and each contains samples which must all possess a
unique ID.  Datatables are created in CREATE, and have data added to them
in ADD_DATA tables.  IDs from a given datatable can be put into new groups
using a GROUP table.

Adding data can use various layouts:
\itemize{
\item Sample-first layout:  Typically used for measurements made on each sample
individually - sample IDs are in first col, subsequent cols contain the
measurement data.
\item Variable-first layout: Typically used to add datetimes of procedures -
procedure titles are laid out in first col, and subsequent cols are GORUPS
which are processed together.  This allows timing data to be added cleanly
and easily across all samples in a group (including the special group ALL).
\item TIMETABLE: Used for measuring the actual timings used during optimisation
of a protocol.  When time is varied in a protocol it is very difficult to
plan and track this.  The timetable provides a convenient layout for planning
the groups and the timings of changes over the procedure where timing is being
optimised.  It then allows the ACTUAL TIMINGS to be inserted into the table
as the procedures are followed - and these are linked to the original samples,
making this data available in further analyses.
}

Samples in datatables can be subsampled - destroying the original sample,
and creating one or more sub-samples.  The parent samples can no longer
have data added to them, only the existing sub-samples.

Samples in datatables are be exported to destination notes, and imported
from source notes.  If exported from this source note, the samples no longer
exist in the current note - they cannot be manipulated in this note.

If imported from a previous source note, the samples now exist in this note.
They can have data added to them, grouped, subsampled, and exported.
}
