% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_read_vector}
\alias{datatable_read_vector}
\title{Read Datatables and check validity from character vector}
\usage{
datatable_read_vector(rmd_contents)
}
\description{
Read Datatables and check validity from character vector
}
