% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_resample}
\alias{datatable_resample}
\title{Resample a set of samples}
\usage{
datatable_resample(
  contents,
  datatable_name,
  resample_vector,
  rep_vector = c(1),
  dt_length = 120
)
}
\arguments{
\item{contents}{character vector containing current document with
existing data tables.}

\item{datatable_name}{String of data table name: This data table MUST ALREADY
EXIST/BE DECLARED BEFORE rmd_line in rmd_path.  The resampling will be
default include all sample IDs from this datatable.  The user may then
delete and sample IDs as is appropriate.}

\item{resample_vector}{A character vector that contains all the resample
codes to be added to each sample/rep in the resample table.}

\item{dt_length}{Int of data table max length in characters - default 120.}
}
\description{
Creates a resample datatable consisting of: sample ID column containing all
EXISTING sample IDs, sample reps column (if applicable) containing all EXISTING
sample reps, and then the resample and rep columns, which represent the
resampling.
}
\details{
The function assumes all EXISTING samples and reps from selected datatable are
resampled: ALL are added to the output resample datatable.

The User is free to DELETE sample IDs and reps from the resample table, and
the deleted samples will no longer be resampled when read into tibbles.

This functionality means the user can resample an INITIAL SUBSET of samples
in an experiment (perhaps as a pilot - for example resampling a block of CNS at
different thicknesses 100um 200um 400um 500um etc), and then later to resample
the remaining samples in the experiment (eg. now the pilot shows 200um is the
best thickness, all other samples are resampled to 200um thick sections).

These two resampling events can be documented in two separate resample tables,
and these will be collated into a single resample/rep columns in the final
compiled datatable tibble.
}
