% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{datatable_resample_template_rmd}
\alias{datatable_resample_template_rmd}
\title{Resample Sample from TEMPLATE and insert into Rmd}
\usage{
datatable_resample_template_rmd(
  rmd_path,
  rmd_startline,
  rmd_endline,
  datatable_name,
  settings,
  dt_length = 100
)
}
\arguments{
\item{rmd_path}{path to Rmd file}

\item{rmd_startline}{line in Rmd file where TEMPLATE datatable starts}

\item{rmd_endline}{line in Rmd file where TEMPLATE datatable ends}

\item{datatable_name}{String of data table name.}

\item{dt_length}{Int of data table max length in characters - default 100.}

\item{IDs}{Character vector of sample IDs}
}
\description{
Creates a Sample DataTable in specified Rmd file from the template that
should be present between specified start & end line.
}
\details{
The template Sample DataTable will contain an initial ID column containing
a default IDs vector (which will be replaced with IDs from selected existing
datatable), and any other columns will have their initial data
value replicated along the length of the datatable.
}
