% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{default_google_auth}
\alias{default_google_auth}
\title{Authenticate with Google Calendar using projectmanagr's built-in OAuth client}
\usage{
default_google_auth(
  gar_set_client = googleAuthR::gar_set_client,
  gar_auth = googleAuthR::gar_auth
)
}
\value{
Invisibly returns the result of \code{googleAuthR::gar_auth()}
}
\description{
This function configures \code{googleAuthR} using a pre-bundled OAuth 2.0 client
JSON, and sets the token cache to a user-specific location using \code{rappdirs}.
}
\details{
It then attempts to authenticate the user by selecting an existing cached
token automatically (using \code{googleAuthR::gar_auth(email = TRUE)}), or
prompting browser sign-in if no token exists yet.

This function sets a global R option:
\itemize{
\item \code{options(gargle_oauth_cache = rappdirs::user_cache_dir("projectmanagr"))}
}
to store refreshable tokens in a consistent location across sessions.
}
