% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-functions.R
\name{extract_todos}
\alias{extract_todos}
\title{Extract All TODO lists from a Collection of Project Notes}
\usage{
extract_todos(
  location,
  date = get_date(split = "-"),
  todoGDTExtractionTemplate = "Todo-GDT-Extraction-Template.Rmd",
  todoExtractionTemplate = "Todo-Extraction-Template.Rmd",
  todoProgCollectionTemplate = "Todo-Programme-Collection-Template.Rmd",
  todoCollectionTemplate = "Todo-Collection-Template.Rmd"
)
}
\arguments{
\item{location}{A path to a Directory or Project Note file inside an
Organisation where TODO lists should be extracted from each GDT.  Can be
Organisation root, a programme, any sub-directory that contains project
notes, or a single project note.}

\item{date}{The date of extraction, default is today's date. Should be in
char format as 'YYYY-MM-DD' - added to header of Todo Collection.}

\item{todoGDTExtractionTemplate}{Rmd template for todo GDT extraction}

\item{todoExtractionTemplate}{Rmd template for todo Extraction}

\item{todoProgCollectionTemplate}{Rmd template for todo Prog Collection}

\item{todoCollectionTemplate}{Rmd template for todo Collection}
}
\value{
Character Vector of Todo Collection.
}
\description{
Extracted to Todo-Extraction-Template.Rmd that provides a template to
extract an absolute filesystem link to the project note, summarise each
Doc GDT with link, and copy all incompleted TODOs under this Doc GDT.
}
\details{
Todo-Collection-Template.Rmd that provides a template for organising multiple
Project Notes, listing notes by PROGRAMME in organisation directory tree.

incompleted TODOs can then be reviewed, and prioritised by re-ordering &
deleting the project note / doc GDT / TODOs.

The filled Todo-Extraction-Template.Rmd is returned as character vector.
}
