% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_header_Rmd_path}
\alias{find_header_Rmd_path}
\title{Find Group Note Header Rmd Path}
\usage{
find_header_Rmd_path(subNoteRmdPath, settings)
}
\description{
From the \code{subNoteRmdPath} - an absolute path.
}
