% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_prog_dir}
\alias{find_prog_dir}
\title{Find Programme Dir}
\usage{
find_prog_dir(fileSystemPath, settings)
}
\description{
Searches fileSystemPath's parent directories to identify
a Programme directory.  This is identified by
finding a 'PROJECTS/' directory and a 'SOP' directory.
}
\details{
If a Programme path is identified, it is returned, otherwise
the function returns a BLANK string "".
}
