% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{find_prog_dirs}
\alias{find_prog_dirs}
\title{Find Programme Dirs from Organisation path}
\usage{
find_prog_dirs(orgPath, settings)
}
\description{
Searches the orgPath to identify all programmes.
}
