% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{findOrgDir}
\alias{findOrgDir}
\title{Find Organisation Dir}
\usage{
findOrgDir(fileSystemPath)
}
\arguments{
\item{fileSystemPath}{an absolute path in the filesystem - should be within an Organisation.}
}
\description{
Searches fileSystemPath's parent directories to identify
a Project Organisation directory.  This is identified by
finding a 'config/' directory and a 'config/templates/' directory.
}
\details{
If an Organisation path is identified, it is returned, otherwise
the function returns a BLANK string "".
}
