% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_description}
\alias{get_content_description}
\title{Get Content Description}
\usage{
get_content_description(contentDeclContents, settings, orgPath)
}
\arguments{
\item{contentDeclContents}{The Content Declaration Contents - character vector that
includes all lines between content separators.}

\item{settings}{ProjectManagr organisation settings.yml file}

\item{orgPath}{The path to the root of the organisation}
}
\value{
the Content Description value as string
}
\description{
Extract content description from the content Contents, using the ContentDescriptionField
parameter declared in settings.
}
