% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_content_title}
\alias{get_content_title}
\title{Get Content Title}
\usage{
get_content_title(rmdContents, startSepLineIndex, settings, orgPath)
}
\description{
From the note rmdContents, the content with startSep at startSepLineIndex.
}
