% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_contents}
\alias{get_contents}
\title{Get Insertable Contents from file path}
\usage{
get_contents(f, contentSepContents, settings, orgPath)
}
\description{
Get insertable contents from file path (\code{f})
}
\details{
Returns named list of \code{contentTitle} with elements: \code{projectNotePath},
\code{startLineNum}
}
