% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_datetime}
\alias{get_datetime}
\title{internal function to get datetime in YYYY/MM/DD:hh:mm format}
\usage{
get_datetime(timezone = "UTC", split = "-", splitTime = ":")
}
\description{
internal function to get datetime in YYYY/MM/DD:hh:mm format
}
