% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_file_contents_headers}
\alias{get_file_contents_headers}
\title{Get File Contents Headers}
\usage{
get_file_contents_headers(contents)
}
\arguments{
\item{contents}{Character vector containing the contents of the file.}
}
\value{
Character vector of every header line (lines starting with \code{#})
after code blocks have been removed.
}
\description{
Returns every Markdown header line from the text file (all lines that begin
with \code{#}), while skipping over code blocks (anything between lines that
start with triple backticks).
}
\details{
If it detects an odd number of triple-backtick lines, it warns you that
some fences may not be at the start of the line. This can happen if a code
fence is preceded by spaces.
}
\examples{
\dontrun{
  txt <- readLines("somefile.Rmd")
  get_file_contents_headers(txt)
}
}
