% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_file_type}
\alias{get_file_type}
\title{Get File Type}
\usage{
get_file_type(filePath, settings)
}
\arguments{
\item{filePath}{the Absolute path to the file.}

\item{settings}{A named list of the settings.yml file in projectmanagr organisation.}
}
\description{
Determine if the current file is a PROJECT DOC (it is inside the PROJECTS/ Directory),
a HEADER GROUP NOTE (contains the string "-00~_"), or a SIMPLE or SUB Project Note (any
other file).
}
\details{
Returns the relevant String:  DOC, HEAD, SUB, NOTE, or UNKNOWN for unidentified file.
}
