% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{get_goal_title}
\alias{get_goal_title}
\title{get goal title}
\usage{
get_goal_title(goal, settings)
}
\description{
`goal` as collected from selection[["goal"]] - fitting the format specified
in settings[["ProjectGoalHeader"]].
}
