% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem-functions.R
\name{get_project_note_paths}
\alias{get_project_note_paths}
\title{Get Project Note Paths}
\usage{
get_project_note_paths(parentDirectory, settings)
}
\description{
Get project note paths from directory tree.  Assumes `parentDirectory` is in
Organisation, returns blank vector if no project notes are identified.
}
