% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_config_dir}
\alias{get_rstudio_config_dir}
\title{Get RStudio Config Dir}
\usage{
get_rstudio_config_dir()
}
\description{
Returns config directory depending on OS.type (unix or Windows) & rstudio
version (1.2 and prior OR 1.3+)
}
\details{
See :

https://support.posit.co/hc/en-us/articles/206382178
'Customizing Keyboard Shortcuts in the RStudio IDE'

https://support.posit.co/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
'Resetting RStudio Desktop's State'
}
