% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{get_rstudio_internal_state_dir}
\alias{get_rstudio_internal_state_dir}
\title{Get RStudio Internal State Directory}
\usage{
get_rstudio_internal_state_dir()
}
\value{
A filesystem path to the RStudio session state directory (as an
\code{fs::path}), or \code{NULL} if unavailable.
}
\description{
Returns the path to RStudio Desktop's internal session state directory,
taking into account RStudio version and operating system.
}
\details{
\itemize{
\item RStudio ≥ 1.4:
\itemize{
\item macOS/Linux: \verb{~/.local/share/rstudio}
\item Windows: \verb{\%LOCALAPPDATA\%/RStudio}
}
\item RStudio ≤ 1.3:
\itemize{
\item macOS/Linux: \verb{~/.rstudio-desktop}
\item Windows: \verb{\%LOCALAPPDATA\%/RStudio-Desktop}
}
}

RStudio Server is not supported.

see: https://support.posit.co/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
\keyword{internal}
