% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{getProjectPrefixFromName}
\alias{getProjectPrefixFromName}
\title{Get Project Prefix From Name}
\usage{
getProjectPrefixFromName(projectName)
}
\description{
Extracts the Project Prefix from the projectName - all characters before "~_"
}
