% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{insert_header_link_subnote}
\alias{insert_header_link_subnote}
\title{Insert Header Link into Subnote}
\usage{
insert_header_link_subnote(
  subNoteContents,
  headerNoteFileName,
  headerNoteRmdPath,
  subNoteRmdPath,
  headerNoteContentLinkContents,
  settings,
  orgPath
)
}
\arguments{
\item{subNoteContents}{Character vector containing the contents of the
subnote Rmd file.}

\item{headerNoteFileName}{A string specifying the header note file name.}

\item{headerNoteRmdPath}{A string specifying the full file path of the
header note Rmd.}

\item{subNoteRmdPath}{A string specifying the full file path of the
subnote Rmd.}

\item{headerNoteContentLinkContents}{Character vector containing the template
for the header link content.}

\item{settings}{A list of configuration settings used for loading parameter
vectors and template values.}

\item{orgPath}{A string specifying the root directory path of the
organisation.}
}
\description{
This function inserts a hyperlink to the header note into the subnote
contents. It creates the link using the header note file name and path,
replaces the "{{SUB_NOTE_CONTENT_LINK}}" placeholder in the provided
template, and inserts the resulting content into the subnote.
}
\details{
The function creates a hyperlink by calling
\code{create_hyperlink()} with the header note file name and paths. It then
substitutes the placeholder \code{"{{SUB_NOTE_CONTENT_LINK}}"} in the given
template using \code{sub_template_param()}. The insertion point in the
subnote is determined by locating header and footer markers using
\code{load_param_vector()}, \code{match_line_index()}, and
\code{grep_line_index_from()}. Finally, the generated link content is
inserted into the subnote via \code{insert_at_indices()}.
}
\note{
This function assumes that the subnote contains valid header and
footer markers as defined in the settings. Missing markers may lead to
unexpected behavior.
}
\examples{
\dontrun{
# Read subnote contents and header link template.
subNoteContents <- read_file("subnote.Rmd")
headerTemplate <- read_file("HeaderContentLinkTemplate.Rmd")

# Define header note details.
headerFileName <- "header_note.Rmd"
headerPath <- "/path/to/header_note.Rmd"
subNotePath <- "/path/to/subnote.Rmd"

# Load project settings and organisation path.
settings <- get_settings_yml("/path/to/org")
orgPath <- "/path/to/org"

# Insert the header link into the subnote contents.
newSubNoteContents <- insert_header_link_subnote(subNoteContents,
                           headerFileName, headerPath, subNotePath,
                           headerTemplate, settings, orgPath)

# Write the updated subnote contents back to file.
write_file(newSubNoteContents, subNotePath)
}

}
\seealso{
create_hyperlink, sub_template_param, load_param_vector,
match_line_index, grep_line_index_from, insert_at_indices
}
