% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-functions.R
\name{insert_protocol}
\alias{insert_protocol}
\title{Insert a Protocol from Source Project Note into a new Project Note}
\usage{
insert_protocol(
  selectionSource,
  selectionDestination,
  protocolInsertionTemplate = "Protocol-Insertion-Template.Rmd"
)
}
\arguments{
\item{selectionSource}{Selection object from Project Note file containing
the Protocol to be inserted.  The selection must be on the FIRST PROTOCOL
DELIMITER that indicates the start of the Protocol. Use
`projectmanagr::cursor_selection()` or `projectmanagr::user_selection()` to
create this object.}

\item{selectionDestination}{Selection object from the Project Note file where
the protocol is to be inserted.  Use `projectmanagr::cursor_selection()` or
`projectmanagr::user_selection()` to create this object.}

\item{protocolInsertionTemplate}{Template file that contains boilerplate content
for protocol insertion into project note.}
}
\description{
This Function adds a Protocol to a Project Note - protocols are created and
 formed in a Source Project Note.
}
\details{
* Recommended to place all Project Notes that define Protocols into a
 Programme-wide directory with clear name (eg. protocol/), so all protocols
 for a programme can be easily and quickly located.

All links in the Protocol are UPDATED to work from the destination Project
 Note.

All graphics included in a knitr::include_graphics() r code chunk are
 transferred to the new Project Note DIR and linked appropriately.
}
