% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertProtocol.R
\name{insertProtocol}
\alias{insertProtocol}
\title{Insert a Protocol into a Project Note}
\usage{
insertProtocol(
  projectNotePath,
  projectNoteRow,
  protocolName,
  includeNotes = FALSE,
  includeEquip = TRUE
)
}
\arguments{
\item{projectNotePath}{The ABSOLUTE path of the Project Note.}

\item{projectNoteRow}{Index to insert the protocol into project notes}

\item{protocolName}{The name of the Protocol Rmd file}

\item{includeNotes}{Boolean to indicate whether all Notes bullets are included or not.}

\item{includeEquip}{Boolean to indicate whether the initial equipment and material section should be included.}
}
\description{
This Function adds a Protocol to a Project Note - protocols are found in the
 PROGRAMMES' SOP/ Dir.
}
\details{
Inserted protocols add each Procedure defined under the '# PROTOCOL' section
 of the document. Notes sections (FALSE by default) and the Equipment section
(TRUE by default) can be optionally included in the inserted protocol.

All links in the Protocol are UPDATED to work from the destination Project
 Note.

All graphics included in a knitr::include_graphics() r code chunk are
 transferred to the new Project Note DIR and linked appropriately.
}
