% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{link_add_section}
\alias{link_add_section}
\title{Add Section to Link}
\usage{
link_add_section(NoteLink, toFileSection)
}
\arguments{
\item{NoteLink}{Character. A path to the file (without the section).}

\item{toFileSection}{Character. The actual heading text in the Rmd file to link to.}
}
\value{
A character string with the section anchor appended (e.g. "#my-special-section").
}
\description{
\code{NoteLink} is just a path to a file to link to.
\code{toFileSection} is the Markdown header of a section in \code{NoteLink}.
}
\details{
This function appends the header as an HTML-friendly anchor, using pandoc-like
rules. For example, a heading \verb{My Special--Section!} becomes \verb{#my-special-section}.
}
\examples{
\dontrun{
  link_add_section("path/to/myDoc.Rmd", "My Special--Section!")
}
}
