% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-doc-functions.R
\name{link_sub_note_doc}
\alias{link_sub_note_doc}
\title{Link Sub Note to Project Doc GDT}
\usage{
link_sub_note_doc(
  selection,
  settings,
  subNoteRmdPath,
  subNoteContents,
  headerNoteRmdPath,
  headerNoteRmdContents,
  projNoteLinkContents,
  projNoteLinkSummaryContents,
  todoContents,
  projNoteSummaryContents,
  subNoteSummaryContents,
  linkNoteRmdPath,
  linkNoteRmdContents,
  orgPath
)
}
\description{
Single internal function to perform the linking between all project doc GDT
links the \code{linkNoteRmdContents} and the \code{subNoteRmdPath}.
}
\details{
Assumes these are both in the same project note group -
\code{linkNoteRmdContents} should be either the header note of the group, or a
subnote, and \code{subNoteRmdPath} should be a new subnote in the project note
group with no links. This function is used exclusively for the addition of
new sub-notes to a project note group.
}
