% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{match_vector}
\alias{match_vector}
\title{Match vector in parent}
\usage{
match_vector(vector, parent, nomatch = 0L)
}
\description{
Identifies each initial index of \code{vector} in \code{parent} where all elements in
\code{vector} match elements in \code{parent} in order.
}
