% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-addins.R
\name{navigate_markdown_link}
\alias{navigate_markdown_link}
\title{Navigate link to markdown file and header}
\usage{
navigate_markdown_link()
}
\description{
Navigate link to markdown file and header
}
