% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-addins.R
\name{navigate_markdown_link}
\alias{navigate_markdown_link}
\title{Navigate link to markdown file and header}
\usage{
navigate_markdown_link()
}
\description{
This deals with links that exist over multiple lines in the Markdown file: if
the first line of the link is selected, that contains the NAME (in \verb{[]}) and
the FIRST PART of the PATH (ie. the first \code{(}), then will search subsequent
lines to identify the remaining path to navigate.
}
