% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables-functions.R
\name{parse_datatable_create}
\alias{parse_datatable_create}
\title{CREATE a tibble from a datatable Rmd vector}
\usage{
parse_datatable_create(dt_vector, dt_list, table_name)
}
\description{
This datatable has the FUNCTION : CREATE
}
\details{
It MUST have IDs in the first column.  It can have subsequent columns with
further data.
}
