% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-functions.R
\name{rename_project_doc}
\alias{rename_project_doc}
\title{Rename a Project Doc}
\usage{
rename_project_doc(
  projectDocPath,
  newProjectDocName,
  newProjectDocTitle = "",
  replaceLinksFileExtensions = list("Rmd")
)
}
\arguments{
\item{projectDocPath}{defines the path to the Project Doc - should point
to the Project Doc Rmd file}

\item{newProjectDocName}{defines the NEW Project Doc File name}

\item{newProjectDocTitle}{defines the NEW Project Doc Title - written to
its Rmd file.  By default this is the projectDocPath, with spaces replacing
"-" and "_".}
}
\description{
Renames a Project Doc at projectDocPath with file name newProjectDocName,
and replaces title with newProjectDocTitle if set.
}
\details{
This also modifies all LINKS to this file throughout the Organisation -
updating them to use the new project Doc name and title.

Doc - this DOES NOT modify the PREFIX for a Project Doc.
}
