% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-functions.R
\name{rename_project_doc_goal}
\alias{rename_project_doc_goal}
\title{Rename a Project Doc Goal}
\usage{
rename_project_doc_goal(goalSelection, newGoalName)
}
\arguments{
\item{goalSelection}{Selection from projectmanagr Project Doc file, made on
the Goal to be renamed.}

\item{newGoalName}{The new name to use for the selected Goal.}
}
\description{
Rename a Project Doc Goal
}
