% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-utils-functions.R
\name{replace_knitr_include_graphics_link}
\alias{replace_knitr_include_graphics_link}
\title{Replace Knitr include_graphics link}
\usage{
replace_knitr_include_graphics_link(
  contentContents,
  sourceFilePath,
  destinationFilePath,
  knitrGraphics = "knitr::include_graphics(",
  knitrGraphicsEnd = ")"
)
}
\arguments{
\item{contentContents}{Character vector of lines}

\item{sourceFilePath}{Path of source file (where content was read from)}

\item{destinationFilePath}{Path of destination file}

\item{knitrGraphics}{Function call to search for (default: 'knitr::include_graphics(')}

\item{knitrGraphicsEnd}{End of function call (default: ')')}
}
\value{
Modified character vector with updated relative links
}
\description{
Replace the path inside knitr::include_graphics() with a relative
path based on source and destination file paths.
}
