% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{replace_params_with_vector}
\alias{replace_params_with_vector}
\title{Replace every instance of `templateParam` in `templateContents` with `paramContents`.}
\usage{
replace_params_with_vector(templateContents, templateParam, paramContents)
}
\description{
Replace every instance of `templateParam` in `templateContents` with `paramContents`.
}
