% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{replace_sep_values}
\alias{replace_sep_values}
\title{Replace SEP Values}
\usage{
replace_sep_values(templateContents, orgPath)
}
\description{
In a string vector, replace every instance of SEP values - syntax used to denote
separators between document sections in projectmanagr templates.
}
\details{
* Six separators strings are defined in config/templates in projectmanagr org
in the files: `SEP01.txt` to `SEP06.txt`

* Syntax in projectmanagr templates for inserting the separators in projectmanagr
templates: `{{SEP01}}` to `{{SEP06}}`
}
