% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-functions.R
\name{set_rstudio_keybindings_addins}
\alias{set_rstudio_keybindings_addins}
\title{Set RStudio Keybindings to Addings}
\usage{
set_rstudio_keybindings_addins()
}
\description{
Sets the default keybindings for projectmanagr in rstudio, via the
keybindings/addins.json file
}
