% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{update_links_filenames}
\alias{update_links_filenames}
\title{Update links}
\usage{
update_links_filenames(
  oldFileName,
  newName,
  dirTree,
  settings,
  fileExtensions = list("Rmd")
)
}
\arguments{
\item{oldFileName}{defines the OLD name that will be in Hyperlinks.  Should be
the FILE NAME - with no spaces, including PREFIX and .Rmd EXTENSION}

\item{newName}{defines the NEW name to be written into Hyperlinks.  Should
be the FILE NAME - with no spaces.}

\item{dirTree}{Directory tree to search for files for replacing links in.}

\item{fileExtensions}{List of file extensions indicating what types of files
should be searched for links and have links replaced.  Must be plaintext file
type.  Default is Rmd files only.}

\item{oldTitle}{defines the OLD TITLE that will be in Hyperlinks.  By default
replaces - & _ with spaces from name.}

\item{newTitle}{defines the NEW TITLE that will be in Hyperlinks.  By default
replaces - & _ with spaces from name.}
}
\description{
Updates every hyperlink in all files within \code{dirTree}, replacing
\code{oldFileName} (the basename of the file, including prefix and extension) with
\code{newName} (the new file name, without prefix or extension).
}
