% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-functions.R
\name{updateAllLinks}
\alias{updateAllLinks}
\title{Update all links}
\usage{
updateAllLinks(orgPath, oldName, oldTitle, newName, newTitle)
}
\arguments{
\item{orgPath}{defines the path to the Organisation - should point to the root directory of the Organisation.}

\item{oldName}{defines the NEW name to be written into Hyperlinks.  Should be the FILE NAME - with no spaces.}
}
\description{
Updates every hyperlink in all Rmd files within orgPath directory tree, from oldName
to newName.
}
