% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-functions.R
\name{yaml_keys}
\alias{yaml_keys}
\title{Retrieve Expected YAML Keys}
\usage{
yaml_keys()
}
\value{
Character vector. List of expected YAML keys.
}
\description{
Returns a character vector of all expected keys that should be present in a
YAML configuration file.
}
